split particle files
