+++ config/GALE01/splits.txt
@@ -5450,15 +5450,48 @@ sysdolphin/baselib/gobjinit.c:
 	.text       start:0x803912E0 end:0x80391580
 	.data       start:0x80408620 end:0x8040862C
 
+sysdolphin/baselib/hsd_3915.c:
+	.text       start:0x80391580 end:0x80392CCC
+	.data       start:0x80408630 end:0x8040A540
+	.bss        start:0x804CE3F8 end:0x804CEB40
+	.sdata      start:0x804D6070 end:0x804D6358
+	.sbss       start:0x804D7850 end:0x804D78D0
+	.sdata2     start:0x804DE890 end:0x804DE930
+
+sysdolphin/baselib/hsd_392C.c:
+	.text       start:0x80392CCC end:0x803932D0
+	.data       start:0x8040A540 end:0x8040A920
+
+sysdolphin/baselib/hsd_3933.c:
+	.text       start:0x803932D0 end:0x80394314
+	.data       start:0x8040A920 end:0x8040AB00
+	.bss        start:0x804CEB40 end:0x804CF810
+	.sdata2     start:0x804DE930 end:0x804DE948
+
+sysdolphin/baselib/debugconsole_main.c:
+	.text       start:0x80394314 end:0x803983A4
+	.data       start:0x8040AB00 end:0x8040BFB0
+	.bss        start:0x804CF810 end:0x804D08E8
+
 sysdolphin/baselib/particle.c:
-	extab       start:0x80005520 end:0x800055F8
-	extabindex  start:0x800056C0 end:0x80005804
-	.text       start:0x80391580 end:0x8039F78C
-	.data       start:0x80408630 end:0x8040C300
-	.bss        start:0x804CE3F8 end:0x804D0FC0
-	.sdata      start:0x804D6070 end:0x804D6380
-	.sbss       start:0x804D7850 end:0x804D7908
-	.sdata2     start:0x804DE890 end:0x804DEA10
+	extab       start:0x80005520 end:0x80005588
+	extabindex  start:0x800056C0 end:0x8000575C
+	.text       start:0x803983A4 end:0x8039D1E4
+	.data       start:0x8040BFB0 end:0x8040C248
+	.bss        start:0x804D08E8 end:0x804D0F90
+	.sdata      start:0x804D6358 end:0x804D6368
+	.sbss       start:0x804D78D0 end:0x804D78F8
+	.sdata2     start:0x804DE948 end:0x804DE9A8
+
+sysdolphin/baselib/generator.c:
+	extab       start:0x80005588 end:0x800055F8
+	extabindex  start:0x8000575C end:0x80005804
+	.text       start:0x8039D1E4 end:0x8039F78C
+	.data       start:0x8040C248 end:0x8040C300
+	.bss        start:0x804D0F90 end:0x804D0FC0
+	.sdata      start:0x804D6368 end:0x804D6380
+	.sbss       start:0x804D78F8 end:0x804D7908
+	.sdata2     start:0x804DE9A8 end:0x804DEA10
 
 sysdolphin/baselib/psdisp.c:
 	extab       start:0x800055F8 end:0x80005618
