+++ config/GALE01/symbols.txt
@@ -25728,11 +25728,20 @@ lbl_8040A76C = .data:0x8040A76C; // type:object size:0xE scope:local data:string
 lbl_8040A77C = .data:0x8040A77C; // type:object size:0x18 scope:local data:string
 lbl_8040A794 = .data:0x8040A794; // type:object size:0x14 scope:local data:string
 jumptable_8040A7A8 = .data:0x8040A7A8; // type:object size:0x58 scope:local
-lbl_8040A800 = .data:0x8040A800; // type:object size:0xE0 scope:local
+lbl_8040A800 = .data:0x8040A800; // type:object size:0x1A data:string
+lbl_8040A81C = .data:0x8040A81C; // type:object size:0xD data:string
+lbl_8040A82C = .data:0x8040A82C; // type:object size:0xC data:string
+lbl_8040A838 = .data:0x8040A838; // type:object size:0x1F data:string
+lbl_8040A858 = .data:0x8040A858; // type:object size:0x1F data:string
+lbl_8040A878 = .data:0x8040A878; // type:object size:0xA data:string
+lbl_8040A884 = .data:0x8040A884; // type:object size:0x29 data:string
+lbl_8040A8B0 = .data:0x8040A8B0; // type:object size:0xB data:string
+lbl_8040A8BC = .data:0x8040A8BC; // type:object size:0x21 data:string
 lbl_8040A8E0 = .data:0x8040A8E0; // type:object size:0x23 scope:local data:string
 lbl_8040A904 = .data:0x8040A904; // type:object size:0x1A scope:local data:string
 lbl_8040A920 = .data:0x8040A920; // type:object size:0x1B scope:local data:string
-lbl_8040A93C = .data:0x8040A93C; // type:object size:0x94 scope:local
+lbl_8040A93C = .data:0x8040A93C; // type:object size:0x80 scope:local
+lbl_8040A9BC = .data:0x8040A9BC; // type:object size:0xE data:string
 psUsbMessages = .data:0x8040A9D0; // type:object size:0x130 scope:local
 lbl_8040AB00 = .data:0x8040AB00; // type:object size:0x20 scope:local
 lbl_8040AB20 = .data:0x8040AB20; // type:object size:0x20 scope:local
@@ -25814,7 +25823,8 @@ lbl_8040B838 = .data:0x8040B838; // type:object size:0x14 scope:local data:strin
 lbl_8040B84C = .data:0x8040B84C; // type:object size:0x15 scope:local data:string
 lbl_8040B864 = .data:0x8040B864; // type:object size:0xB scope:local data:string
 jumptable_8040B870 = .data:0x8040B870; // type:object size:0x3C scope:local
-lbl_8040B8AC = .data:0x8040B8AC; // type:object size:0x2C scope:local
+lbl_8040B8AC = .data:0x8040B8AC; // type:object size:0x18 data:string
+lbl_8040B8C4 = .data:0x8040B8C4; // type:object size:0x14
 lbl_8040B8D8 = .data:0x8040B8D8; // type:object size:0x29 scope:local data:string
 lbl_8040B904 = .data:0x8040B904; // type:object size:0x10 scope:local
 jumptable_8040B914 = .data:0x8040B914; // type:object size:0x24 scope:local
@@ -25867,7 +25877,9 @@ lbl_8040BE6C = .data:0x8040BE6C; // type:object size:0x10 scope:local data:strin
 lbl_8040BE7C = .data:0x8040BE7C; // type:object size:0x17 scope:local data:string
 lbl_8040BE94 = .data:0x8040BE94; // type:object size:0x17 scope:local data:string
 lbl_8040BEAC = .data:0x8040BEAC; // type:object size:0x17 scope:local data:string
-lbl_8040BEC4 = .data:0x8040BEC4; // type:object size:0x4C scope:local
+lbl_8040BEC4 = .data:0x8040BEC4; // type:object size:0x1C scope:local
+lbl_8040BEE0 = .data:0x8040BEE0; // type:object size:0x14 data:string
+lbl_8040BEF4 = .data:0x8040BEF4; // type:object size:0x17 data:string
 lbl_8040BF10 = .data:0x8040BF10; // type:object size:0x32 scope:local data:string
 lbl_8040BF48 = .data:0x8040BF48; // type:object size:0x14 scope:local
 lbl_8040BF5C = .data:0x8040BF5C; // type:object size:0x14 scope:local
@@ -26457,10 +26469,11 @@ hsd_804CF7E8 = .bss:0x804CF7E8; // type:object size:0x28 scope:local data:byte
 hsd_804CF810 = .bss:0x804CF810; // type:object size:0xD8 scope:local data:byte
 hsd_804CF8E8 = .bss:0x804CF8E8; // type:object size:0x1000 scope:local
 hsd_804D08E8 = .bss:0x804D08E8; // type:object size:0x20 scope:local data:4byte
-hsd_804D0908 = .bss:0x804D0908; // type:object size:0x248 scope:local data:4byte
-psTexGroupArray = .bss:0x804D0B50; // type:object size:0x104 scope:local data:4byte
-psNumCmdList = .bss:0x804D0C54; // type:object size:0x104 scope:local data:4byte
-psCmdListArray = .bss:0x804D0D58; // type:object size:0x104 scope:local data:4byte
+hsd_804D0908 = .bss:0x804D0908; // type:object size:0x144 scope:local data:4byte
+psFormGroupArray = .bss:0x804D0A4C; // type:object size:0x104 scope:local data:4byte
+psTexGroupArray = .bss:0x804D0B50; // type:object size:0x104 scope:global data:4byte
+psNumCmdList = .bss:0x804D0C54; // type:object size:0x104 scope:global data:4byte
+psCmdListArray = .bss:0x804D0D58; // type:object size:0x104 scope:global data:4byte
 ptclref_804D0E5C = .bss:0x804D0E5C; // type:object size:0x104 scope:local data:4byte
 hsd_804D0F60 = .bss:0x804D0F60; // type:object size:0x30 scope:local data:byte
 hsd_804D0F90 = .bss:0x804D0F90; // type:object size:0x30 scope:local data:byte
@@ -28056,7 +28069,7 @@ HSD_DevCom_804D6050 = .sdata:0x804D6050; // type:object size:0x4 scope:global da
 lbl_804D6070 = .sdata:0x804D6070; // type:object size:0x4 scope:local align:4 data:float
 lbl_804D6074 = .sdata:0x804D6074; // type:object size:0x4 scope:local align:4 data:float
 lbl_804D6078 = .sdata:0x804D6078; // type:object size:0x8 scope:local data:byte
-lbl_804D6080 = .sdata:0x804D6080; // type:object size:0x8 scope:local data:4byte
+lbl_804D6080 = .sdata:0x804D6080; // type:object size:0x4 scope:local data:4byte
 lbl_804D6088 = .sdata:0x804D6088; // type:object size:0x4 scope:local data:4byte
 lbl_804D608C = .sdata:0x804D608C; // type:object size:0x4 scope:local data:4byte
 lbl_804D6090 = .sdata:0x804D6090; // type:object size:0x4 scope:local data:4byte
@@ -29301,20 +29314,20 @@ hsd_804D78C0 = .sbss:0x804D78C0; // type:object size:0x8 scope:local data:4byte
 hsd_804D78C8 = .sbss:0x804D78C8; // type:object size:0x4 scope:local data:4byte
 hsd_804D78CC = .sbss:0x804D78CC; // type:object size:0x4 scope:local data:4byte
 hsd_804D78D0 = .sbss:0x804D78D0; // type:object size:0x4 scope:local align:4 data:float
-hsd_804D78D4 = .sbss:0x804D78D4; // type:object size:0x4 scope:local data:4byte
+psCallback = .sbss:0x804D78D4; // type:object size:0x4 scope:local data:4byte
 hsd_804D78D8 = .sbss:0x804D78D8; // type:object size:0x2 scope:local data:2byte
 hsd_804D78DA = .sbss:0x804D78DA; // type:object size:0x2 scope:local data:2byte
-hsd_804D78DC = .sbss:0x804D78DC; // type:object size:0x2 scope:local data:2byte
+numPeakParticles = .sbss:0x804D78DC; // type:object size:0x2 scope:local data:2byte
 hsd_804D78DE = .sbss:0x804D78DE; // type:object size:0x2 scope:local data:2byte
 hsd_804D78E0 = .sbss:0x804D78E0; // type:object size:0x2 scope:local data:2byte
-hsd_804D78E2 = .sbss:0x804D78E2; // type:object size:0x6 scope:local data:2byte
+hsd_804D78E2 = .sbss:0x804D78E2; // type:object size:0x4 scope:local data:2byte
 hsd_804D78E8 = .sbss:0x804D78E8; // type:object size:0x4 scope:local data:4byte
 hsd_804D78EC = .sbss:0x804D78EC; // type:object size:0x4 scope:local data:4byte
-hsd_804D78F0 = .sbss:0x804D78F0; // type:object size:0x4 scope:local data:4byte
+psCamera = .sbss:0x804D78F0; // type:object size:0x4 scope:global data:4byte
 hsd_804D78F4 = .sbss:0x804D78F4; // type:object size:0x4 scope:local data:4byte
 hsd_804D78F8 = .sbss:0x804D78F8; // type:object size:0x4 scope:local data:4byte
 hsd_804D78FC = .sbss:0x804D78FC; // type:object size:0x4 scope:local data:4byte
-hsd_804D7900 = .sbss:0x804D7900; // type:object size:0x8 scope:local data:4byte
+hsd_804D7900 = .sbss:0x804D7900; // type:object size:0x4 scope:local data:4byte
 HSD_PSDisp_804D7908 = .sbss:0x804D7908; // type:object size:0x4 scope:global data:4byte
 HSD_PSDisp_804D790C = .sbss:0x804D790C; // type:object size:0x4 scope:global data:4byte
 HSD_PSDisp_804D7910 = .sbss:0x804D7910; // type:object size:0x4 scope:global data:4byte
