+++ configure.py
@@ -1415,7 +1415,7 @@ config.libs = [
             Object(Matching, "melee/if/ifprize.c"),
             Object(NonMatching, "melee/if/ifcoget.c"),
             Object(NonMatching, "melee/if/soundtest.c"),
-            Object(NonMatching, "melee/if/textdraw.c"),
+            Object(NonMatching, "melee/if/textdraw.c", extra_cflags=["-requireprotos"]),
             Object(NonMatching, "melee/if/textlib.c"),
         ],
     ),
@@ -1765,13 +1765,13 @@ config.libs = [
             Object(Matching, "sysdolphin/baselib/gobjuserdata.c"),
             Object(Matching, "sysdolphin/baselib/gobj.c"),
             Object(Matching, "sysdolphin/baselib/gobjinit.c"),
-            Object(
-                NonMatching,
-                "sysdolphin/baselib/particle.c",
-                extra_cflags=["-Cpp_exceptions on"],
-            ),
-            Object(
-                NonMatching,
+            Object(NonMatching, "sysdolphin/baselib/hsd_3915.c"),
+            Object(NonMatching, "sysdolphin/baselib/hsd_392C.c"),
+            Object(NonMatching, "sysdolphin/baselib/hsd_3933.c"),
+            Object(NonMatching, "sysdolphin/baselib/debugconsole_main.c"),
+            Object(NonMatching, "sysdolphin/baselib/particle.c", extra_cflags=["-Cpp_exceptions on"]),
+            Object(NonMatching, "sysdolphin/baselib/generator.c", extra_cflags=["-Cpp_exceptions on"]),
+            Object(NonMatching,
                 "sysdolphin/baselib/psdisp.c",
                 extra_cflags=["-Cpp_exceptions on"],
             ),
