+++ src/melee/db/dberror.c
@@ -4,7 +4,8 @@
 #include <dolphin/db.h>
 #include <dolphin/os.h>
 #include <sysdolphin/baselib/debug.h>
-#include <sysdolphin/baselib/particle.h>
+#include <sysdolphin/baselib/debugconsole_main.h>
+#include <sysdolphin/baselib/hsd_3933.h>
 #include <sysdolphin/baselib/video.h>
 #include <melee/lb/lb_0195.h>
 #include <melee/lb/lbcardnew.h>
