+++ src/melee/gm/gmtoulib.c
@@ -40,7 +40,7 @@
 #include <baselib/gobjproc.h>
 #include <baselib/jobj.h>
 #include <baselib/mobj.h>
-#include <baselib/particle.h>
+#include <baselib/hsd_3915.h>
 #include <baselib/random.h>
 #include <baselib/sislib.h>
