+++ src/melee/lb/lbbgflash.c
@@ -53,7 +53,7 @@ extern BgFlashData lbl_80433658;
 #include <baselib/jobj.h>
 #include <baselib/mtx.h>
 #include <baselib/objalloc.h>
-#include <baselib/particle.h>
+#include <baselib/hsd_3915.h>
 #include <baselib/quatlib.h>
 #include <melee/lb/lbarchive.h>
 #include <melee/lb/lbspdisplay.h>
