+++ src/sysdolphin/baselib/hsd_392C.h
@@ -1,0 +1,12 @@
+#ifndef HSD_392C_H
+#define HSD_392C_H
+
+#include <platform.h>
+
+/* 392CCC */ s32 fn_80392CCC(s32);
+/* 392CD8 */ u8 fn_80392CD8(char*);
+/* 392E2C */ void fn_80392E2C(s32);
+/* 392E80 */ int hsd_80392E80(void);
+/* 3931A4 */ bool hsd_803931A4(s32);
+
+#endif
