+++ src/sysdolphin/baselib/hsd_3933.h
@@ -1,0 +1,30 @@
+#ifndef HSD_3933_H
+#define HSD_3933_H
+
+#include <placeholder.h>
+#include <platform.h>
+
+/* 3932D0 */ void fn_803932D0(s32, u32, s32);
+/* 393328 */ s32 hsd_80393328(void);
+/* 393440 */ void hsd_80393440(void*, void*);
+/* 393840 */ void hsd_80393840(void);
+/* 393844 */ void hsd_80393844(void);
+/* 393A04 */ bool hsd_80393A04(void);
+/* 393A54 */ void hsd_80393A54(int);
+/* 393A5C */ int hsd_80393A5C(char*, int, int);
+/* 393C14 */ void fn_80393C14(const u8* buf, size_t);
+/* 393D2C */ s32 hsd_80393D2C(s32);
+/* 393DA0 */ void hsd_80393DA0(u8*, size_t);
+/* 393E34 */ void hsd_80393E34(s32* col_out, s32* row_out);
+/* 393E68 */ void hsd_80393E68(u32, u32);
+/* 393EF4 */ void hsd_80393EF4(int, int);
+/* 394068 */ u8 hsd_80394068(void);
+/* 394128 */ u8 hsd_80394128(s32, s32);
+/* 3941E8 */ s32 hsd_803941E8(void*, void*);
+/* 394314 */ void hsd_80394314(void);
+/* 394434 */ void hsd_80394434(void*);
+/* 394544 */ void hsd_80394544(s32, s32, u32, u32, s32, s32, s32, s32, s32,
+                               s32, s32, void*);
+/* 394668 */ void hsd_80394668(void);
+
+#endif
