+++ src/sysdolphin/baselib/particle.c
@@ -1,5 +1,7 @@
 #include "particle.h"
 
+#include "generator.h"
+
 typedef struct {
     /* 0x00 */ void* next;
     /* 0x04 */ s32 type;
@@ -16,8 +18,6 @@ typedef struct {
         } gradient[16];
     } content;
 } PerfDispItem;
-
-static PerfDispItem hsd_804CE3F8[6];
 
 #include "particle.static.h"
 
@@ -45,46 +45,14 @@ static PerfDispItem hsd_804CE3F8[6];
 #include <MSL/math_ppc.h>
 #include <MSL/trigf.h>
 
-typedef struct {
-    s32 x0;
-    u32 x4;
-    s32 x8;
-} ParticleLogEntry;
-
-typedef struct _ExcptNode {
-    /* 0x1 */ struct _ExcptNode* next;
-    /* 0x4 */ void (*callback)(struct _ExcptNode*);
-} ExcptNode;
-
-typedef struct _EventData {
-    /* 0x00 */ u8 _pad[0x10];
-    /* 0x10 */ char** entries;
-    /* 0x14 */ s32 index;
-} EventData;
-
-typedef struct _MCCPacket {
-    /* 0x0 */ s32 x0;
-    /* 0x4 */ u8 x4_b7 : 1;
-    /* 0x4 */ u8 _x4_pad : 7;
-    /* 0x5 */ u8 x5;
-} MCCPacket;
-
-typedef struct _ParticleFontData {
-    /* 0x0000 */ u8 _pad0[0x700];
-    /* 0x0700 */ u8 x700[0x38];
-    /* 0x0738 */ u8 _pad738[0x230];
-    /* 0x0968 */ u8 x968[0x38];
-    /* 0x09A0 */ u8 _pad9A0[0x38];
-    /* 0x09D8 */ u8 x9D8[0x38];
-    /* 0x0A10 */ u8 _padA10[0x1110];
-    /* 0x1B20 */ u8 x1B20[0x38];
-} ParticleFontData;
-
-typedef struct _DispData {
-    /* 0x00 */ u8 _pad[0x10];
-    /* 0x10 */ s32 x10;
-    /* 0x14 */ s32 x14;
-} DispData;
+static HSD_JObj* hsd_804D08E8[8];
+static void* hsd_804D0908[0x144 / 4];
+static HSD_PSFormGroup** psFormGroupArray[65];
+/* 4D0B50 */ static HSD_PSTexGroup** psTexGroupArray[65];
+/* 4D0D58 */ static int psNumCmdList[65];
+/* 4D0C54 */ static HSD_PSCmdList** psCmdListArray[65];
+/* 4D0E5C */ static u32* ptclref_804D0E5C[65];
+/* 4D0F60 */ static struct hsd_804D0F60_t hsd_804D0F60;
 
 typedef struct PSNode {
     /* 0x00 */ struct PSNode* child;
@@ -92,4500 +60,10 @@ typedef struct PSNode {
     /* 0x08 */ void (*callback)(struct PSNode*);
 } PSNode;
 
-typedef struct {
-    /* 0x00 */ s32 x0;
-    /* 0x04 */ void (*callback)(void* dst, s32 x, s32 y, s32 val, void* self);
-} GlyphEntry;
-
-typedef struct {
-    s32 count;
-    u32 color;
-} DispBar;
-
-typedef struct _DispItem {
-    /* 0x00 */ struct _DispItem* next;
-    /* 0x04 */ s32 type;
-    /* 0x08 */ union {
-        char text[128];
-        DispBar bars[1];
-        u8 gradient[8];
-    } content;
-} DispItem;
-
-typedef DispItem* (*DispCallback)(void*);
-
 typedef union {
     f32 f;
     u8 bytes[4];
 } ParticleFloatBytes;
-
-struct EventPriority {
-    Event event;
-    int priority;
-};
-
-static u16 numPeakParticles;
-
-void DrawRectangle(f32 x_min, f32 y_min, f32 w, f32 h, GXColor* color)
-{
-    f32 x_max;
-    f32 y_max;
-    GXColor c;
-    u8 r, g, b, a;
-
-    GXBegin(0x80, 0, 4);
-
-    x_max = x_min + w;
-    y_max = y_min + h;
-
-    // Send the corners in clockwise order, starting with top left
-
-    GXWGFifo.f32 = x_min;
-    GXWGFifo.f32 = y_min;
-    a = color->a;
-    b = color->b;
-    g = color->g;
-    r = color->r;
-    GXWGFifo.u8 = r;
-    GXWGFifo.u8 = g;
-    GXWGFifo.u8 = b;
-    GXWGFifo.u8 = a;
-
-    GXWGFifo.f32 = x_max;
-    GXWGFifo.f32 = y_min;
-    GXWGFifo.u8 = r;
-    GXWGFifo.u8 = g;
-    GXWGFifo.u8 = b;
-    GXWGFifo.u8 = a;
-
-    GXWGFifo.f32 = x_max;
-    GXWGFifo.f32 = y_max;
-    GXWGFifo.u8 = r;
-    GXWGFifo.u8 = g;
-    GXWGFifo.u8 = b;
-    GXWGFifo.u8 = a;
-
-    GXWGFifo.f32 = x_min;
-    GXWGFifo.f32 = y_max;
-    GXWGFifo.u8 = r;
-    GXWGFifo.u8 = g;
-    GXWGFifo.u8 = b;
-    GXWGFifo.u8 = a;
-
-    GXEnd();
-}
-
-static u8 lbl_80408630[0x268] = {
-    0x00, 0x60, 0x60, 0x66, 0x66, 0x06, 0x06, 0x00, 0xFF, 0x00, 0x00, 0x00,
-    0x00, 0x30, 0x36, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x60, 0x60, 0x63, 0x63, 0x03, 0x03, 0x06, 0x06, 0x66,
-    0xFF, 0x00, 0x00, 0x00, 0x60, 0x60, 0x66, 0x66, 0x06, 0x13, 0x63, 0xFF,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x03, 0x63, 0x40, 0x46, 0xFF, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x60, 0x00, 0x00, 0x03, 0x03, 0x63, 0x63,
-    0x66, 0x66, 0x06, 0xFF, 0x00, 0x00, 0x00, 0x06, 0x06, 0x66, 0x66, 0x63,
-    0x63, 0x03, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x60, 0x60, 0x62, 0x62,
-    0x43, 0x43, 0x46, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x60, 0x60, 0x66,
-    0x66, 0x06, 0x06, 0x00, 0x03, 0x63, 0xFF, 0x00, 0x00, 0x63, 0x03, 0x03,
-    0x00, 0x00, 0x60, 0x60, 0x66, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x06, 0x30,
-    0x30, 0x66, 0x14, 0x54, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    0x40, 0x40, 0x61, 0x61, 0x66, 0x66, 0x06, 0x06, 0x00, 0x03, 0x63, 0xFF,
-    0x60, 0x20, 0x20, 0x03, 0x03, 0x26, 0x26, 0x66, 0xFF, 0x00, 0x00, 0x00,
-    0x00, 0x00, 0x30, 0x30, 0x63, 0x63, 0x36, 0x36, 0x06, 0x10, 0x16, 0xFF,
-    0x00, 0x00, 0x60, 0x00, 0x00, 0x06, 0x06, 0x66, 0x03, 0x53, 0xFF, 0x00,
-    0x00, 0x00, 0x00, 0x60, 0x00, 0x00, 0x06, 0x02, 0x42, 0xFF, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x60, 0x30, 0x30, 0x02, 0x02, 0x06, 0x06, 0x66,
-    0x66, 0x63, 0x63, 0x33, 0xFF, 0x00, 0x06, 0x60, 0x66, 0x03, 0x63, 0xFF,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x10, 0x50, 0x16, 0x56, 0x30, 0x36,
-    0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x50, 0x56, 0x56, 0x16, 0x16,
-    0x14, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x60, 0x03,
-    0x03, 0x66, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x06,
-    0x66, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x00,
-    0x00, 0x33, 0x33, 0x60, 0x60, 0x66, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x06,
-    0x00, 0x00, 0x66, 0x66, 0x60, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    0x30, 0x63, 0x63, 0x36, 0x36, 0x03, 0x03, 0x30, 0xFF, 0x00, 0x00, 0x00,
-    0x00, 0x06, 0x00, 0x00, 0x50, 0x50, 0x61, 0x61, 0x62, 0x62, 0x53, 0x53,
-    0x03, 0xFF, 0x06, 0x02, 0x02, 0x60, 0x60, 0x64, 0x64, 0x06, 0x24, 0x46,
-    0x46, 0x66, 0xFF, 0x06, 0x00, 0x00, 0x60, 0x60, 0x62, 0x62, 0x02, 0x32,
-    0x66, 0xFF, 0x00, 0x00, 0x60, 0x00, 0x00, 0x66, 0x66, 0x06, 0xFF, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x60, 0x30, 0x36, 0xFF, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x06, 0x66, 0x66, 0x60,
-    0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x02, 0x36, 0x36,
-    0x62, 0x62, 0x60, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x16, 0x16, 0x56,
-    0x56, 0x60, 0x32, 0x36, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x60,
-    0x06, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x32,
-    0x32, 0x60, 0x32, 0x36, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    0x60, 0x60, 0x06, 0x06, 0x66, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    0x03, 0x63, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    0x00, 0x06, 0x66, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    0x00, 0x00, 0x03, 0x22, 0x22, 0x44, 0x44, 0x63, 0xFF, 0x00, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x63, 0x63, 0x06, 0xFF, 0x00, 0x00, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x50, 0x16, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x22, 0x22, 0x24, 0x24, 0x46, 0xFF,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x42, 0x42, 0x44, 0x44, 0x26,
-    0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x44, 0x26, 0xFF, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x60, 0x03, 0x03, 0x66,
-    0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x30, 0x36, 0x03,
-    0x63, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x61, 0x01,
-    0x01, 0x65, 0x65, 0x05, 0x30, 0x36, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00,
-};
-static float lbl_804D6070 = 12.0F;
-static float lbl_804D6074 = 12.0F;
-
-f32 DrawASCII(int chr, float x, float y, GXColor* color)
-{
-    s32 index;
-    u8* glyph;
-    s32 i;
-    u8 p0, p1;
-    PAD_STACK(16);
-
-    if (chr >= '0' && chr <= '9') {
-        index = chr - '0';
-    } else if (chr >= 'A' && chr <= 'Z') {
-        index = chr - 'A' + 10;
-    } else if (chr >= 'a' && chr <= 'z') {
-        index = chr - 'a' + 10;
-    } else {
-        switch (chr) {
-        case '.': {
-            f32 px;
-            f32 py;
-            GXBegin(0xB8, 0, 1);
-            px = (f32) ((f64) lbl_804D6070 * 0.3 + (f64) x);
-            py = (f32) ((f64) lbl_804D6074 * 0.9 + (f64) y);
-            GXPosition2f32(px, py);
-            GXColor4u8(color->r, color->g, color->b, color->a);
-            return lbl_804D6070;
-        }
-        case ':': {
-            GXBegin(0xB8, 0, 2);
-            GXPosition2f32((f32) ((f64) lbl_804D6070 * 0.3 + (f64) x),
-                           (f32) ((f64) lbl_804D6074 * 0.3 + (f64) y));
-            GXColor4u8(color->r, color->g, color->b, color->a);
-            GXPosition2f32((f32) ((f64) lbl_804D6070 * 0.3 + (f64) x),
-                           (f32) ((f64) lbl_804D6074 * 0.7 + (f64) y));
-            GXColor4u8(color->r, color->g, color->b, color->a);
-            return lbl_804D6070;
-        }
-        case '-':
-            index = 0x24;
-            break;
-        case '_':
-            index = 0x25;
-            break;
-        case '~':
-            index = 0x26;
-            break;
-        case '>':
-            index = 0x27;
-            break;
-        case '/':
-            index = 0x28;
-            break;
-        case '(':
-            index = 0x29;
-            break;
-        case ')':
-            index = 0x2A;
-            break;
-        case ',':
-            index = 0x2B;
-            break;
-        case '<':
-            index = 0x2C;
-            break;
-        case '+':
-            index = 0x2D;
-            break;
-        case '$':
-            index = 0x2E;
-            break;
-        default:
-            return lbl_804D6070;
-        }
-    }
-
-    glyph = &lbl_80408630[index * 13];
-    i = 0;
-    while (i < 0x29 && (u8) glyph[i] != 0xFF) {
-        p0 = glyph[i++];
-        p1 = glyph[i++];
-        GXBegin(0xA8, 0, 2);
-        GXPosition2f32(lbl_804D6070 * (0.11F * (p0 >> 4)) + x,
-                       lbl_804D6074 * (0.11F * (p0 & 0xF)) + y);
-        GXColor4u8(color->r, color->g, color->b, color->a);
-        GXPosition2f32(lbl_804D6070 * (0.11F * (p1 >> 4)) + x,
-                       lbl_804D6074 * (0.11F * (p1 & 0xF)) + y);
-        GXColor4u8(color->r, color->g, color->b, color->a);
-    }
-    return lbl_804D6070;
-}
-
-static u8 lbl_804D6078 = 12;
-
-void hsd_80391A04(float scale_x, float scale_y, int line_width)
-{
-    Mtx view_mtx;
-
-    lbl_804D6070 = scale_x;
-    lbl_804D6074 = scale_y;
-    lbl_804D6078 = line_width;
-    HSD_ClearVtxDesc();
-    GXSetCurrentMtx(0);
-    HSD_CObjGetViewingMtx(HSD_CObjGetCurrent(), view_mtx);
-    GXLoadPosMtxImm(view_mtx, 0);
-    HSD_StateSetLineWidth(lbl_804D6078, 5);
-    HSD_StateSetPointSize(lbl_804D6078 * 2, 5);
-    HSD_SetupRenderMode(0x68000002);
-    GXSetVtxAttrFmt(GX_VTXFMT0, GX_VA_POS, GX_TEX_S, GX_RGBA6, 0);
-    GXSetVtxAttrFmt(GX_VTXFMT0, GX_VA_CLR0, GX_TEX_ST, GX_RGBA8, 0);
-    GXSetVtxDesc(GX_VA_POS, GX_DIRECT);
-    GXSetVtxDesc(GX_VA_CLR0, GX_DIRECT);
-    HSD_StateSetCullMode(0);
-}
-
-static inline int hexval(int ch)
-{
-    if (ch < 0x47) {
-        if (ch < 0x0A) {
-            if (ch >= 0) {
-                goto sub_30;
-            }
-            goto ret_zero;
-        }
-        if (ch >= 0x41) {
-            goto sub_37;
-        }
-        goto ret_zero;
-    }
-    if (ch < 0x67) {
-        if (ch >= 0x61) {
-            goto sub_57;
-        }
-        goto ret_zero;
-    }
-    goto ret_zero;
-sub_30:
-    return ch - 0x30;
-sub_57:
-    return ch - 0x57;
-sub_37:
-    return ch - 0x37;
-ret_zero:
-    return 0;
-}
-
-// @TODO: Currently 98.73% match - hexval branch condition inversion
-s32 hsd_80391AC8(char* str, GXColor* color, f32 x, f32 y)
-{
-    GXColor col;
-    GXColor* col_ptr;
-    s32 i;
-    u8* p;
-
-    if (str == NULL) {
-        return (s32) x;
-    }
-    col_ptr = &col;
-    for (i = 0; (s8) str[i] != 0; i++) {
-        if ((s8) str[i] != '\\') {
-            *col_ptr = *color;
-            x += DrawASCII((s8) str[i], x, y, col_ptr);
-        } else {
-            switch ((s8) str[++i]) {
-            case 'c':
-            case 'C':
-                p = (u8*) (str + i);
-                color->r = (u8) (hexval(p[2]) + hexval(p[1]) * 16);
-                color->g = (u8) (hexval(p[4]) + hexval(p[3]) * 16);
-                color->b = (u8) (hexval(p[6]) + hexval(p[5]) * 16);
-                i += 6;
-                color->a = 0xFF;
-                break;
-            }
-        }
-    }
-    return (s32) x;
-}
-
-void hsd_80391E18(u8* list, f32 x1, f32 y1, f32 x2, f32 y2)
-{
-    f32 dx;
-    f32 dy;
-    f64 t;
-    f32 prev_x;
-    f32 prev_y;
-    u32 color;
-    u8 r, g, b, a;
-    PAD_STACK(8);
-
-    if (list == NULL) {
-        return;
-    }
-
-    prev_x = x1;
-    prev_y = y1;
-    dx = x2 - x1;
-    dy = y2 - y1;
-
-    for (;;) {
-        t = *(f32*) list;
-        color = *(u32*) (list + 4);
-        list += 8;
-
-        if (t < 0.0) {
-            break;
-        }
-
-        GXBegin(0xA8, 0, 2);
-
-        GXWGFifo.f32 = prev_x;
-        GXWGFifo.f32 = prev_y;
-
-        prev_x = t * dx + x1;
-        prev_y = t * dy + y1;
-
-        r = ((u8*) &color)[0];
-        g = ((u8*) &color)[1];
-        b = ((u8*) &color)[2];
-        a = ((u8*) &color)[3];
-
-        GXWGFifo.u8 = r;
-        GXWGFifo.u8 = g;
-        GXWGFifo.u8 = b;
-        GXWGFifo.u8 = a;
-
-        GXWGFifo.f32 = prev_x;
-        GXWGFifo.f32 = prev_y;
-        GXWGFifo.u8 = r;
-        GXWGFifo.u8 = g;
-        GXWGFifo.u8 = b;
-        GXWGFifo.u8 = a;
-    }
-}
-
-// @TODO: Currently 88.12% match - register allocation differences remain
-static inline void hsd_80391F28_calc_perp_and_begin(f32 dy, f32 dx,
-                                                    f32* perp_x, f32* perp_y)
-{
-    f32 len;
-
-    len = (dx * dx) + (dy * dy);
-    if (len > 0.0F) {
-        f64 e = __frsqrte(len);
-        e = 0.5 * e * -(len * (e * e) - 3.0);
-        e = 0.5 * e * -(len * (e * e) - 3.0);
-        e = 0.5 * e * -(len * (e * e) - 3.0);
-        e = len * e;
-        {
-            volatile f32 temp = (f32) e;
-            len = temp;
-        }
-    }
-    *perp_x = -dy / len;
-    *perp_y = -dx / len;
-
-    GXBegin(0xA8, 0, 2);
-}
-
-static inline void hsd_80391F28_calc_tick4_y(f32 perp_y, f32* tick4_y)
-{
-    *tick4_y = 4.0F * perp_y;
-}
-
-void hsd_80391F28(GXColor* color, f32 x1, f32 y1, f32 x2, f32 y2, f32 count)
-{
-    f32 tick6_x;
-    f32 dx;
-    f32 dy;
-    f32 perp_x;
-    f32 perp_y;
-    f32 tick6_y;
-    f32 tick4_x;
-    f32 tick4_y;
-    f32 step_y;
-    f32 step_x;
-    f32 len;
-    s32 i;
-    u8 r, g, b, a;
-
-    dy = y2 - y1;
-    dx = x2 - x1;
-    hsd_80391F28_calc_perp_and_begin(dy, dx, &perp_x, &perp_y);
-
-    step_y = dy / count;
-    step_x = dx / count;
-
-    GXWGFifo.f32 = x1;
-    i = 0;
-    a = color->a;
-    GXWGFifo.f32 = y1;
-
-    b = color->b;
-    g = color->g;
-    r = color->r;
-
-    tick6_y = 6.0F * perp_y;
-    {
-        f32 tmp_tick6_x = 6.0F * perp_x;
-        tick6_x = tmp_tick6_x;
-    }
-    tick4_x = 4.0F * perp_x;
-    hsd_80391F28_calc_tick4_y(perp_y, &tick4_y);
-
-    GXWGFifo.u8 = r;
-    GXWGFifo.u8 = g;
-    GXWGFifo.u8 = b;
-    GXWGFifo.u8 = a;
-
-    GXWGFifo.f32 = x2;
-    GXWGFifo.f32 = y2;
-    GXWGFifo.u8 = r;
-    GXWGFifo.u8 = g;
-    GXWGFifo.u8 = b;
-    GXWGFifo.u8 = a;
-
-    while ((f32) i <= count) {
-        GXBegin(0xA8, 0, 2);
-        GXWGFifo.f32 = x1;
-        GXWGFifo.f32 = y1;
-        a = color->a;
-        b = color->b;
-        g = color->g;
-        r = color->r;
-        GXWGFifo.u8 = r;
-        GXWGFifo.u8 = g;
-        GXWGFifo.u8 = b;
-        GXWGFifo.u8 = a;
-        if ((i % 5) == 0) {
-            GXWGFifo.f32 = x1 + tick6_x;
-            GXWGFifo.f32 = y1 + tick6_y;
-        } else {
-            GXWGFifo.f32 = x1 + tick4_x;
-            GXWGFifo.f32 = y1 + tick4_y;
-        }
-        a = color->a;
-        b = color->b;
-        g = color->g;
-        r = color->r;
-        x1 += step_x;
-        y1 += step_y;
-        i++;
-        GXWGFifo.u8 = r;
-        GXWGFifo.u8 = g;
-        GXWGFifo.u8 = b;
-        GXWGFifo.u8 = a;
-    }
-}
-
-void hsd_80392194(u8* dst, s32 flags, void* unused1, void* unused2, u8* src)
-{
-    u8 b;
-    dst[0] = src[0];
-    if (flags & 1) {
-        b = src[2];
-    } else {
-        b = src[1];
-    }
-    dst[1] = b;
-}
-
-extern GlyphEntry lbl_80408898[4];
-
-// @TODO: Currently 72.88% match - needs register allocation fixes
-void hsd_803921B8(void* bitmap, s32 x, s32 y, s32 dst, s32 w, s32 h,
-                  s32 stride, void* tbl)
-{
-    volatile s32 v_stride;
-    s32 bit_x;
-    s32 cur_dst;
-    u32 max_x;
-    s32 off_x;
-    s32 off_y;
-    GlyphEntry* table;
-    s32 bit_off;
-    s32 x2;
-    s32 data_off;
-    u32 word;
-    s32 shift;
-    s32 val;
-    u8* bmp;
-    GlyphEntry* entry;
-
-    v_stride = stride;
-    bmp = bitmap;
-    off_y = 0;
-    off_x = 0;
-    if ((table = tbl) == NULL) {
-        table = lbl_80408898;
-    }
-    if (y < 0) {
-        off_y = -y;
-        y = 0;
-    }
-    max_x = (u32) (x + 11) < w ? x + 11 : w;
-    if (x < 0) {
-        off_x = -x;
-        x = 0;
-    }
-    x2 = x * 2;
-    data_off = off_y * 4;
-    while (off_y < 14 && (u32) y < h) {
-        s32 col;
-        col = x;
-        bit_x = off_x;
-        cur_dst = dst + (y * v_stride) + x2;
-        while ((u32) col < max_x) {
-            word =
-                *(u32*) (bmp + data_off + (((u32) bit_x >> 2) & 0x3FFFFFFC));
-            shift = cur_dst;
-            bit_off = bit_x & 0xF;
-            while (bit_off < 16 && (u32) col < max_x) {
-                val = (word >> ((15 - bit_off) * 2)) & 3;
-                entry = &table[val];
-                entry->callback((void*) shift, col, y, val, entry);
-                bit_off++;
-                bit_x++;
-                shift += 2;
-                cur_dst += 2;
-                col++;
-            }
-        }
-        off_y++;
-        data_off += 4;
-        y++;
-    }
-}
-
-// @TODO: Currently 85% match - needs register allocation fixes
-void hsd_803922FC(void* bitmap, s32 x, s32 y, s32 parity, s32 dst, s32 w,
-                  s32 h, s32 stride, void* tbl)
-{
-    volatile s32 v_stride;
-    s32 bit_x;
-    s32 cur_dst;
-    u32 max_x;
-    s32 off_x;
-    s32 off_y;
-    s32 bit_off;
-    s32 x2;
-    s32 data_off;
-    s32 row_idx;
-    u32 word;
-    s32 shift;
-    s32 val;
-    u8* bmp;
-    GlyphEntry* entry;
-
-    bmp = bitmap;
-    off_y = 0;
-    off_x = 0;
-    if (tbl == NULL) {
-        tbl = lbl_80408898;
-    }
-    v_stride = stride;
-    if (y < 0) {
-        off_y = -y;
-        y = 0;
-    }
-    if (y % 2 != parity) {
-        y++;
-        off_y++;
-    }
-    max_x = (u32) (x + 11) < w ? x + 11 : w;
-    if (x < 0) {
-        off_x = -x;
-        x = 0;
-    }
-    row_idx = y / 2;
-    x2 = x * 2;
-    data_off = off_y * 4;
-    while (off_y < 14 && (u32) y < h) {
-        s32 col;
-        col = x;
-        bit_x = off_x;
-        cur_dst = dst + (row_idx * v_stride) + x2;
-        while ((u32) col < max_x) {
-            word =
-                *(u32*) (bmp + data_off + (((u32) bit_x >> 2) & 0x3FFFFFFC));
-            shift = cur_dst;
-            bit_off = bit_x & 0xF;
-            while (bit_off < 16 && (u32) col < max_x) {
-                val = (word >> ((15 - bit_off) * 2)) & 3;
-                entry = &((GlyphEntry*) tbl)[val];
-                entry->callback((void*) shift, col, y, val, entry);
-                bit_off++;
-                bit_x++;
-                shift += 2;
-                cur_dst += 2;
-                col++;
-            }
-        }
-        off_y += 2;
-        data_off += 8;
-        row_idx++;
-        y += 2;
-    }
-}
-
-void hsd_80392474(void)
-{
-    hsd_804D7850 = NULL;
-}
-
-HSD_SList* fn_80392480(Event event, int priority)
-{
-    HSD_SList* prev = NULL;
-    HSD_SList* cur = hsd_804D7850;
-
-    goto loop_5;
-block_1: {
-    HSD_SList* ret = cur->data;
-    if (event != (Event) ret->next) {
-        if (((struct EventPriority*) ret)->priority <= priority) {
-            prev = cur;
-        }
-        cur = cur->next;
-    loop_5:
-        if (cur != NULL) {
-            goto block_1;
-        }
-        {
-            struct EventPriority* data = HSD_MemAlloc(8);
-            data->event = event;
-            data->priority = priority;
-            if (prev != NULL) {
-                return HSD_SListAllocAndAppend(prev, data);
-            }
-            ret = HSD_SListAllocAndPrepend(hsd_804D7850, data);
-        }
-        hsd_804D7850 = ret;
-    }
-    return ret;
-}
-}
-
-#pragma push
-#pragma dont_inline on
-void hsd_80392528(Event event)
-{
-    fn_80392480(event, 0x80);
-}
-#pragma pop
-
-static GXColor lbl_804D6080;
-static u32 lbl_804D6084;
-static const u32 lbl_804DE8E0 = 0xFFFFFFFF;
-
-// @TODO: Currently 89.78% match - needs minor control flow and register fixes
-void hsd_8039254C(void)
-{
-    f32 line;
-    f32 bar_y;
-    f32 bar_x;
-    f32 t2;
-    GXColor default_col;
-    GXColor bg_col0;
-    GXColor bg_col1;
-    GXColor txt_col;
-    GXColor bg_col2;
-    GXColor bg_col3;
-    GXColor bar_col;
-    s32 char_count;
-    s32 count;
-    GXColor* p_bg_col0;
-    GXColor* p_bg_col1;
-    GXColor* p_txt_col;
-    GXColor* p_bg_col2;
-    GXColor* p_bg_col3;
-    GXColor* p_bar_col;
-    s32 col_pos;
-    s32 first;
-    HSD_SList* event_node;
-
-    p_bar_col = &bar_col;
-    p_bg_col3 = &bg_col3;
-    p_bg_col2 = &bg_col2;
-    p_bg_col1 = &bg_col1;
-    p_txt_col = &txt_col;
-    p_bg_col0 = &bg_col0;
-    col_pos = 60;
-    first = 1;
-    line = 1.0F;
-    event_node = (HSD_SList*) hsd_804D7850;
-    default_col = *(GXColor*) &lbl_804DE8E0;
-
-    while (event_node != NULL) {
-        DispItem* item;
-        DispCallback cb;
-        cb = (DispCallback) ((void**) event_node->data)[0];
-        item = cb(event_node->data);
-
-        while (item != NULL) {
-            if (first != 0) {
-                if (lbl_804D6080.a != 0) {
-                    hsd_80391A04(10.0F, 10.0F, 6);
-                    bg_col0 = lbl_804D6080;
-                    DrawRectangle(-10.0F, 5.0F, 620.0F, 10.0F, p_bg_col0);
-                }
-                first = 0;
-            }
-            switch (item->type) {
-            case 0:
-                if ((&item->content) == NULL) {
-                    char_count = 0;
-                } else {
-                    char_count = 0;
-                    {
-                        s32 j = 0;
-                        s32 len = strlen(item->content.text);
-                        char_count = j;
-                        while (j < len) {
-                            if ((s8) item->content.text[j] != '\\') {
-                                char_count++;
-                            } else {
-                                j++;
-                                {
-                                    char* tmp = item->content.text;
-                                    switch ((s8) tmp[j]) {
-                                    case 'c':
-                                    case 'C':
-                                        j += 6;
-                                        break;
-                                    }
-                                }
-                            }
-                            j++;
-                        }
-                    }
-                }
-                if (col_pos + char_count > 60) {
-                    line -= 1.0F;
-                    col_pos = 0;
-                    if (lbl_804D6080.a != 0) {
-                        bg_col1 = lbl_804D6080;
-                        DrawRectangle(-10.0F, (10.0F * line) - 5.0F, 620.0F,
-                                      10.0F, p_bg_col1);
-                    }
-                }
-                hsd_80391A04(10.0F, 10.0F, 6);
-                txt_col = default_col;
-                hsd_80391AC8(item->content.text, p_txt_col,
-                             (f32) (col_pos * 10), 10.0F * line);
-                col_pos = col_pos + (2 + char_count);
-                break;
-            case 2:
-                if (col_pos != 0) {
-                    line = (f32) ((f64) line - 0.5);
-                    if (lbl_804D6080.a != 0) {
-                        bg_col2 = lbl_804D6080;
-                        DrawRectangle(-10.0F, (10.0F * line) - 5.0F, 620.0F,
-                                      5.0F, p_bg_col2);
-                    }
-                }
-                hsd_80391A04(10.0F, 10.0F, 24);
-                t2 = (10.0F * line) + 2.0F;
-                hsd_80391E18(item->content.gradient, 0.0F, t2, 600.0F, t2);
-                col_pos = 60;
-                break;
-            case 1: {
-                DispItem* bar_ptr;
-                DispItem* bar_draw_ptr;
-
-                bar_ptr = item;
-                char_count = 0;
-                {
-                    while ((count = bar_ptr->content.bars[0].count) > 0) {
-                        char_count += count;
-                        bar_ptr = (DispItem*) ((DispBar*) bar_ptr + 1);
-                    }
-                }
-                if (char_count > 0) {
-                    if (col_pos != 0) {
-                        line = (f32) ((f64) line - 0.5);
-                        if (lbl_804D6080.a != 0) {
-                            bg_col3 = lbl_804D6080;
-                            DrawRectangle(-10.0F, (10.0F * line) - 5.0F,
-                                          620.0F, 5.0F, p_bg_col3);
-                        }
-                    }
-                    hsd_80391A04(10.0F, 10.0F, 12);
-                    bar_y = (10.0F * line) + 2.0F;
-                    bar_x = 0.0F;
-                    {
-                        bar_draw_ptr = item;
-                        while ((count = bar_draw_ptr->content.bars[0].count) >
-                               0)
-                        {
-                            f32 prev_x;
-                            prev_x = bar_x;
-                            bar_col =
-                                *(GXColor*) &bar_draw_ptr->content.bars[0]
-                                     .color;
-                            bar_x += (600.0F / (f32) char_count) * (f32) count;
-                            hsd_80391F28(
-                                p_bar_col, prev_x, bar_y, bar_x, bar_y,
-                                (f32) bar_draw_ptr->content.bars[0].count);
-                            bar_draw_ptr =
-                                (DispItem*) ((DispBar*) bar_draw_ptr + 1);
-                        }
-                    }
-                    col_pos = 60;
-                }
-                break;
-            }
-            }
-            item = item->next;
-        }
-        event_node = event_node->next;
-    }
-}
-
-void fn_80392934(void)
-{
-    f32 cpu;
-    f32 draw;
-    f32 total;
-
-    cpu = HSD_PerfLastStat.cpu_time;
-    hsd_804D7858 = cpu;
-    draw = HSD_PerfLastStat.draw_time;
-    hsd_804D785C = draw;
-    total = HSD_PerfLastStat.total_time;
-    hsd_804D7860 = total;
-
-    if (hsd_804D7864 < cpu || hsd_804D7870-- < 0) {
-        hsd_804D7864 = cpu;
-        hsd_804D7870 = 60;
-    }
-    if (hsd_804D7868 < draw || hsd_804D7874-- < 0) {
-        hsd_804D7868 = draw;
-        hsd_804D7874 = 60;
-    }
-    if (hsd_804D786C < total || hsd_804D7878-- < 0) {
-        hsd_804D786C = total;
-        hsd_804D7878 = 60;
-    }
-    if (hsd_804D787C < cpu) {
-        hsd_804D787C = cpu;
-    }
-    if (hsd_804D7880 < draw) {
-        hsd_804D7880 = draw;
-    }
-    if (hsd_804D7884 < total) {
-        hsd_804D7884 = total;
-    }
-}
-
-static s32 lbl_804D6088 = 4;
-static s32 lbl_804D608C = 1;
-
-#pragma push
-#pragma dont_inline on
-void fn_80392A08(int mode, int scale, int enable)
-{
-    lbl_804D6088 = mode;
-    lbl_804D608C = scale;
-    if (hsd_804D7888 == 0 && enable != 0) {
-        hsd_804D787C = 0.0F;
-        hsd_804D7880 = 0.0F;
-        hsd_804D7884 = 0.0F;
-    }
-    hsd_804D7888 = enable;
-}
-#pragma pop
-
-static s32 lbl_804D6090 = -1;
-static s32 lbl_804D6094 = (s32) 0xFF0000FF;
-static s32 lbl_804D6098 = 0x00FF00FF;
-static s32 lbl_804D609C = 0x00FFFFFF;
-static s32 lbl_804D60A0 = 0x8080FF;
-static s32 lbl_804D60A4 = (s32) 0xC0C000FF;
-
-// @TODO: Currently 89.63% match - needs register allocation and expression
-// fixes
-static inline PerfDispItem* get_perf_disp_item(s32 count)
-{
-    return &hsd_804CE3F8[count];
-}
-
-void* fn_80392A3C(void)
-{
-    volatile s32 green;
-    s32 numFrames;
-    s32 bar_count;
-    s32 count;
-    PerfDispItem* entry;
-    PerfDispItem* entry2;
-
-    count = 0;
-    numFrames = lbl_804D6088;
-    if (0 != numFrames) {
-        u8* counts = &hsd_804CE3F8[0].content.bytes[0];
-        u8* colors_base = &hsd_804CE3F8[0].content.bytes[4];
-        u8* colors = colors_base;
-        hsd_804CE3F8[0].type = 1;
-        *(s32*) counts = 1;
-        bar_count = count;
-        bar_count++;
-        green = lbl_804D6098;
-        *(s32*) colors = green;
-        if (numFrames > 1) {
-            s32 val = numFrames - 1;
-            if (val > 3) {
-                val = 3;
-            }
-            *(s32*) (counts + bar_count * 8) = val;
-            *(s32*) (colors + bar_count * 8) = lbl_804D60A4;
-            bar_count = 2;
-        }
-        if (numFrames > 4) {
-            *(s32*) (counts + bar_count * 8) = numFrames - 4;
-            *(s32*) (colors + bar_count * 8) = lbl_804D6094;
-            bar_count++;
-        }
-        *(s32*) (counts + bar_count * 8) = -1;
-        hsd_804CE3F8[0].next = &hsd_804CE3F8[1];
-        hsd_804CE3F8[1].type = 2;
-        count = 4;
-        hsd_804CE3F8[1].content.gradient[0].pos =
-            hsd_804D7860 / (f32) numFrames;
-        hsd_804CE3F8[1].content.gradient[0].color = lbl_804D609C;
-        hsd_804CE3F8[1].content.gradient[1].pos =
-            (f32) (s32) (0.9999F + hsd_804D7860) / (f32) numFrames;
-        hsd_804CE3F8[1].content.gradient[1].color = lbl_804D60A0;
-        hsd_804CE3F8[1].content.gradient[2].pos = -1.0F;
-        hsd_804CE3F8[1].next = &hsd_804CE3F8[2];
-        hsd_804CE3F8[2].type = 2;
-        hsd_804CE3F8[2].content.gradient[0].pos =
-            hsd_804D785C / (f32) numFrames;
-        hsd_804CE3F8[2].content.gradient[0].color = lbl_804D6090;
-        hsd_804CE3F8[2].content.gradient[1].pos = -1.0F;
-        hsd_804CE3F8[2].next = &hsd_804CE3F8[3];
-        hsd_804CE3F8[3].type = 2;
-        hsd_804CE3F8[3].content.gradient[0].pos =
-            hsd_804D7858 / (f32) numFrames;
-        hsd_804CE3F8[3].content.gradient[0].color = green;
-        hsd_804CE3F8[3].content.gradient[1].pos = -1.0F;
-        hsd_804CE3F8[3].next = &hsd_804CE3F8[4];
-    }
-    if (lbl_804D608C != 0) {
-        hsd_804CE3F8[count].type = 0;
-        entry = get_perf_disp_item(count);
-        sprintf(entry->content.text,
-                "\\c00ff00%2.3f \\cffffff%2.3f \\c00ffff%2.3f  "
-                "\\c00ff00%2.3f \\cffffff%2.3f \\c00ffff%2.3f",
-                hsd_804D7858, hsd_804D785C, hsd_804D7860, hsd_804D7864,
-                hsd_804D7868, hsd_804D786C);
-        entry->next = &hsd_804CE3F8[count + 1];
-        count++;
-        if (hsd_804D7888 != 0) {
-            hsd_804CE3F8[count].type = 0;
-            entry2 = &hsd_804CE3F8[count];
-            sprintf(entry2->content.text,
-                    "\\c00ff00%2.3f \\cffffff%2.3f \\c00ffff%2.3f",
-                    hsd_804D787C, hsd_804D7880, hsd_804D7884);
-            entry2->next = &hsd_804CE3F8[count + 1];
-            count++;
-        }
-    }
-    if (count == 0) {
-        return NULL;
-    }
-    hsd_804CE3F8[count - 1].next = NULL;
-    return hsd_804CE3F8;
-}
-
-s32 fn_80392CCC(s32 channel)
-{
-    hsd_804D7894 = channel;
-    return 0;
-}
-
-u8 fn_80392CD8(char* caller)
-{
-    char* msg;
-    u8 err;
-
-    err = MCCGetLastError();
-
-    switch (err) {
-    case 0:
-        return err;
-    case 1:
-        msg = "MCC is no initialize";
-        break;
-    case 2:
-        msg = "No responce";
-        break;
-    case 3:
-        msg = "PING error";
-        break;
-    case 4:
-        msg = "Could not initialize HIO";
-        break;
-    case 5:
-        msg = "Could not read HIO mailbox";
-        break;
-    case 6:
-        msg = "Could not write HIO mailbox";
-        break;
-    case 7:
-        msg = "Could not read HIO memory";
-        break;
-    case 8:
-        msg = "Could not write HIO memory";
-        break;
-    case 9:
-        msg = "Could not read HIO status";
-        break;
-    case 10:
-        msg = "Could not flush channelInfo";
-        break;
-    case 11:
-        msg = "Could not load channelInfo";
-        break;
-    case 12:
-        msg = "Not enough memory block";
-        break;
-    case 13:
-        msg = "Invalid function parameter";
-        break;
-    case 14:
-        msg = "Invalid channel parameter";
-        break;
-    case 15:
-        msg = "Invalid data size";
-        break;
-    case 16:
-        msg = "Invalid offset parameter";
-        break;
-    case 17:
-        msg = "Channel was (already) opened";
-        break;
-    case 18:
-        msg = "Channel was (already) closed";
-        break;
-    case 19:
-        msg = "Channel was (already) locked";
-        break;
-    case 20:
-        msg = "Channel was (already) unlocked";
-        break;
-    case 21:
-        msg = "Channel (read/write) busy";
-        break;
-    default:
-        msg = "Unknown error";
-        break;
-    }
-
-    if (caller != NULL) {
-        OSReport("%s: MCC Error, %s (%d)\n", caller, msg, err);
-    } else {
-        OSReport("MCC Error, %s (%d)\n", msg, err);
-    }
-    return err;
-}
-
-s32 hsd_804CE728[0x106];
-
-void fn_80392E2C(s32 event_type)
-{
-    s32 idx;
-
-    if (hsd_804D7898 <= 0x100 && (u32) (event_type - 1) <= 1U) {
-        idx = hsd_804D789C + hsd_804D7898;
-        hsd_804CE728[idx % 256] = event_type;
-        hsd_804D7898 += 1;
-    }
-}
-
-ParticleLogEntry hsd_804CEB40[0x100];
-extern s32 hsd_804CF740[42];
-
-extern s32 hsd_804D78A8;
-extern s32 hsd_804D78AC;
-extern int hsd_804D78A0;
-
-// @TODO: Currently 92.84% match - needs minor register allocation fix
-void hsd_80392E80(void)
-{
-    enum MCC_CONNECT status;
-    s32 event;
-    s32 head;
-    s32 startTick;
-    s32 waiting;
-    s32 intr;
-    s32* channel_flags;
-    PAD_STACK(12);
-
-    for (;;) {
-        intr = OSDisableInterrupts();
-        if (hsd_804D7898 == 0) {
-            OSRestoreInterrupts(intr);
-            break;
-        }
-
-        head = hsd_804D789C + 1;
-        event = hsd_804CE728[hsd_804D789C];
-        hsd_804D7898 -= 1;
-        hsd_804D789C = head % 256;
-        OSRestoreInterrupts(intr);
-
-        switch (event) {
-        case 2:
-            OSReport("Connecting...\n");
-            hsd_804D78AC = 0;
-            hsd_804D78A8 = 0;
-            OSReport("FIO Init...");
-            if (FIOInit(hsd_804D7890, 1, 10) == 0) {
-                OSReport("NG\n");
-            } else {
-                OSReport("OK\n");
-                hsd_804D78A0 = 1;
-            }
-            OSReport("MCC Open...");
-            if (MCCOpen(0xF, 1, (MCC_CBEvent) fn_803932D0) == 0) {
-                OSReport("NG\n");
-                OSReport("Reseting...\n");
-                FIOExit();
-                hsd_804D78A0 = 0;
-                MCCExit();
-                if (MCCInit(hsd_804D7890, 0, (MCC_CBSysEvent) fn_80392E2C) ==
-                    0)
-                {
-                    OSReport("MCCInit Failed.\n");
-                }
-            } else {
-                OSReport("Waiting for connection...");
-                waiting = 1;
-                startTick = OSGetTick();
-                for (;;) {
-                    if (MCCGetConnectionStatus(0xF, &status) != 0 &&
-                        status == 3)
-                    {
-                        waiting = 0;
-                    } else if ((u32) ((u32) (OSGetTick() - startTick) /
-                                      OS_TIMER_CLOCK) < 3U)
-                    {
-                        continue;
-                    }
-                    break;
-                }
-                if (waiting != 0) {
-                    OSReport("NG\n");
-                    OSReport("Reseting...\n");
-                    FIOExit();
-                    hsd_804D78A0 = 0;
-                    MCCExit();
-                    if (MCCInit(hsd_804D7890, 0,
-                                (MCC_CBSysEvent) fn_80392E2C) == 0)
-                    {
-                        OSReport("MCCInit Failed.\n");
-                    }
-                } else {
-                    OSReport("OK\n");
-                    channel_flags = (s32*) hsd_804CF740;
-                    channel_flags[0] = 0;
-                    channel_flags[1] = 0;
-                    channel_flags[2] = 0;
-                    channel_flags[3] = 0;
-                    channel_flags[4] = 0;
-                    channel_flags[5] = 0;
-                    channel_flags[6] = 0;
-                    channel_flags[7] = 0;
-                    channel_flags[8] = 0;
-                    channel_flags[9] = 0;
-                    channel_flags[10] = 0;
-                    channel_flags[11] = 0;
-                    channel_flags[12] = 0;
-                    channel_flags[13] = 0;
-                    channel_flags[14] = 0;
-                    channel_flags[15] = 0;
-                    channel_flags[0] = 1;
-                    channel_flags[8] = 1;
-                    channel_flags[15] = 1;
-                }
-            }
-            break;
-        case 1:
-            OSReport("Disconnecting...\n");
-            if (hsd_80393328() == 0) {
-                OSReport("Disconnection failed.\n");
-                OSReport("Reseting...\n");
-                FIOExit();
-                hsd_804D78A0 = 0;
-                MCCExit();
-                if (MCCInit(hsd_804D7890, 0, (MCC_CBSysEvent) fn_80392E2C) ==
-                    0)
-                {
-                    OSReport("MCCInit Failed.\n");
-                }
-            } else {
-                OSReport("Disconnection OK.\n");
-            }
-            break;
-        }
-    }
-    hsd_80393844();
-}
-
-bool hsd_803931A4(s32 exi_channel)
-{
-    s32 channel;
-    s32* channel_flags;
-    PAD_STACK(16);
-
-    channel_flags = (s32*) hsd_804CF740;
-    channel_flags[0] = 0;
-    channel_flags[1] = 0;
-    channel_flags[2] = 0;
-    channel_flags[3] = 0;
-    channel_flags[4] = 0;
-    channel_flags[5] = 0;
-    channel_flags[6] = 0;
-    channel_flags[7] = 0;
-    channel_flags[8] = 0;
-    channel_flags[9] = 0;
-    channel_flags[10] = 0;
-    channel_flags[11] = 0;
-    channel_flags[12] = 0;
-    channel_flags[13] = 0;
-    channel_flags[14] = 0;
-    channel_flags[15] = 0;
-    channel_flags[0] = 1;
-    channel_flags[8] = 1;
-    channel_flags[15] = 1;
-
-    channel = exi_channel;
-
-    if (channel < 0) {
-        hsd_804D7894 = -1;
-        MCCEnumDevices((MCC_CBEnumDevices) fn_80392CCC);
-        channel = hsd_804D7894;
-        if ((s32) hsd_804D7894 < 0) {
-            channel = 0;
-        }
-    }
-
-    hsd_804D7890 = channel;
-    hsd_80393A54(1);
-    hsd_804D78A0 = 0;
-    hsd_804D7898 = 0;
-    hsd_804D789C = 0;
-    hsd_804D78BC = 0;
-    hsd_804D78B8 = 0;
-
-    if (MCCInit(hsd_804D7890, 0, (MCC_CBSysEvent) fn_80392E2C) == 0) {
-        OSReport("EXI_%d:not support USB (Error=%d)\n", channel,
-                 MCCGetLastError());
-        return 0;
-    }
-
-    OSReport("EXI initialized (EXI_%d)\n", channel);
-    return 1;
-}
-
-s32 hsd_804CF740[42];
-
-void fn_803932D0(s32 type, u32 flags, s32 value)
-{
-    s32 count = hsd_804D78BC;
-    ParticleLogEntry* base = hsd_804CEB40;
-    s32 idx;
-    ParticleLogEntry* entry;
-
-    if (count <= 0x100 && flags == 0x100U) {
-        idx = (hsd_804D78B8 + count) % 256;
-        *(s32*) ((u8*) base + idx * (s32) sizeof(ParticleLogEntry)) = type;
-        entry = &base[idx];
-        entry->x4 = flags;
-        entry->x8 = value;
-        hsd_804D78BC += 1;
-    }
-}
-
-s32 hsd_80393328(void)
-{
-    u32 ticksPerUnit;
-    s32 startTick;
-    s32 result;
-    u8 err;
-
-    if (hsd_804D78B4 != 0) {
-        return 0;
-    }
-    hsd_804D78B4 = 1;
-    hsd_804D78B0 = HSD_Rand() % 255;
-    if (MCCNotify(0xF, hsd_804D78B0 + 256) == 0) {
-        err = MCCGetLastError();
-        OSReport("Error(0x%x) in MCCNotify.\n", err);
-    }
-    result = 1;
-    startTick = OSGetTick();
-    ticksPerUnit = (*(u32*) 0x800000F8 >> 2) & 0xFFFFFFFFFFFFFFFF;
-    do {
-        hsd_80392E80();
-        if (hsd_804D78B0 == 0) {
-            result = 0;
-            break;
-        }
-    } while ((u32) (OSGetTick() - startTick) / ticksPerUnit < 3);
-
-    if (result != 0) {
-        hsd_804D78B4 = 0;
-        return 0;
-    }
-    hsd_804D78B4 = 0;
-    return 1;
-}
-
-static void (*lbl_8040A93C[32])(void*, void*) = {
-    (void (*)(void*, void*)) hsd_80393440,
-    (void (*)(void*, void*)) hsd_80393840,
-};
-
-// Evidence: retail hsd_80393A5C loads this table's base into a register
-// (lis/addi 0x8040A9D0) and reads the messages at field displacements
-// +0x20 ("cannot open"), +0x90 ("Done"), +0xB8 ("cannot save"); the
-// remaining messages fill out the retail 0x130 object (see
-// struct ParticleUsbMessages in particle.static.h).
-/* 40A9D0 */ static struct ParticleUsbMessages psUsbMessages = {
-    "loading file (%s) from USB ...",
-    "cannot open file\n",
-    "cannot get files stat\n",
-    "too large file size\n",
-    "cannot allocate memory\n",
-    "cannot load file\n",
-    "Done %s size:%d time:%f spped:%fkbps\n",
-    "cannot save file\n",
-    "Seaching files (%s) via USB ...",
-    "cannot get dir stat\n",
-    "too large dir size\n",
-    "cannot use USB now.\n",
-};
-
-// @TODO: Currently 90.11% match - needs register allocation fix
-void hsd_80393440(void* request, void* response)
-{
-    u16 cmd;
-    s32 channel_mask;
-    s32 streaming;
-    s32 num_blocks;
-    s32 channel;
-    s32 i;
-    u8 free_blocks;
-    u8 err;
-    ParticleLogEntry* base;
-
-    cmd = ((u16*) request)[3];
-    channel_mask = cmd & 0xFF00;
-    base = hsd_804CEB40;
-
-    switch (channel_mask) {
-    case 0x100:
-        if (cmd & 0x80) {
-            streaming = 1;
-        } else {
-            streaming = 0;
-        }
-        num_blocks = cmd & 0xF;
-        for (channel = 2; channel < 16; channel++) {
-            if (((s32*) &base[0x100])[channel] == 0) {
-                break;
-            }
-        }
-
-        if (channel == 16) {
-            ((u16*) response)[3] = (u16) 0x8001;
-            MCCWrite(0xF, (hsd_804D78AC << 5) + 0x1000, response, 0x20, 0);
-            if (MCCNotify(0xF, hsd_804D78AC + 0x80) == 0) {
-                err = MCCGetLastError();
-                OSReport("Error(0x%x) in MCCNotify.\n", err);
-            }
-            hsd_804D78AC = (hsd_804D78AC + 1) % 128;
-            return;
-        }
-
-        free_blocks = MCCGetFreeBlocks(1);
-        if (free_blocks == 0) {
-            err = MCCGetLastError();
-            if (err != 0) {
-                OSReport("MCCError: %d\n", err);
-            }
-        }
-
-        if (num_blocks > (s32) free_blocks) {
-            ((u16*) response)[3] = (u16) (free_blocks + 0x8010);
-            MCCWrite(0xF, (hsd_804D78AC << 5) + 0x1000, response, 0x20, 0);
-            if (MCCNotify(0xF, hsd_804D78AC + 0x80) == 0) {
-                err = MCCGetLastError();
-                OSReport("Error(0x%x) in MCCNotify.\n", err);
-            }
-            hsd_804D78AC = (hsd_804D78AC + 1) % 128;
-            return;
-        }
-
-        ((s32*) &base[0x100])[channel] = 1;
-        ((u16*) response)[3] = channel;
-        MCCWrite(0xF, (hsd_804D78AC << 5) + 0x1000, response, 0x20, 0);
-        if (MCCNotify(0xF, hsd_804D78AC + 0x80) == 0) {
-            err = MCCGetLastError();
-            OSReport("Error(0x%x) in MCCNotify.\n", err);
-        }
-        hsd_804D78AC = (hsd_804D78AC + 1) % 128;
-        if (streaming != 0) {
-            MCCStreamOpen(channel, num_blocks);
-        } else {
-            MCCOpen(channel, num_blocks, 0);
-        }
-        return;
-
-    case 0x200:
-        i = cmd & 0xF;
-        if (((s32*) &base[0x100])[i] != 1) {
-            ((u16*) response)[3] = (u16) 0x8002;
-            MCCWrite(0xF, (hsd_804D78AC << 5) + 0x1000, response, 0x20, 0);
-            if (MCCNotify(0xF, hsd_804D78AC + 0x80) == 0) {
-                err = MCCGetLastError();
-                OSReport("Error(0x%x) in MCCNotify.\n", err);
-            }
-            hsd_804D78AC = (hsd_804D78AC + 1) % 128;
-            return;
-        }
-        ((u16*) response)[3] = 0;
-        MCCWrite(0xF, (hsd_804D78AC << 5) + 0x1000, response, 0x20, 0);
-        if (MCCNotify(0xF, hsd_804D78AC + 0x80) == 0) {
-            err = MCCGetLastError();
-            OSReport("Error(0x%x) in MCCNotify.\n", err);
-        }
-        hsd_804D78AC = (hsd_804D78AC + 1) % 128;
-        ((s32*) &base[0x100])[i] = 0;
-        MCCClose(i);
-        return;
-    }
-}
-
-void hsd_80393840(void) {}
-
-void hsd_80393844(void)
-{
-    ParticleLogEntry* base = hsd_804CEB40;
-    s32 type;
-    u32 flags;
-    s32 value;
-    BOOL irq;
-    u8 err;
-
-    for (;;) {
-        irq = OSDisableInterrupts();
-        if (hsd_804D78BC == 0) {
-            OSRestoreInterrupts(irq);
-            return;
-        }
-
-        type = base[hsd_804D78B8].x0;
-        flags = base[hsd_804D78B8].x4;
-        value = base[hsd_804D78B8].x8;
-        hsd_804D78BC -= 1;
-        hsd_804D78B8 = (hsd_804D78B8 + 1) % 256;
-
-        OSRestoreInterrupts(irq);
-
-        if (flags != 0x100 || type != 0xF) {
-            continue;
-        }
-
-        if ((value & 0xFFFFFF00) == 0x100) {
-            if (MCCNotify(0xF, (u8) value | 0x200) == 0) {
-                err = MCCGetLastError();
-                OSReport("Error(0x%x) in MCCNotify.\n", err);
-            }
-        } else if ((value & 0xFFFFFF00) == 0x200) {
-            if ((u32) hsd_804D78B0 == (u8) value) {
-                hsd_804D78B0 = 0;
-            }
-        } else if ((value & 0xFFFFFF80) == 0x80) {
-            memset((u8*) &base[0x10A].x8, 0, 0x20);
-            hsd_804D78A8 = value & 0x7F;
-            if (MCCRead(0xF, hsd_804D78A8 << 5, (u8*) &base[0x10A].x8, 0x20,
-                        0) != 0 &&
-                !((MCCPacket*) &base[0x10A].x8)->x4_b7)
-            {
-                u8 cmd;
-                memset((u8*) &base[0x105].x4, 0, 0x20);
-                ((MCCPacket*) &base[0x105].x4)->x4_b7 = 1;
-                *(s32*) &base[0x105].x4 = *(s32*) &base[0x10A].x8;
-                cmd = ((u8*) &base[0x10A].x8)[5];
-                if (((u8*) &base[0x105].x4)[5] = cmd, cmd < 0x20U) {
-                    if (lbl_8040A93C[cmd] != NULL) {
-                        lbl_8040A93C[cmd]((u8*) &base[0x10A].x8,
-                                          (u8*) &base[0x105].x4);
-                    }
-                }
-            }
-        }
-    }
-}
-
-bool hsd_80393A04(void)
-{
-    if (hsd_804D78A0 == 0) {
-        return 0;
-    }
-    if (FIOQuery() == 0) {
-        hsd_804D78A0 = 0;
-        return 0;
-    }
-    return 1;
-}
-
-void hsd_80393A54(int level)
-{
-    hsd_804D78C0 = level;
-}
-
-static inline f32 kbps_scale(void)
-{
-    return 1.0F / 1024.0F;
-}
-
-int hsd_80393A5C(char* filename, int data, int size)
-{
-    int ready;
-    u32 start;
-    int fd;
-    // Evidence: retail materializes the fd argument (mr r3) before the
-    // buffer argument (mr r4) at the FIOFwrite call; the fresh local copy
-    // reproduces that argument scheduling.
-    int fd_arg;
-    f32 written_f;
-    f32 elapsed;
-    u32* data_p;
-    struct ParticleUsbMessages* messages = &psUsbMessages;
-
-    if (hsd_804D78A0 == 0) {
-        ready = 0;
-    } else if (FIOQuery() == 0) {
-        hsd_804D78A0 = ready = 0;
-    } else {
-        ready = 1;
-    }
-
-    if (ready == 0) {
-        return -5;
-    }
-
-    if (*filename == '/') {
-        filename++;
-    }
-
-    start = OSGetTick();
-    fd = FIOFopen(filename, 0xA02);
-
-    if ((u32) (fd + 0x10000) == 0xFFFF) {
-        OSReport(messages->cannot_open);
-        return -1;
-    }
-
-    fd_arg = fd;
-    data_p = (u32*) data;
-    written_f = (f32) (u32) FIOFwrite(fd_arg, data_p, size);
-
-    if ((f32) (s32) size != written_f) {
-        OSReport(messages->cannot_save);
-        FIOFclose(fd);
-        return (s32) written_f;
-    }
-
-    FIOFclose(fd);
-    elapsed = (f32) (OSGetTick() - start) / (f32) (*(u32*) 0x800000F8 >> 2);
-    {
-        f32 bits = 8.0F * (f32) size;
-        OSReport(messages->done, filename, size, elapsed,
-                 bits / elapsed * kbps_scale());
-    }
-    return size;
-}
-
-void fn_80393C14(const u8* buf, size_t size)
-{
-    int i;
-
-    const u32 out_size = hsd_804CF7E8.buf_size;
-
-    int write_pos = hsd_804CF7E8.xC;
-    u8* out_buf = hsd_804CF7E8.out_buf;
-    u8 line_len = hsd_804CF7E8.x11;
-
-    for (i = 0; i < size; i++) {
-        const u32 tmp = out_size - 1;
-        switch (buf[i]) {
-        case '\r':
-            break;
-        case '\n':
-            if (line_len != 0 || out_buf[(write_pos + tmp) % out_size] != '\0')
-            {
-                out_buf[write_pos] = line_len;
-                write_pos = (write_pos + 1) % out_size;
-                hsd_804CF7E8.x1C++;
-                hsd_804CF7E8.x18++;
-                line_len = 0;
-            }
-            break;
-        default:
-            out_buf[write_pos] = buf[i];
-            if (++line_len == 0x36) {
-                write_pos = (write_pos + 1) % out_size;
-                out_buf[write_pos] = line_len;
-                hsd_804CF7E8.x1C++;
-                hsd_804CF7E8.x18++;
-                line_len = 0;
-            }
-            write_pos = (write_pos + 1) % out_size;
-            hsd_804CF7E8.x1C++;
-            break;
-        }
-    }
-    hsd_804CF7E8.xC = write_pos;
-    hsd_804CF7E8.x11 = line_len;
-}
-
-#pragma push
-#pragma dont_inline on
-s32 hsd_80393D2C(s32 enable)
-{
-    struct ParticleConsoleState* sp = &hsd_804CF7E8;
-    s32 old = sp->x0_b1;
-
-    if (enable != 0) {
-        sp->x0_b1 = 1;
-        HSD_SetReportCallback(fn_80393C14);
-    } else {
-        sp->x0_b1 = 0;
-        HSD_SetReportCallback(NULL);
-    }
-    return old;
-}
-#pragma pop
-
-#pragma push
-#pragma pool_data off
-void hsd_80393DA0(u8* buf, size_t size)
-{
-    PAD_STACK(4);
-    memset(&hsd_804CF7E8, 0, sizeof(hsd_804CF7E8) - sizeof(hsd_804CF7E8.x24));
-    hsd_804CF7E8.out_buf = buf;
-    hsd_804CF7E8.buf_size = size;
-    memset(buf, 0, size);
-    hsd_804CF7E8.x0_b0 = true;
-    hsd_804CF7E8.x0_b1 = true;
-    HSD_SetReportCallback(fn_80393C14);
-}
-#pragma pop
-
-#pragma push
-#pragma dont_inline on
-void hsd_80393E34(s32* col_out, s32* row_out)
-{
-    if (col_out != NULL) {
-        *col_out = hsd_804CF7E8.x14;
-    }
-    if (row_out != NULL) {
-        *row_out = hsd_804CF7E8.x18;
-    }
-}
-#pragma pop
-
-#pragma push
-#pragma dont_inline on
-void hsd_80393E68(u32 col, u32 row)
-{
-    struct ParticleConsoleState* sp = &hsd_804CF7E8;
-    u32 byte_val;
-    u32 buf_size;
-    u8* out_buf;
-    u32 sum;
-    u32 pos;
-    u32 counter;
-
-    if (!sp->x0_b0) {
-        return;
-    }
-
-    byte_val = sp->x11;
-    counter = 0;
-    buf_size = sp->buf_size;
-    out_buf = sp->out_buf;
-    sum = sp->xC + buf_size;
-    pos = byte_val;
-
-    while (counter < row && pos < buf_size) {
-        pos++;
-        byte_val = out_buf[(sum - pos) % buf_size];
-        counter++;
-        pos += byte_val;
-    }
-
-    sp->x18 = counter;
-    sp->x1C = pos;
-    sp->x20 = byte_val;
-
-    if (col < byte_val) {
-        sp->x14 = col;
-    } else {
-        sp->x14 = byte_val;
-    }
-}
-#pragma pop
-
-#pragma push
-#pragma dont_inline on
-void hsd_80393EF4(int col_delta, int row_delta)
-{
-    u32 byte_val;
-    u32 buf_size;
-    int* pos_ptr;
-    u32 counter;
-    u32 pos;
-    u32 sum;
-    u8* out_buf;
-    struct ParticleConsoleState* sp = &hsd_804CF7E8;
-
-    if (!sp->x0_b0) {
-        return;
-    }
-
-    buf_size = sp->buf_size;
-
-    if ((u32) * (pos_ptr = &sp->x1C) >= buf_size) {
-        hsd_80393E68(sp->x14 + col_delta, sp->x18 + row_delta);
-        return;
-    }
-
-    out_buf = sp->out_buf;
-    pos = *pos_ptr;
-    byte_val = sp->x20;
-    sum = sp->xC + buf_size;
-
-    if (row_delta > 0) {
-        counter = 0;
-
-        while (counter < (u32) row_delta && pos < buf_size) {
-            pos++;
-            byte_val = out_buf[(sum - pos) % buf_size];
-            counter++;
-            pos += byte_val;
-        }
-
-        sp->x18 += counter;
-        *pos_ptr = pos;
-
-        if (col_delta >= 0) {
-            int* col_ptr = &sp->x14;
-            *col_ptr += col_delta;
-            if ((u32) *col_ptr > byte_val) {
-                *col_ptr = byte_val;
-            }
-        } else {
-            int* col_ptr;
-            u32 col = *(col_ptr = &sp->x14);
-            col_delta = -col_delta;
-            if (col > (u32) col_delta) {
-                *col_ptr = col - (u32) col_delta;
-            } else {
-                *col_ptr = 0;
-            }
-        }
-    } else if (row_delta < 0) {
-        if (col_delta < 0 && (u32) sp->x14 < (u32) -col_delta) {
-            col_delta = 0;
-        } else {
-            col_delta = sp->x14 + col_delta;
-        }
-
-        if (row_delta < 0 && (u32) sp->x18 < (u32) -row_delta) {
-            row_delta = 0;
-        } else {
-            row_delta = sp->x18 + row_delta;
-        }
-
-        hsd_80393E68(col_delta, row_delta);
-    }
-}
-#pragma pop
-
-#pragma push
-#pragma pool_data off
-#pragma dont_inline on
-u8 hsd_80394068(void)
-{
-    u8 result;
-
-    if (!hsd_804CF7E8.x0_b0) {
-        return 0;
-    }
-    if ((u32) hsd_804CF7E8.x1C > hsd_804CF7E8.buf_size) {
-        return 0;
-    }
-    if ((u32) hsd_804CF7E8.x14 < (u32) hsd_804CF7E8.x20) {
-        result =
-            hsd_804CF7E8.out_buf[(hsd_804CF7E8.x14 +
-                                  (hsd_804CF7E8.xC + hsd_804CF7E8.buf_size -
-                                   hsd_804CF7E8.x1C)) %
-                                 hsd_804CF7E8.buf_size];
-        hsd_804CF7E8.x14++;
-    } else {
-        result = 0;
-        hsd_80393EF4(0, -1);
-        hsd_804CF7E8.x14 = 0;
-    }
-    return result;
-}
-#pragma pop
-
-#pragma push
-#pragma dont_inline on
-u8 hsd_80394128(s32 col, s32 row)
-{
-    struct {
-        u8 value;
-    } result;
-
-    hsd_80393E68(col, row);
-    if (!hsd_804CF7E8.x0_b0) {
-        result.value = 0;
-    } else if ((u32) hsd_804CF7E8.x1C > hsd_804CF7E8.buf_size) {
-        result.value = 0;
-    } else if ((u32) hsd_804CF7E8.x14 < (u32) hsd_804CF7E8.x20) {
-        result.value =
-            hsd_804CF7E8.out_buf[(hsd_804CF7E8.x14 +
-                                  (hsd_804CF7E8.xC + hsd_804CF7E8.buf_size -
-                                   hsd_804CF7E8.x1C)) %
-                                 hsd_804CF7E8.buf_size];
-        hsd_804CF7E8.x14++;
-    } else {
-        result.value = 0;
-        hsd_80393EF4(0, -1);
-        hsd_804CF7E8.x14 = result.value;
-    }
-    return result.value;
-}
-#pragma pop
-
-s32 hsd_803941E8(void* xfb_out_ptr, void* xfb_cur_ptr)
-{
-    s32* xfb_out = xfb_out_ptr;
-    u32* xfb_cur = xfb_cur_ptr;
-    s32 last_draw;
-    u8* vi_base;
-    s32 nb_xfb;
-    s32 i;
-    u32 buf;
-    PAD_STACK(8);
-
-    HSD_VIWaitXFBFlushNoYield();
-    last_draw = HSD_VIGetXFBLastDrawDone();
-
-    if (last_draw != -1) {
-        *xfb_cur = (u32) HSD_VIData.xfb[last_draw].buffer;
-    }
-
-    vi_base = (u8*) &HSD_VIData;
-    nb_xfb = HSD_VIData.nb_xfb;
-    for (i = 0; i < nb_xfb; i++) {
-        if (i == last_draw) {
-            goto next1;
-        }
-        buf = *(u32*) (vi_base + 0x58);
-        xfb_out[0] = buf;
-        if (buf != 0) {
-            break;
-        }
-    next1:
-        vi_base += 0x60;
-    }
-
-    i++;
-    vi_base = (u8*) &HSD_VIData + i * 0x60;
-    for (; i < nb_xfb; i++) {
-        if (i == last_draw) {
-            goto next2;
-        }
-        buf = *(u32*) (vi_base + 0x58);
-        xfb_out[1] = buf;
-        if (buf != 0) {
-            break;
-        }
-    next2:
-        vi_base += 0x60;
-    }
-
-    if ((u32) xfb_out[0] == 0) {
-        if (xfb_cur != NULL) {
-            xfb_out[0] = (s32) *xfb_cur;
-            *xfb_cur = 0;
-        } else {
-            return 0;
-        }
-    }
-
-    if ((u32) xfb_out[1] != 0) {
-        return 2;
-    }
-    return 1;
-}
-
-extern u8 lbl_804088B8[];
-
-void hsd_80394314(void)
-{
-    void* sp = &hsd_804CF810;
-
-    memset(sp, 0, 0xD8);
-    hsd_803941E8(&((struct ParticleScreenState*) sp)->x24,
-                 &((struct ParticleScreenState*) sp)->x2C);
-
-    {
-        s32 mode;
-        if (((struct ParticleScreenState*) sp)->x28 != 0) {
-            mode = 2;
-        } else {
-            mode = 1;
-        }
-        ((struct ParticleScreenState*) sp)->x38 = mode;
-    }
-
-    ((struct ParticleScreenState*) sp)->x34 = 0;
-    ((struct ParticleScreenState*) sp)->x30 = &HSD_VIData;
-    ((struct ParticleScreenState*) sp)->x3C =
-        ((u16*) ((struct ParticleScreenState*) sp)->x30)[2];
-    ((struct ParticleScreenState*) sp)->x40 =
-        ((u16*) ((struct ParticleScreenState*) sp)->x30)[4];
-    ((struct ParticleScreenState*) sp)->x44 =
-        (((u16) ((struct ParticleScreenState*) sp)->x3C + 15) * 2) & 0x1FFE0;
-    ((struct ParticleScreenState*) sp)->x48 =
-        ((struct ParticleScreenState*) sp)->x44 *
-        ((struct ParticleScreenState*) sp)->x40;
-    ((struct ParticleScreenState*) sp)->x4 = 0;
-    ((struct ParticleScreenState*) sp)->x8 =
-        ((struct ParticleScreenState*) sp)->x40;
-    ((struct ParticleScreenState*) sp)->x18 = 0;
-    ((struct ParticleScreenState*) sp)->x14 = 0;
-    ((struct ParticleScreenState*) sp)->x20 =
-        (u32) (((struct ParticleScreenState*) sp)->x3C - 0x28) / 11;
-    ((struct ParticleScreenState*) sp)->x1C =
-        (u32) (((struct ParticleScreenState*) sp)->x40 - 0x50) / 14;
-    ((struct ParticleScreenState*) sp)->x4C = lbl_804088B8;
-    ((struct ParticleScreenState*) sp)->x50 = 0;
-    ((struct ParticleScreenState*) sp)->xC4 = 0;
-}
-
-// @TODO: Currently 94.99% match - obj file has extra addi for lis/addi
-/// pair resolution; linked code matches 100%
-void hsd_80394434(void* text)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    s32* x4_ptr = (s32*) sp + 1;
-    s32 x = sp->x4;
-    s32 y = sp->x8;
-    s32 mode = sp->x0_b7;
-    s32 interlace = sp->x0_b6;
-    u8* font = (u8*) sp->x4C;
-    s8* ptr = text;
-
-    while (*ptr != 0) {
-        switch (*ptr) {
-        case '\r':
-            break;
-        case '\n':
-            x = *x4_ptr;
-            y += 14;
-            /* fallthrough */
-        default:
-            if (mode != 0) {
-                hsd_803922FC(font + (*ptr & 0x7F) * 0x38, x, y, interlace,
-                             (&sp->x24)[sp->x34], sp->x3C, sp->x40, sp->x44,
-                             sp->x50);
-            } else {
-                hsd_803921B8(font + (*ptr & 0x7F) * 0x38, x, y,
-                             (&sp->x24)[sp->x34], sp->x3C, sp->x40, sp->x44,
-                             sp->x50);
-            }
-            x += 11;
-            break;
-        }
-        ptr++;
-    }
-}
-
-// @TODO: Currently 91.32% match - needs register allocation fix
-void hsd_80394544(s32 col, s32 row, u32 num_cols, u32 num_rows, s32 x, s32 y,
-                  s32 xfb_buf, s32 xfb_w, s32 xfb_h, s32 xfb_stride,
-                  s32 font_data, void* color_data)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    s32 cur_x;
-    s32 interlace;
-    u32 r;
-    u32 c;
-    s32 mode;
-    u8 ch;
-
-    mode = sp->x0_b7;
-    interlace = sp->x0_b6;
-    r = 0;
-
-    while (r < num_rows) {
-        hsd_80393E68(col, row);
-        y -= 14;
-
-        for (cur_x = x + (c = 0) * 11; c < num_cols; c++) {
-            ch = hsd_80394068();
-            if (ch == 0) {
-                break;
-            }
-            if (mode != 0) {
-                hsd_803922FC((void*) (font_data + (ch & 0x7F) * 0x38), cur_x,
-                             y, interlace, xfb_buf, xfb_w, xfb_stride, xfb_h,
-                             color_data);
-            } else {
-                hsd_803921B8((void*) (font_data + (ch & 0x7F) * 0x38), cur_x,
-                             y, xfb_buf, xfb_w, xfb_stride, xfb_h, color_data);
-            }
-            cur_x += 11;
-        }
-        row++;
-        r++;
-    }
-}
-
-extern u8 lbl_8040AB40[];
-
-struct ParticleScreenBuffer {
-    u8 data[1];
-};
-
-#pragma push
-#pragma global_optimizer off
-void hsd_80394668(void)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    s32 pos;
-    PAD_STACK(24);
-
-    {
-        s32* dst_base;
-        struct ParticleScreenBuffer* dst;
-        u32 size;
-        struct ParticleScreenBuffer* src;
-
-        src = (struct ParticleScreenBuffer*) sp->x2C;
-        if ((u32) src != 0) {
-            /* Copy XFB data with brightness adjustment */
-            dst_base = (s32*) sp + sp->x34;
-            dst = (struct ParticleScreenBuffer*) dst_base[9];
-            size = sp->x48;
-
-            for (pos = 0; pos < size; pos += 2) {
-                u8* s1 = src->data + pos;
-                u8* d1 = dst->data + pos;
-                d1[0] = (u8) (((s1[0] - 0x10) & 0xFFFFFF) + 0x10);
-                d1[1] = s1[1];
-            }
-        } else {
-            /* Draw console text to framebuffer */
-            void* saved_color;
-            s32 cur_x;
-            s32 y_off;
-            void** x50_ptr;
-            s32* x40_ptr;
-            s32* x8_ptr;
-            s32* x4_ptr;
-            s32 row;
-            s32 col;
-
-            x50_ptr = &sp->x50;
-            saved_color = *x50_ptr;
-            *x50_ptr = lbl_8040AB40;
-            x4_ptr = &sp->x4;
-            x40_ptr = &sp->x40;
-            x8_ptr = &sp->x8;
-            row = 0;
-
-            while (row <= sp->x1C) {
-                y_off = (row + 1) * 14;
-                col = 0;
-                cur_x = 0x14;
-
-                while (col < sp->x20) {
-                    *x4_ptr = cur_x;
-                    *x8_ptr = (*x40_ptr - 0x28) - y_off;
-                    {
-                        s32 interlace = sp->x0_b6;
-                        if (sp->x0_b7 != 0) {
-                            hsd_803922FC(((ParticleFontData*) sp->x4C)->x700,
-                                         *x4_ptr, *x8_ptr, interlace,
-                                         (&sp->x24)[sp->x34], sp->x3C,
-                                         *x40_ptr, sp->x44, *x50_ptr);
-                        } else {
-                            hsd_803921B8(((ParticleFontData*) sp->x4C)->x700,
-                                         *x4_ptr, *x8_ptr, (&sp->x24)[sp->x34],
-                                         sp->x3C, *x40_ptr, sp->x44, *x50_ptr);
-                        }
-                    }
-                    cur_x += 11;
-                    col++;
-                }
-                row++;
-            }
-            *x50_ptr = saved_color;
-        }
-    }
-}
-#pragma pop
-
-// @TODO: Register swap in second loop (r29/r30)
-void hsd_80394950(OSContext* ctx)
-{
-    OSContext tmp;
-    OSContext* saved;
-    BOOL irq;
-    s32 i;
-    u8* p;
-
-    if (!(ctx->state & OS_CONTEXT_STATE_FPSAVED)) {
-        return;
-    }
-
-    irq = OSDisableInterrupts();
-    saved = OSGetCurrentContext();
-    OSClearContext(&tmp);
-    OSSetCurrentContext(&tmp);
-
-    OSReport("- FPR -----------------------------------------------\n");
-
-    i = 0;
-    p = (u8*) ctx;
-    do {
-        f64 fval = *(f64*) (p + 0x90);
-        OSReport("R%02d=%08X:%08X (%e)\n", i, *(u32*) (p + 0x90),
-                 *(u32*) (p + 0x94), fval);
-        i++;
-        p += 8;
-    } while (i < 32);
-
-    OSReport("- PSF -----------------------------------------------\n");
-
-    i = 0;
-    p = (u8*) ctx;
-    do {
-        u32 val0 = ((u32*) p)[0x24];
-        u32 val1 = ((u32*) p)[0x25];
-        OSReport("R%02d=%08X:%08X (%e, %e)\n", i, val0, val1, ((f32*) p)[0x24],
-                 ((f32*) p)[0x25]);
-        i++;
-        p += 8;
-    } while (i < 32);
-
-    OSClearContext(&tmp);
-    OSSetCurrentContext(saved);
-    OSRestoreInterrupts(irq);
-}
-
-/// unused function to force data ordering for these report strings
-static void unused(OSContext* ctx)
-{
-    OSReport("- MISC ----------------------------------------------\n");
-    OSReport("SRR0=%08X SRR1=%08X\n", ctx->srr0, ctx->srr1);
-    OSReport("CR  =%08X LR  =%08X\n", ctx->cr, ctx->lr);
-    OSReport("CTR =%08X XER =%08X\n", ctx->ctr, ctx->xer);
-    OSReport("FPSCR=%08X\n", ctx->fpscr);
-    OSReport("GQR%d=%08X GQR%d=%08X\n", 0, ctx->gqr[0], 1, ctx->gqr[1]);
-}
-
-// @TODO: Currently 99.94% match - minor relocation difference
-void Exception_ReportStackTrace(OSContext* ctx, int max_depth)
-{
-    u32 i;
-    u32* sp;
-
-    OSReport("- STACK ---------------------------------------------\n");
-    OSReport(" Address:  Back Chain  LR Save\n");
-
-    sp = (u32*) ctx->gpr[1];
-    i = 0;
-
-    while (sp != NULL && (u32) (sp + 0x4000) != 0xFFFF && i < (u32) max_depth)
-    {
-        if ((u32) sp < 0x80000000u) {
-            break;
-        }
-        if ((s64) (u32) sp >= (s64) OSGetPhysicalMemSize() + 0x800000000) {
-            break;
-        }
-        OSReport("%08X:   %08X   %08X\n", sp, sp[0], sp[1]);
-        sp = (u32*) sp[0];
-        i++;
-    }
-}
-
-// @TODO: Currently 99.81% match - instructions match, string relocations
-// differ
-void Exception_ReportCodeline(u16 error, int dsisr, int dar, OSContext* ctx)
-{
-    char* exception_type;
-    char* around_or_at;
-    int srr1;
-
-    OSReport("- UNHANDLED EXCEPTION -------------------------------\n");
-    OSReport("DSISR=%08X DAR=%08X\n", dsisr, dar);
-    error = (u16) error;
-    OSReport("ERROR %d: ", error);
-
-    switch (error) {
-    case 0:
-        OSReport("(SYSTEM RESET)\n");
-        return;
-    case 1:
-        OSReport("(MACHINE CHECK)\n");
-        return;
-    case 2:
-        OSReport("(DSI)\n");
-        OSReport("Instruction at %08X (read from SRR0) attempted\n"
-                 "to access invalid address %08X (read from DAR)\n",
-                 ctx->srr0, dar);
-        if (dsisr & 0x02000000) {
-            OSReport("Store operation.\n");
-        } else {
-            OSReport("Load operation.\n");
-        }
-        if (dsisr & 0x40000000) {
-            OSReport("No appropriate page entry or DBAT exist.\n");
-        }
-        if (dsisr & 0x08000000) {
-            OSReport("Memory access is not permitted by the page or DBAT.\n");
-        }
-        if (dsisr & 0x400000) {
-            OSReport("DABR match occurs.\n");
-        }
-        return;
-    case 3:
-        OSReport("(ISI)\n");
-        OSReport("Attempted to fetch instruction from invalid address\n"
-                 "%08X (read from SRR0)\n",
-                 ctx->srr0);
-        if (ctx->srr1 & 0x40000000) {
-            OSReport("No appropriate page entry or IBAT exist.\n");
-        }
-        if (ctx->srr1 & 0x10000000) {
-            OSReport("Illegal address.\n");
-        }
-        if (ctx->srr1 & 0x08000000) {
-            OSReport("Memory access is not permitted by the page or IBAT.\n");
-        }
-        break;
-    case 4:
-        OSReport("(EXTERNAL INTERRUPT)\n");
-        return;
-    case 5:
-        OSReport("(ALIGNMNET)\n");
-        OSReport("Instruction at %08X (read from SRR0) attempted\n"
-                 "to access unaligned address %08X (read from DAR)\n",
-                 ctx->srr0, dar);
-        return;
-    case 6:
-        OSReport("(PROGRAM)\n");
-        srr1 = ctx->srr1;
-        around_or_at = (srr1 & 0x10000) ? "around" : "at";
-        if (srr1 & 0x100000) {
-            exception_type = "IEEE floating point enabled program exception";
-        } else if (srr1 & 0x80000) {
-            exception_type = "Illegal instruction";
-        } else if (srr1 & 0x40000) {
-            exception_type = "Privileged instruction";
-        } else if (srr1 & 0x20000) {
-            exception_type = "Trap program exception";
-        } else {
-            exception_type = "Unknown type of program exception";
-        }
-        OSReport("%s %s %08X (read from SRR0)\n", exception_type, around_or_at,
-                 ctx->srr0);
-        return;
-    case 7:
-        OSReport("(FLOATING POINT)\n");
-        return;
-    case 8:
-        OSReport("(DECREMENTER)\n");
-        return;
-    case 9:
-        OSReport("(SYSTEM CALL)\n");
-        return;
-    case 10:
-        OSReport("(TRACE)\n");
-        return;
-    case 11:
-        OSReport("(PERFORMANCE MONITOR)\n");
-        return;
-    case 12:
-        OSReport("(BREAKPOINT)\n");
-        return;
-    case 13:
-        OSReport("(SYSTEM INTERRPUT)\n");
-        return;
-    case 14:
-        OSReport("(THERMAL INTERRPUT)\n");
-        return;
-    default:
-        OSReport("(UNKNOWN)\n");
-        break;
-    }
-}
-
-void fn_80394DF4(void* node_ptr)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    ExcptNode* node = node_ptr;
-    ExcptNode* prev;
-    ExcptNode* head;
-    ExcptNode** head_ptr;
-
-    if (node == NULL || (head = *(head_ptr = (ExcptNode**) &sp->xD0)) == NULL)
-    {
-        return;
-    }
-    if (head == node) {
-        *head_ptr = head->next;
-        node->next = NULL;
-        sp->x0_b5 = 1;
-        return;
-    }
-    prev = head;
-    while (prev != NULL) {
-        if (prev->next == node) {
-            prev->next = node->next;
-            node->next = NULL;
-            sp->x0_b5 = 1;
-            return;
-        }
-        prev = prev->next;
-    }
-}
-
-void hsd_80394E8C(void* node_ptr)
-{
-    void** head;
-    ExcptNode* node = node_ptr;
-
-    if (hsd_804D78C8 < 1) {
-        return;
-    }
-    head = (void**) &hsd_804CF810.xD0;
-    if (*head == node_ptr) {
-        return;
-    }
-    if (*head == NULL) {
-        *head = node_ptr;
-        node->next = NULL;
-    } else {
-        fn_80394DF4(node_ptr);
-        node->next = ((ExcptNode*) *head)->next;
-        ((ExcptNode*) *head)->next = NULL;
-        *head = node_ptr;
-    }
-    if (node->callback != NULL) {
-        node->callback(node);
-    }
-    hsd_804CF810.x0_b5 = 1;
-}
-
-extern u8 lbl_8040AB00[];
-extern u8 lbl_8040AB20[];
-
-// @TODO: Currently 94.46% match - BSS relocation and register allocation
-// differences remain
-void hsd_80394F48(void* data)
-{
-    u8* base_color = lbl_8040AB00;
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    EventData* dp = data;
-    s32 num_entries;
-    void** px50;
-    s32* pxC8;
-    s32* pxCC;
-    s32* px4;
-    s32* px40;
-    s32* px8;
-    s32 col_start;
-    s32 row_start;
-    void* hi_color;
-    s32 x_base;
-    s32 cur_row;
-    s32 i;
-    s32 b6;
-
-    PAD_STACK(64);
-
-    num_entries = strlen(dp->entries[0]);
-    px50 = &sp->x50;
-    pxC8 = &sp->xC8;
-    pxCC = &sp->xCC;
-    px4 = &sp->x4;
-    px40 = &sp->x40;
-    px8 = &sp->x8;
-    *px50 = base_color;
-    col_start = *pxC8;
-    row_start = *pxCC;
-
-    *px4 = col_start * 11 + 0x14;
-    *px8 = (*px40 - 0x28) - (row_start + 1) * 14;
-
-    b6 = sp->x0_b6;
-    if (sp->x0_b7 != 0) {
-        hsd_803922FC(((ParticleFontData*) sp->x4C)->x968, *px4, *px8, b6,
-                     (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-    } else {
-        hsd_803921B8(((ParticleFontData*) sp->x4C)->x968, *px4, *px8,
-                     (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-    }
-
-    i = 0;
-    *px4 += 11;
-    while (i < num_entries) {
-        b6 = sp->x0_b6;
-        if (sp->x0_b7 != 0) {
-            hsd_803922FC(((ParticleFontData*) sp->x4C)->x9D8, *px4, *px8, b6,
-                         (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-        } else {
-            hsd_803921B8(((ParticleFontData*) sp->x4C)->x9D8, *px4, *px8,
-                         (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-        }
-        i++;
-        *px4 += 11;
-    }
-
-    b6 = sp->x0_b6;
-    if (sp->x0_b7 != 0) {
-        hsd_803922FC(((ParticleFontData*) sp->x4C)->x968, *px4, *px8, b6,
-                     (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-    } else {
-        hsd_803921B8(((ParticleFontData*) sp->x4C)->x968, *px4, *px8,
-                     (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-    }
-
-    *px4 += 11;
-    x_base = col_start * 11 + 0x14;
-    hi_color = base_color + 0x20;
-    cur_row = row_start - 1;
-
-    for (i = 0; dp->entries[i] != 0; i++) {
-        *px4 = x_base;
-        *px8 = (*px40 - 0x28) - (cur_row-- + 1) * 14;
-
-        b6 = sp->x0_b6;
-        if (sp->x0_b7 != 0) {
-            hsd_803922FC(((ParticleFontData*) sp->x4C)->x1B20, *px4, *px8, b6,
-                         (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-        } else {
-            hsd_803921B8(((ParticleFontData*) sp->x4C)->x1B20, *px4, *px8,
-                         (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-        }
-
-        if (i == dp->index) {
-            *px50 = hi_color;
-        } else {
-            *px50 = base_color;
-        }
-
-        *px4 += 11;
-        hsd_80394434(dp->entries[i]);
-
-        *px4 += strlen(dp->entries[i]) * 11;
-        *px50 = base_color;
-
-        b6 = sp->x0_b6;
-        if (sp->x0_b7 != 0) {
-            hsd_803922FC(((ParticleFontData*) sp->x4C)->x1B20, *px4, *px8, b6,
-                         (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-        } else {
-            hsd_803921B8(((ParticleFontData*) sp->x4C)->x1B20, *px4, *px8,
-                         (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-        }
-    }
-
-    *px50 = base_color;
-    *px4 = x_base;
-    *px8 = (*px40 - 0x28) - (cur_row + 1) * 14;
-
-    b6 = sp->x0_b6;
-    if (sp->x0_b7 != 0) {
-        hsd_803922FC(((ParticleFontData*) sp->x4C)->x968, *px4, *px8, b6,
-                     (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-    } else {
-        hsd_803921B8(((ParticleFontData*) sp->x4C)->x968, *px4, *px8,
-                     (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-    }
-
-    {
-        s32 j = 0;
-        *px4 += 11;
-        while (j < num_entries) {
-            b6 = sp->x0_b6;
-            if (sp->x0_b7 != 0) {
-                hsd_803922FC(((ParticleFontData*) sp->x4C)->x9D8, *px4, *px8,
-                             b6, (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44,
-                             *px50);
-            } else {
-                hsd_803921B8(((ParticleFontData*) sp->x4C)->x9D8, *px4, *px8,
-                             (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44,
-                             *px50);
-            }
-            j++;
-            *px4 += 11;
-        }
-    }
-
-    b6 = hsd_804CF810.x0_b6;
-    if (sp->x0_b7 != 0) {
-        hsd_803922FC(((ParticleFontData*) sp->x4C)->x968, *px4, *px8, b6,
-                     (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-    } else {
-        hsd_803921B8(((ParticleFontData*) sp->x4C)->x968, *px4, *px8,
-                     (&sp->x24)[sp->x34], sp->x3C, *px40, sp->x44, *px50);
-    }
-
-    *px4 += 11;
-    *pxC8 += 4;
-    *pxCC -= dp->index + 1;
-}
-
-s32 hsd_80395550(void* event_ptr)
-{
-    EventData* data = event_ptr;
-    u32 flags = hsd_804CF810.xBC;
-    u32 bit = 1;
-
-    while (bit <= flags) {
-        switch (flags & bit) {
-        case 0x8: {
-            s32 counter = data->index;
-            if (counter > 0) {
-                data->index = counter - 1;
-            } else {
-                char** base = data->entries;
-                s32 count = 0;
-                while (base[count + 1] != 0) {
-                    count++;
-                }
-                data->index = count;
-            }
-            return 1;
-        }
-        case 0x4: {
-            s32 idx = data->index;
-            char** base = data->entries;
-            if (base[idx + 1] != 0) {
-                data->index = idx + 1;
-            } else {
-                data->index = 0;
-            }
-            return 1;
-        }
-        case 0x100:
-            return 2;
-        case 0x200:
-        case 0x1000:
-            return -1;
-        default:
-            bit <<= 1;
-            break;
-        }
-    }
-    return 0;
-}
-
-extern u8 lbl_8040AB20[];
-extern u8 lbl_8040B8AC[];
-extern u8 lbl_8040B904[];
-
-void hsd_80395644(void)
-{
-    struct ParticleScreenState* sp =
-        (struct ParticleScreenState*) &hsd_804CF810;
-    void* saved;
-    void** p = &sp->x50;
-    s32 val_x20;
-    s32 val_x1C;
-
-    PAD_STACK(16);
-    saved = *p;
-    *p = lbl_8040AB20;
-    val_x20 = sp->x20;
-    val_x1C = sp->x1C;
-    sp->x4 = (val_x20 - 21) * 11 + 20;
-    sp->x8 = (sp->x40 - 40) - (val_x1C + 1) * 14;
-    if (hsd_804D78C8 >= 1) {
-        hsd_80394434(lbl_8040B8AC);
-    }
-    *p = saved;
-}
-
-s32 hsd_803956D8(void* disp_ptr)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    DispData* data = disp_ptr;
-    s32 val;
-
-    if (sp->xBC & 0x1000) {
-        hsd_80394E8C(lbl_8040B904);
-        sp->x14 = 0;
-        data->x10 = 0;
-        return 1;
-    }
-
-    val = data->x10;
-    if (val < 120) {
-        data->x10 = val + 1;
-    } else {
-        data->x10 = 0;
-        {
-            s32* pp = &sp->x14;
-            *pp += sp->x1C;
-            if (!(u8) hsd_80394128(0, *pp)) {
-                if (!(u8) hsd_80394128(0, *pp + 1)) {
-                    *pp = 0;
-                    hsd_80394E8C(lbl_8040B904);
-                }
-            }
-        }
-        return 1;
-    }
-    return 0;
-}
-
-static inline void get_screen_data(void** data)
-{
-    *data = hsd_804CF810.x50;
-}
-
-static inline void set_default_screen_data(void)
-{
-    hsd_804CF810.x50 = lbl_8040AB20;
-}
-
-// @TODO: BSS relocation and register allocation differences remain
-void hsd_803957C0(u8 input)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    s32* x10_ptr = &sp->x10;
-    void** x50_ptr;
-    s32* x8_ptr;
-    s32* x40_ptr;
-    s32* x4_ptr;
-    s32 col;
-    s32 row;
-    s32 b6;
-    u8 ch;
-    void* saved;
-    PAD_STACK(24);
-
-    col = sp->x0C + sp->x18;
-    row = *x10_ptr + sp->x14;
-
-    while (col >= 0) {
-        ch = (u8) hsd_80394128(col, row);
-        if (ch != 0) {
-            break;
-        }
-        col--;
-    }
-
-    if (ch == 0) {
-        if (col < 0) {
-            col = 0;
-        }
-        ch = ' ';
-    }
-
-    get_screen_data(&saved);
-    set_default_screen_data();
-    sp->x8 = (sp->x40 - 40) - (*x10_ptr + 1) * 14;
-    sp->x4 = col * 11 + 20;
-
-    x50_ptr = &sp->x50;
-    {
-        struct ParticleScreenState* draw_sp = &hsd_804CF810;
-        x4_ptr = &draw_sp->x4;
-        x40_ptr = &draw_sp->x40;
-        (void) x40_ptr;
-        x8_ptr = &draw_sp->x8;
-        b6 = draw_sp->x0_b6;
-
-        if (draw_sp->x0_b7) {
-            hsd_803922FC((void*) ((u8*) draw_sp->x4C + (ch & 0x7F) * 0x38),
-                         *x4_ptr, *x8_ptr, b6, (&draw_sp->x24)[draw_sp->x34],
-                         draw_sp->x3C, *x40_ptr, draw_sp->x44, *x50_ptr);
-        } else {
-            hsd_803921B8((void*) ((u8*) draw_sp->x4C + (ch & 0x7F) * 0x38),
-                         *x4_ptr, *x8_ptr, (&draw_sp->x24)[draw_sp->x34],
-                         draw_sp->x3C, *x40_ptr, draw_sp->x44, *x50_ptr);
-        }
-    }
-
-    if (input == (u32) sp->xD0) {
-        struct ParticleScreenState* final_sp = &hsd_804CF810;
-        *x4_ptr = (final_sp->x20 - 0x27) * 11 + 20;
-        *x8_ptr = (*x40_ptr - 40) - (final_sp->x1C + 1) * 14;
-        hsd_80394434("<<PUSH START BUTTON TO SEE INSTRUCTION>>");
-    }
-
-    *x50_ptr = saved;
-}
-
-#pragma push
-#pragma dont_inline on
-s32 hsd_80395970(void)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    s32 saved_x;
-    s32 saved_y;
-    s32 result;
-    s32 i;
-    u8 ch;
-
-    result = 0;
-    hsd_80393E34(&saved_x, &saved_y);
-    hsd_80393E68(sp->x0C + sp->x18, sp->x10 + sp->x14);
-
-    for (i = 0; i < 8; i++) {
-        ch = hsd_80394068();
-        if (ch == 0) {
-            ch = hsd_80394068();
-            if (ch == 0) {
-                break;
-            }
-        }
-        if (ch >= '0' && ch <= '9') {
-            result = (result << 4) + ch - '0';
-        } else if (ch >= 'A' && ch <= 'F') {
-            result = (result << 4) + ch - 'A' + 10;
-        } else if (ch >= 'a' && ch <= 'f') {
-            result = (result << 4) + ch - 'a' + 10;
-        } else {
-            break;
-        }
-    }
-
-    result = (result & 0x0FFFFFF0) | (s32) 0x80000000;
-    hsd_80393E68(saved_x, saved_y);
-    return result;
-}
-
-#pragma pop
-
-extern u8 lbl_8040BA5C[];
-
-extern struct {
-    /* 0x00 */ u8 _pad[0x10];
-    /* 0x10 */ u32 x10;
-} lbl_8040BAF0;
-
-extern struct {
-    /* 0x00 */ u8 _pad[0x10];
-    /* 0x10 */ u32 x10;
-    /* 0x14 */ s32 x14;
-    /* 0x18 */ void* x18;
-} lbl_8040BC3C;
-
-s32 hsd_80395A78(void)
-{
-    u32 bit;
-    s32 new_col;
-    s32 new_scroll;
-    s32 sum;
-    void* sp = &hsd_804CF810;
-    PAD_STACK(8);
-
-    bit = 1;
-    while (bit <= ((struct ParticleScreenState*) sp)->xBC) {
-        switch (((struct ParticleScreenState*) sp)->xBC & bit) {
-        case 0x8:
-            if (((struct ParticleScreenState*) sp)->x10 <
-                ((struct ParticleScreenState*) sp)->x1C - 1)
-            {
-                ((struct ParticleScreenState*) sp)->x10 =
-                    ((struct ParticleScreenState*) sp)->x10 + 1;
-            } else {
-                ((struct ParticleScreenState*) sp)->x14 += 1;
-            }
-            return 1;
-        case 0x4:
-            if (((struct ParticleScreenState*) sp)->x10 > 0) {
-                ((struct ParticleScreenState*) sp)->x10 -= 1;
-            } else if (((struct ParticleScreenState*) sp)->x14 > 0) {
-                ((struct ParticleScreenState*) sp)->x14 -= 1;
-            } else {
-                break;
-            }
-            return 1;
-        case 0x1:
-            if (((struct ParticleScreenState*) sp)->x0C > 0) {
-                ((struct ParticleScreenState*) sp)->x0C -= 1;
-            } else if (((struct ParticleScreenState*) sp)->x18 > 0) {
-                ((struct ParticleScreenState*) sp)->x18 -= 1;
-            } else {
-                break;
-            }
-            while (!(u8) hsd_80394128(
-                ((struct ParticleScreenState*) sp)->x0C +
-                    ((struct ParticleScreenState*) sp)->x18,
-                ((struct ParticleScreenState*) sp)->x10 +
-                    ((struct ParticleScreenState*) sp)->x14))
-            {
-                if (((struct ParticleScreenState*) sp)->x0C > 0) {
-                    ((struct ParticleScreenState*) sp)->x0C -= 1;
-                } else if (((struct ParticleScreenState*) sp)->x18 > 0) {
-                    ((struct ParticleScreenState*) sp)->x18 -= 1;
-                } else {
-                    break;
-                }
-            }
-            return 1;
-        case 0x2:
-            new_col = ((struct ParticleScreenState*) sp)->x0C;
-            new_scroll = ((struct ParticleScreenState*) sp)->x18;
-            if ((u32) new_col <
-                (u32) (((struct ParticleScreenState*) sp)->x20 - 1))
-            {
-                new_col += 1;
-            } else {
-                new_scroll += 1;
-            }
-            sum = new_col + new_scroll;
-            if ((u8) hsd_80394128(sum,
-                                  ((struct ParticleScreenState*) sp)->x10 +
-                                      ((struct ParticleScreenState*) sp)->x14))
-            {
-                ((struct ParticleScreenState*) sp)->x0C = new_col;
-                ((struct ParticleScreenState*) sp)->x18 = new_scroll;
-                return 1;
-            }
-            break;
-        case 0x100:
-            lbl_8040BC3C.x10 = hsd_80395970();
-            lbl_8040BC3C.x18 = &lbl_8040BAF0;
-            if ((ExcptNode*) &lbl_8040BC3C != NULL) {
-                fn_80394DF4((ExcptNode*) &lbl_8040BC3C);
-                ((ExcptNode*) &lbl_8040BC3C)->next =
-                    (ExcptNode*) ((struct ParticleScreenState*) sp)->xD0;
-                ((struct ParticleScreenState*) sp)->xD0 = &lbl_8040BC3C;
-                if (((ExcptNode*) &lbl_8040BC3C)->callback != NULL) {
-                    ((ExcptNode*) &lbl_8040BC3C)
-                        ->callback((ExcptNode*) &lbl_8040BC3C);
-                }
-                ((struct ParticleScreenState*) sp)->x0_b5 = 1;
-            }
-            return 1;
-        case 0x400:
-            lbl_8040BC3C.x18 = &lbl_8040BAF0;
-            if ((ExcptNode*) &lbl_8040BC3C != NULL) {
-                fn_80394DF4((ExcptNode*) &lbl_8040BC3C);
-                ((ExcptNode*) &lbl_8040BC3C)->next =
-                    (ExcptNode*) ((struct ParticleScreenState*) sp)->xD0;
-                ((struct ParticleScreenState*) sp)->xD0 = &lbl_8040BC3C;
-                if (((ExcptNode*) &lbl_8040BC3C)->callback != NULL) {
-                    ((ExcptNode*) &lbl_8040BC3C)
-                        ->callback((ExcptNode*) &lbl_8040BC3C);
-                }
-                ((struct ParticleScreenState*) sp)->x0_b5 = 1;
-            }
-            return 1;
-        case 0x1000:
-            if ((ExcptNode*) lbl_8040BA5C != NULL) {
-                fn_80394DF4((ExcptNode*) lbl_8040BA5C);
-                ((ExcptNode*) lbl_8040BA5C)->next =
-                    (ExcptNode*) ((struct ParticleScreenState*) sp)->xD0;
-                ((struct ParticleScreenState*) sp)->xD0 = lbl_8040BA5C;
-                if (((ExcptNode*) lbl_8040BA5C)->callback != NULL) {
-                    ((ExcptNode*) lbl_8040BA5C)
-                        ->callback((ExcptNode*) lbl_8040BA5C);
-                }
-                ((struct ParticleScreenState*) sp)->x0_b5 = 1;
-            }
-            return 1;
-        default:
-            break;
-        }
-        bit <<= 1;
-    }
-    return 0;
-}
-extern u8 lbl_8040BEC4[];
-
-static inline void ps_remove_node(struct ParticleScreenState* sp, void* node)
-{
-    void** head_ptr;
-    void* cur;
-
-    if (node == NULL) {
-        return;
-    }
-    if (*(head_ptr = &sp->xD0) == NULL) {
-        return;
-    }
-    while ((cur = *head_ptr) != NULL && cur != node) {
-        void* next = *(void**) cur;
-        *(void**) cur = NULL;
-        *head_ptr = next;
-    }
-    if (cur == node) {
-        void* next = *(void**) node;
-        *(void**) node = NULL;
-        *head_ptr = next;
-    }
-    sp->x0_b5 = 1;
-}
-
-static inline void ps_clear_nodes(struct ParticleScreenState* sp)
-{
-    void** head_ptr;
-    ExcptNode* cur;
-
-    cur = *(head_ptr = &sp->xD0);
-    while (cur != NULL) {
-        ExcptNode* next = cur->next;
-        cur->next = NULL;
-        cur = next;
-    }
-    *head_ptr = NULL;
-    sp->x0_b5 = 1;
-}
-
-// @TODO: Currently 99.38% match - remaining register allocation differences
-s32 hsd_80395D88(void* data)
-{
-    char* msg = (char*) lbl_8040AB00;
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    s32 result;
-
-    result = hsd_80395550(data);
-
-    switch (result) {
-    case 1:
-        return 1;
-    case 2: {
-        u32 cmd = *(u32*) ((u8*) data + 0x14);
-        switch (cmd) {
-        case 0:
-            sp->xBC = 8;
-            return 0;
-        case 1:
-            sp->xBC = 4;
-            return 0;
-        case 2:
-            sp->xBC = 1;
-            return 0;
-        case 3:
-            sp->xBC = 2;
-            return 0;
-        case 4:
-            sp->xBC = 0x100;
-            goto remove_return_0;
-        case 5:
-            sp->xBC = 0x400;
-        remove_return_0:
-            ps_remove_node(sp, data);
-            return 0;
-        case 6: {
-            s32 i;
-            OSContext* ctx;
-            OSContext** ctx_ptr;
-            s32 saved;
-
-            if (*(ctx_ptr = &sp->xD4) != NULL) {
-                saved = hsd_80393D2C(1);
-                ctx = *ctx_ptr;
-                OSReport(msg + 0x728);
-                i = 0;
-                do {
-                    OSReport(msg + 0x760, i, ctx->gpr[i], ctx->gpr[i],
-                             i + 0x10, ctx->gpr[i + 0x10], ctx->gpr[i + 0x10]);
-                    i++;
-                } while (i < 0x10);
-                hsd_80394950(*ctx_ptr);
-                ctx = *ctx_ptr;
-                OSReport(msg + 0x828);
-                OSReport(msg + 0x860, ((u32*) ctx)[0x198 / 4],
-                         ((u32*) ctx)[0x19C / 4]);
-                OSReport(msg + 0x878, ((u32*) ctx)[0x80 / 4],
-                         ((u32*) ctx)[0x84 / 4]);
-                OSReport(msg + 0x890, ((u32*) ctx)[0x88 / 4],
-                         ((u32*) ctx)[0x8C / 4]);
-                OSReport(msg + 0x8A8, ((u32*) ctx)[0x194 / 4]);
-                i = 0;
-                do {
-                    OSReport(msg + 0x8B4, i, ((u32*) ctx)[i], i + 4,
-                             ((u32*) ctx)[0x1B4 / 4 + i]);
-                    i++;
-                } while (i < 4);
-                hsd_80393D2C(saved);
-            }
-            ps_remove_node(sp, data);
-            return 1;
-        }
-        case 7:
-            hsd_80394E8C(lbl_8040BEC4);
-            return 1;
-        case 8:
-            ps_clear_nodes(sp);
-            return 1;
-        default:
-            break;
-        }
-        break;
-    }
-    case -1:
-        ps_remove_node(sp, data);
-        return 1;
-    }
-    return 0;
-}
-
-void hsd_80396130(void)
-{
-    u32 old = lbl_8040BC3C.x10;
-    u32 memsize;
-
-    memsize = OSGetPhysicalMemSize();
-    old &= 0x0FFFFFFF;
-    lbl_8040BAF0.x10 = (old + memsize) % memsize + 0x80000000;
-    PAD_STACK(16);
-}
-
-static char* lbl_804D62CC = "+- MEMORY DUMP ------------------------------+";
-static char* lbl_804D62D0 = "|%08X=%08X:%08X:%08X:%08X|";
-static char* lbl_804D62D4 = "+--------------------------------------------+";
-
-static inline s32 hsd_80396188_calc_col(void)
-{
-    return ((hsd_804CF810.x20 - 0x2E) / 2) * 11 + 20;
-}
-
-// @TODO: Currently 96.11% match - needs register allocation fix
-static inline void hsd_80396188_draw_rows(char* buf, s32 col, u32** addr,
-                                          s32* i)
-{
-    s32* x8 = &hsd_804CF810.x8;
-    s32* x4 = &hsd_804CF810.x4;
-    s32* x40 = &hsd_804CF810.x40;
-
-    hsd_80394434(lbl_804D62CC);
-    *i = 0;
-    do {
-        s32 row = 4 - *i;
-        *x4 = col;
-        *x8 = (*x40 - 0x28) - (row + 1) * 14;
-        sprintf(buf, lbl_804D62D0, *addr, (*addr)[0], (*addr)[1], (*addr)[2],
-                (*addr)[3]);
-        hsd_80394434(buf);
-        {
-            u32 memsize = OSGetPhysicalMemSize();
-            *addr = (u32*) ((((u32) *addr & 0x0FFFFFFF) + memsize + 0x10) %
-                                memsize +
-                            0x80000000);
-        }
-        (*i)++;
-    } while (*i < 4);
-    *x4 = col;
-}
-
-void hsd_80396188(void)
-{
-    void* saved;
-    s32 col;
-    s32 i;
-    u32* addr;
-    char buf[64];
-    PAD_STACK(20);
-
-    addr = (u32*) lbl_8040BAF0.x10;
-    saved = hsd_804CF810.x50;
-    col = hsd_80396188_calc_col();
-    hsd_804CF810.x50 = lbl_8040AB00;
-    hsd_804CF810.x4 = col;
-    hsd_804CF810.x8 = hsd_804CF810.x40 - 0x7C;
-    hsd_80396188_draw_rows(buf, col, &addr, &i);
-    hsd_804CF810.x8 = hsd_804CF810.x40 - 0x36;
-    hsd_80394434(lbl_804D62D4);
-    hsd_804CF810.x50 = saved;
-}
-
-extern ExcptNode lbl_8040BBE8[];
-
-static char lbl_804D62D8[] = "\n";
-static char lbl_804D62DC[] = "[%08X";
-static char lbl_804D62E4[] = ":";
-static char lbl_804D62E8[] = "%02X";
-static char lbl_804D62F0[] = "]\n";
-
-s32 hsd_803962A8(void* data)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    u32 bit;
-    s32 old8, old4, old1, old2;
-    u32 *ptr8, *ptr4, *ptr1, *ptr2;
-    u8* addr;
-    s32 i, j, k;
-    PAD_STACK(32);
-
-    bit = 1;
-    while (bit <= sp->xBC) {
-        switch (sp->xBC & bit) {
-        case 0x8: {
-            ptr8 = &lbl_8040BAF0.x10;
-            old8 = *ptr8;
-            {
-                u32 memsize = OSGetPhysicalMemSize();
-                old8 &= 0x0FFFFFFF;
-                *ptr8 = ((old8 + memsize - 0x10) % memsize) + 0x80000000;
-            }
-            return 1;
-        }
-        case 0x4: {
-            ptr4 = &lbl_8040BAF0.x10;
-            old4 = *ptr4;
-            {
-                u32 memsize = OSGetPhysicalMemSize();
-                old4 &= 0x0FFFFFFF;
-                *ptr4 = ((old4 + memsize + 0x10) % memsize) + 0x80000000;
-            }
-            return 1;
-        }
-        case 0x1: {
-            ptr1 = &lbl_8040BAF0.x10;
-            old1 = *ptr1;
-            {
-                u32 memsize = OSGetPhysicalMemSize();
-                old1 &= 0x0FFFFFFF;
-                *ptr1 = ((old1 + memsize - 0x40) % memsize) + 0x80000000;
-            }
-            return 1;
-        }
-        case 0x2: {
-            ptr2 = &lbl_8040BAF0.x10;
-            old2 = *ptr2;
-            {
-                u32 memsize = OSGetPhysicalMemSize();
-                old2 &= 0x0FFFFFFF;
-                *ptr2 = ((old2 + memsize + 0x40) % memsize) + 0x80000000;
-            }
-            return 1;
-        }
-        case 0x400:
-            hsd_80393D2C(1);
-            addr = (u8*) lbl_8040BAF0.x10;
-            if ((u8) hsd_80394128(0, 0)) {
-                OSReport(lbl_804D62D8);
-            }
-            i = 0;
-            do {
-                OSReport(lbl_804D62DC, (u32) addr);
-                j = 0;
-                do {
-                    OSReport(lbl_804D62E4);
-                    k = 0;
-                    do {
-                        OSReport(lbl_804D62E8, *addr++);
-                        k++;
-                    } while (k < 4);
-                    j++;
-                } while (j < 4);
-                OSReport(lbl_804D62F0);
-                i++;
-            } while (i < 4);
-            hsd_80393D2C(0);
-            ps_remove_node(sp, data);
-            sp->x0C = 0;
-            sp->x10 = 0;
-            sp->x18 = 0;
-            sp->x14 = 0;
-            return 1;
-        case 0x100:
-            lbl_8040BC3C.x10 = (s32) lbl_8040BAF0.x10;
-            lbl_8040BC3C.x18 = &lbl_8040BAF0;
-            if (&lbl_8040BC3C != NULL) {
-                fn_80394DF4(&lbl_8040BC3C);
-                ((ExcptNode*) &lbl_8040BC3C)->next = (ExcptNode*) sp->xD0;
-                sp->xD0 = &lbl_8040BC3C;
-                if (((ExcptNode*) &lbl_8040BC3C)->callback != NULL) {
-                    ((ExcptNode*) &lbl_8040BC3C)
-                        ->callback((ExcptNode*) &lbl_8040BC3C);
-                }
-                sp->x0_b5 = 1;
-            }
-            return 1;
-        case 0x200:
-            ps_remove_node(sp, data);
-            return 1;
-        case 0x1000:
-            if (lbl_8040BBE8 != NULL) {
-                fn_80394DF4(lbl_8040BBE8);
-                lbl_8040BBE8->next = (ExcptNode*) sp->xD0;
-                sp->xD0 = lbl_8040BBE8;
-                if (lbl_8040BBE8->callback != NULL) {
-                    lbl_8040BBE8->callback(lbl_8040BBE8);
-                }
-                sp->x0_b5 = 1;
-            }
-            return 1;
-        default:
-            bit <<= 1;
-            break;
-        }
-    }
-    return 0;
-}
-
-s32 hsd_803966A0(void* data)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    s32 result;
-
-    result = hsd_80395550(data);
-    switch (result) {
-    case 1:
-        return 1;
-    case 2: {
-        u32 cmd = *(u32*) ((u8*) data + 0x14);
-        switch (cmd) {
-        case 0:
-            sp->xBC = 8;
-            break;
-        case 1:
-            sp->xBC = 4;
-            break;
-        case 2:
-            sp->xBC = 1;
-            break;
-        case 3:
-            sp->xBC = 2;
-            break;
-        case 4:
-            sp->xBC = 0x100;
-            break;
-        case 5:
-            sp->xBC = 0x400;
-            break;
-        case 6:
-            sp->xBC = 0x200;
-            break;
-        }
-        ps_remove_node(sp, data);
-        return 0;
-    }
-    case -1:
-        ps_remove_node(sp, data);
-        return 1;
-    default:
-        return 0;
-    }
-}
-
-void hsd_80396868(void)
-{
-    lbl_8040BC3C.x10 = (lbl_8040BC3C.x10 & 0x0FFFFFF0) | 0x80000000;
-}
-
-static char* lbl_804D62F4 = "+--------------------------+";
-static char* lbl_804D62F8 = "| INPUT ADDRESS : 8%07X |";
-
-// @TODO: Currently 92.87% match - .bss.0 relocation caps match percentage
-void hsd_80396884(void)
-{
-    char buf[32];
-    s32 b6;
-    s32 x_base;
-    u8 ch;
-    void* saved;
-    PAD_STACK(28);
-
-    x_base = (hsd_804CF810.x20 - 30) / 2;
-    sprintf(buf, "| INPUT ADDRESS : 8%07X |", lbl_8040BC3C.x10 & 0x0FFFFFFF);
-    saved = hsd_804CF810.x50;
-    hsd_804CF810.x50 = lbl_8040AB00;
-    hsd_804CF810.x4 = x_base * 11 + 20;
-    hsd_804CF810.x8 = hsd_804CF810.x40 - 82;
-    hsd_80394434("+--------------------------+");
-    hsd_804CF810.x4 = x_base * 11 + 20;
-    hsd_804CF810.x8 = hsd_804CF810.x40 - 68;
-    hsd_80394434(buf);
-    hsd_804CF810.x4 = x_base * 11 + 20;
-    hsd_804CF810.x8 = hsd_804CF810.x40 - 54;
-    hsd_80394434("+--------------------------+");
-    hsd_804CF810.x4 = (lbl_8040BC3C.x14 + 19 + x_base) * 11 + 20;
-    hsd_804CF810.x8 = hsd_804CF810.x40 - 68;
-    hsd_804CF810.x50 = lbl_8040AB20;
-    b6 = hsd_804CF810.x0_b6;
-    ch = *(&buf[19] + lbl_8040BC3C.x14);
-    if (hsd_804CF810.x0_b7) {
-        void* bitmap = (void*) ((u8*) hsd_804CF810.x4C + (ch & 0x7F) * 0x38);
-        s32 x = hsd_804CF810.x4;
-        s32 dst = (&hsd_804CF810.x24)[hsd_804CF810.x34];
-        hsd_803922FC(bitmap, x, hsd_804CF810.x8, b6, dst, hsd_804CF810.x3C,
-                     hsd_804CF810.x40, hsd_804CF810.x44, hsd_804CF810.x50);
-    } else {
-        void* bitmap = (void*) ((u8*) hsd_804CF810.x4C + (ch & 0x7F) * 0x38);
-        s32 x = hsd_804CF810.x4;
-        s32 dst = (&hsd_804CF810.x24)[hsd_804CF810.x34];
-        hsd_803921B8(bitmap, x, hsd_804CF810.x8, dst, hsd_804CF810.x3C,
-                     hsd_804CF810.x40, hsd_804CF810.x44, hsd_804CF810.x50);
-    }
-    hsd_804CF810.x50 = saved;
-}
-
-s32 hsd_80396A20(void* data)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    ExcptNode* node = data;
-    u32 val = lbl_8040BC3C.x10;
-    s32 shift = 24 - (lbl_8040BC3C.x14 * 4);
-    u32 bit = 1;
-    s32 mask = 0xF << shift;
-
-    while (bit <= sp->xBC) {
-        switch (sp->xBC & bit) {
-        case 0x8:
-            lbl_8040BC3C.x10 =
-                (val & ~mask) | ((((val >> shift) - 1) & 0xF) << shift);
-            return 1;
-        case 0x4:
-            lbl_8040BC3C.x10 =
-                (val & ~mask) | ((((val >> shift) + 1) & 0xF) << shift);
-            return 1;
-        case 0x1:
-            if (lbl_8040BC3C.x14 > 0) {
-                lbl_8040BC3C.x14 = lbl_8040BC3C.x14 - 1;
-            }
-            return 1;
-        case 0x2:
-            if (lbl_8040BC3C.x14 < 5) {
-                lbl_8040BC3C.x14 = lbl_8040BC3C.x14 + 1;
-            }
-            return 1;
-        case 0x100:
-            if ((lbl_8040BC3C.x10 & 0x0FFFFFFF) < OSGetPhysicalMemSize()) {
-                hsd_80394E8C(lbl_8040BC3C.x18);
-                return 1;
-            }
-            goto default_case;
-        case 0x400:
-            lbl_8040BC3C.x10 = val & ~mask;
-            return 1;
-        case 0x200:
-            ps_remove_node(sp, node);
-            return 1;
-        case 0x1000: {
-            extern ExcptNode lbl_8040BD74[];
-            if (lbl_8040BD74 != NULL) {
-                fn_80394DF4(lbl_8040BD74);
-                lbl_8040BD74->next = sp->xD0;
-                sp->xD0 = lbl_8040BD74;
-                if (lbl_8040BD74->callback != NULL) {
-                    lbl_8040BD74->callback(lbl_8040BD74);
-                }
-                sp->x0_b5 = 1;
-            }
-            return 1;
-        }
-        default:
-        default_case:
-            bit <<= 1;
-            break;
-        }
-    }
-    return 0;
-}
-
-s32 hsd_80396C78(void* data)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    s32 result;
-    DispData* disp = data;
-
-    result = hsd_80395550(data);
-    switch (result) {
-    case 1:
-        return 1;
-    case 2:
-        switch (disp->x14) {
-        case 0:
-            sp->xBC = 8;
-            break;
-        case 1:
-            sp->xBC = 4;
-            break;
-        case 2:
-            sp->xBC = 1;
-            break;
-        case 3:
-            sp->xBC = 2;
-            break;
-        case 4:
-            sp->xBC = 0x100;
-            break;
-        case 5:
-            sp->xBC = 0x400;
-            break;
-        case 6:
-            sp->xBC = 0x200;
-            break;
-        }
-        ps_remove_node(sp, data);
-        return 0;
-    case -1:
-        ps_remove_node(sp, data);
-        return 1;
-    default:
-        return 0;
-    }
-}
-
-static char* lbl_804D62FC = "+- DBAT ------------------------+";
-static char* lbl_804D6300 = "+- IBAT ------------------------+";
-static char* lbl_804D6304 = "| %d BL xxxxxxxxxxx  %c%c %c%c%c%c %s |";
-static char* lbl_804D6308 = "|   BEPI %08X BRPN %08X |";
-static char* lbl_804D630C = "+-------------------------------+";
-static char lbl_804D6310[] = "N/A";
-static char lbl_804D6314[] = "R/W";
-static char lbl_804D6318[] = "R/O";
-
-// @TODO: Currently 86.10% match - .bss.0 relocation affects register
-// allocation
-void hsd_80396E40(s32 keycode)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    void** px50 = &sp->x50;
-    s32* px4 = &sp->x4;
-    s32* px40 = &sp->x40;
-    s32* px8 = &sp->x8;
-    char buf[64];
-    void* saved;
-    s32 row;
-    s32 i;
-    s32 spr_u;
-    s32 spr_l;
-    s32 bat_u;
-    s32 bat_l;
-    char* perm;
-    s32 var_g;
-    s32 var_m;
-    s32 var_i;
-    s32 var_w;
-    s32 var_u;
-    s32 var_s;
-    s32 bit;
-    char* ptr;
-    s32 j;
-    char ch;
-
-    saved = *px50;
-    row = sp->x1C - 1;
-    *px50 = lbl_8040AB00;
-    *px4 = 0x106;
-    *px8 = (*px40 - 0x28) - (row + 1) * 14;
-    row--;
-    if (keycode < 0x220 && keycode >= 0x218) {
-        hsd_80394434("+- DBAT ------------------------+");
-    } else {
-        hsd_80394434("+- IBAT ------------------------+");
-    }
-    i = 0;
-    spr_u = 0x219;
-    spr_l = 0x218;
-    do {
-        bat_u = baselib_mfspr(spr_u);
-        bat_l = baselib_mfspr(spr_l);
-        switch (bat_u & 2) {
-        case 0:
-            perm = lbl_804D6310;
-            break;
-        case 2:
-            perm = lbl_804D6314;
-            break;
-        default:
-        case 1:
-            perm = lbl_804D6318;
-            break;
-        }
-        if (bat_l & 8) {
-            var_g = 'G';
-        } else {
-            var_g = 'G';
-        }
-        if (bat_l & 0x10) {
-            var_m = 'M';
-        } else {
-            var_m = 'm';
-        }
-        if (bat_l & 0x20) {
-            var_i = 'I';
-        } else {
-            var_i = 'i';
-        }
-        if (bat_l & 0x40) {
-            var_w = 'W';
-        } else {
-            var_w = 'w';
-        }
-        if (bat_l & 1) {
-            var_u = 'U';
-        } else {
-            var_u = 'u';
-        }
-        if (bat_l & 2) {
-            var_s = 'S';
-        } else {
-            var_s = 's';
-        }
-        sprintf(buf, "| %d BL xxxxxxxxxxx  %c%c %c%c%c%c %s |", i, (int) var_s,
-                (int) var_u, (int) var_w, (int) var_i, (int) var_m,
-                (int) var_g, perm);
-        ptr = buf;
-        for (j = 0; j < 11; j++) {
-            if (bat_l & (1 << (0x1F - (j + 0x13)))) {
-                ch = '1';
-            } else {
-                ch = '0';
-            }
-            ptr[7] = ch;
-            ptr++;
-        }
-        *px4 = 0x106;
-        *px8 = (*px40 - 0x28) - (row + 1) * 14;
-        row--;
-        hsd_80394434(buf);
-        sprintf(buf, "|   BEPI %08X BRPN %08X |", bat_l & 0xFFFC0000,
-                bat_u & 0xFFFC0000);
-        *px4 = 0x106;
-        *px8 = (*px40 - 0x28) - (row + 1) * 14;
-        row--;
-        hsd_80394434(buf);
-        i++;
-        spr_u += 2;
-        spr_l += 2;
-    } while (i < 4);
-    *px4 = 0x106;
-    *px8 = (*px40 - 0x28) - (row + 1) * 14;
-    hsd_80394434("+-------------------------------+");
-    *px50 = saved;
-}
-
-void fn_803970D8(void) {}
-
-void fn_803970DC(void) {}
-
-void fn_803970E0(void) {}
-
-void fn_803970E4(void) {}
-
-void fn_803970E8(void) {}
-
-void fn_803970EC(void) {}
-
-void fn_803970F0(void) {}
-
-void fn_803970F4(void) {}
-
-void fn_803970F8(void) {}
-
-void fn_803970FC(void) {}
-
-void fn_80397100(void) {}
-
-void fn_80397104(void) {}
-
-void fn_80397108(void) {}
-
-void fn_8039710C(void) {}
-
-static char* lbl_804D631C = "+- SPR --------------+";
-static char* lbl_804D6320 = "| NO   NAME    VAL   |";
-static char* lbl_804D6324 = "|%4d %s%s %08X|";
-static char* lbl_804D6328 = "+---- gekko only ----+";
-static char* lbl_804D632C = "|                    |";
-static char* lbl_804D6330 = "+--------------------+";
-#define SPR_BOX_TOP() lbl_804D631C
-#define SPR_BOX_HEADER() lbl_804D6320
-#define SPR_VALUE_FMT() lbl_804D6324
-#define SPR_GEKKO_ONLY() lbl_804D6328
-#define SPR_BLANK_ROW() lbl_804D632C
-#define SPR_BOX_BOTTOM() lbl_804D6330
-static char lbl_804D6334[] = "     ";
-static char lbl_804D633C[] = "    ";
-static char lbl_804D6344[] = "   ";
-static char lbl_804D6348[] = "  ";
-static char lbl_804D634C[] = " ";
-static char lbl_804D6350[8] = "";
-
-void hsd_80397110(void)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    extern u8 lbl_8040BEC4[];
-    u8* base = lbl_8040AB00;
-    void* saved;
-    u32 offset;
-    struct SPREntry {
-        u32 spr;
-        char* name;
-        u32 unused;
-        void (*callback)(u32);
-    };
-    void** px50 = &sp->x50;
-    s32* px4 = &sp->x4;
-    s32* px40 = &sp->x40;
-    s32* px8 = &sp->x8;
-    char buf[32];
-    s32 row;
-    u32 i;
-    struct SPREntry* spr_entry;
-    char* padding;
-    PAD_STACK(32);
-
-    saved = *px50;
-    row = sp->x1C - 1;
-    *px50 = base;
-    *px4 = 20;
-    *px8 = (*px40 - 0x28) - (row-- + 1) * 14;
-    hsd_80394434(SPR_BOX_TOP());
-    *px4 = 20;
-    *px8 = (*px40 - 0x28) - (row-- + 1) * 14;
-    hsd_80394434(SPR_BOX_HEADER());
-    {
-        u32* bec4 = (u32*) lbl_8040BEC4;
-        i = 0;
-        offset = 0;
-        while (row > 0) {
-            u32 spr_off = bec4[4];
-            if (spr_off + i >= 0x45) {
-                break;
-            }
-            spr_entry =
-                (struct
-                 SPREntry*) ((unsigned char*) ((struct SPREntry*) (base +
-                                                                   0x2D8) +
-                                               spr_off) +
-                             offset);
-            if (i == bec4[5]) {
-                if (spr_entry->callback != 0) {
-                    spr_entry->callback(spr_entry->spr);
-                }
-                *px50 = base + 0x20;
-            } else {
-                *px50 = base;
-            }
-            *px4 = 20;
-            *px8 = (*px40 - 0x28) - (row-- + 1) * 14;
-            if (spr_entry->spr == 0) {
-                hsd_80394434(SPR_GEKKO_ONLY());
-            } else {
-                s32 name_len = strlen(spr_entry->name);
-                switch (name_len) {
-                case 1:
-                    padding = lbl_804D6334;
-                    break;
-                case 2:
-                    padding = lbl_804D633C;
-                    break;
-                case 3:
-                    padding = lbl_804D6344;
-                    break;
-                case 4:
-                    padding = lbl_804D6348;
-                    break;
-                case 5:
-                    padding = lbl_804D634C;
-                    break;
-                default:
-                    padding = lbl_804D6350;
-                    break;
-                }
-                sprintf(buf, SPR_VALUE_FMT(), spr_entry->spr, spr_entry->name,
-                        padding, baselib_mfspr(spr_entry->spr));
-                hsd_80394434(buf);
-            }
-            offset += sizeof(*spr_entry);
-            i++;
-        }
-    }
-    *px50 = base;
-    while (row > 0) {
-        *px4 = 20;
-        *px8 = (*px40 - 0x28) - (row-- + 1) * 14;
-        hsd_80394434(SPR_BLANK_ROW());
-    }
-    *px4 = 20;
-    *px8 = (*px40 - 0x28) - (row + 1) * 14;
-    hsd_80394434(SPR_BOX_BOTTOM());
-    *px50 = saved;
-}
-
-s32 fn_80397374(void* data)
-{
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    extern u8 lbl_8040BEC4[];
-    u32 bit = 1;
-    u32 spr_off = *(u32*) (lbl_8040BEC4 + 0x10);
-    u32 sel = *(u32*) (lbl_8040BEC4 + 0x14);
-
-    while (bit <= sp->xBC) {
-        switch (sp->xBC & bit) {
-        case 0x8:
-            if (sel != 0) {
-                sel--;
-            } else if (spr_off != 0) {
-                spr_off--;
-            } else {
-                u32 max = sp->x1C - 4;
-                sel = max;
-                spr_off = 0x44;
-                spr_off = spr_off - max;
-            }
-            *(u32*) (lbl_8040BEC4 + 0x10) = spr_off;
-            *(u32*) (lbl_8040BEC4 + 0x14) = sel;
-            if (sp->x1C) {
-            }
-            return 1;
-        case 0x4:
-            if (sel < (u32) (sp->x1C - 4)) {
-                sel++;
-            } else if (spr_off + sel < 0x43) {
-                spr_off++;
-            } else {
-                sel = 0;
-                spr_off = 0;
-            }
-            *(u32*) (lbl_8040BEC4 + 0x10) = spr_off;
-            *(u32*) (lbl_8040BEC4 + 0x14) = sel;
-            return 1;
-        case 0x1:
-        case 0x2:
-        case 0x100:
-        case 0x800:
-        case 0x400:
-        case 0x1000:
-            return 1;
-        case 0x200:
-            ps_remove_node(sp, data);
-            return 1;
-        default:
-            bit <<= 1;
-            break;
-        }
-    }
-    return 0;
-}
-
-void hsd_80397520(void* node_ptr)
-{
-    PSNode* child1;
-    PSNode* child2;
-    PSNode* node = node_ptr;
-
-    if (node == NULL) {
-        return;
-    }
-    if ((child1 = node->child) != NULL && child1 != NULL) {
-        if ((child2 = child1->child) != NULL && child2 != NULL) {
-            if (child2->child != NULL) {
-                hsd_80397520(child2->child);
-            }
-            if (child2->callback != NULL) {
-                child2->callback(child2);
-            }
-        }
-        if (child1->callback != NULL) {
-            child1->callback(child1);
-        }
-    }
-    if (node->callback != NULL) {
-        node->callback(node);
-    }
-}
-
-// @TODO: Currently 99.78% match - register allocation differences remain
-void hsd_803975D4(void)
-{
-    struct ParticleInputState {
-        u8 _pad[0x54];
-        PADStatus pads[8];
-        s32 port;
-        s32 repeat;
-    };
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    PADStatus* cur_pads;
-    PADStatus* pads;
-    u32 reset_mask;
-    s32 port;
-    u16 buttons;
-    u16 changed;
-    u32 new_press;
-
-    reset_mask = 0;
-    if (OSGetResetSwitchState() != 0) {
-        sp->xC4 = 1;
-    } else if (sp->xC4 != 0) {
-        OSResetSystem(1, 0, 0);
-        for (;;) {
-        }
-    }
-    cur_pads = (PADStatus*) sp->_pad4;
-    memcpy(&sp->_pad4[0x30], cur_pads, 0x30);
-    PADRead(cur_pads);
-    PADClamp(cur_pads);
-    for (port = 0; port < 4; port++) {
-        if ((s8) sp->_pad4[0xA + port * 0xC] == -1) {
-            reset_mask |= 0x80000000 >> port;
-        }
-    }
-    if (reset_mask != 0) {
-        PADReset(reset_mask);
-    }
-    for (port = 0; port < 4; port++) {
-        if ((s8) sp->_pad4[0x3A + port * 0xC] == 0 &&
-            (s8) sp->_pad4[0xA + port * 0xC] == 0)
-        {
-            break;
-        }
-    }
-    if (port == 4) {
-        sp->xBC = 0;
-        sp->xC0 = 0;
-        return;
-    }
-    ((struct ParticleInputState*) sp)->port = port;
-    pads = ((struct ParticleInputState*) sp)->pads;
-    buttons = pads[((struct ParticleInputState*) sp)->port].button;
-    changed =
-        pads[((struct ParticleInputState*) sp)->port + 4].button ^ buttons;
-    sp->xC0 = buttons;
-    new_press = (u16) (buttons & changed);
-    if (new_press != 0) {
-        ((struct ParticleInputState*) sp)->repeat = 0;
-    } else if (pads[((struct ParticleInputState*) sp)->port].button != 0) {
-        ((struct ParticleInputState*) sp)->repeat += 1;
-        if (((struct ParticleInputState*) sp)->repeat > 30) {
-            new_press = pads[((struct ParticleInputState*) sp)->port].button;
-        }
-    }
-    sp->xBC = new_press;
-}
-
-// @TODO: Currently 98.70% match - register allocation and relocation
-// differences
-static inline PSNode* ps_node_child(PSNode* node)
-{
-    return node->child;
-}
-
-void* fn_80397814(void* arg)
-{
-    u8* base = lbl_8040AB00;
-    void* ctx;
-    struct ParticleScreenState* sp = &hsd_804CF810;
-    u32 retrace;
-    u32 next_retrace;
-    u32* keybuf;
-    PAD_STACK(160);
-
-    ctx = arg;
-    hsd_80393D2C(0);
-    hsd_80394314();
-    PADSetSamplingRate(0);
-
-    /* Wait for first retrace */
-    retrace = VIGetRetraceCount();
-    do {
-        next_retrace = VIGetRetraceCount();
-    } while (next_retrace == retrace);
-    retrace = next_retrace;
-    keybuf = &sp->xC0;
-    do {
-        hsd_803975D4();
-        do {
-            next_retrace = VIGetRetraceCount();
-        } while (next_retrace == retrace);
-        retrace = next_retrace;
-    } while (*keybuf != 0);
-
-    retrace = VIGetRetraceCount();
-    do {
-        next_retrace = VIGetRetraceCount();
-    } while (next_retrace == retrace);
-    retrace = next_retrace;
-    do {
-        hsd_803975D4();
-        do {
-            next_retrace = VIGetRetraceCount();
-        } while (next_retrace == retrace);
-        retrace = next_retrace;
-    } while (*keybuf != 0x70);
-
-    retrace = VIGetRetraceCount();
-    do {
-        next_retrace = VIGetRetraceCount();
-    } while (next_retrace == retrace);
-    retrace = next_retrace;
-    do {
-        hsd_803975D4();
-        do {
-            next_retrace = VIGetRetraceCount();
-        } while (next_retrace == retrace);
-        retrace = next_retrace;
-    } while (*keybuf != 0);
-
-    retrace = VIGetRetraceCount();
-    do {
-        next_retrace = VIGetRetraceCount();
-    } while (next_retrace == retrace);
-    retrace = next_retrace;
-    do {
-        hsd_803975D4();
-        do {
-            next_retrace = VIGetRetraceCount();
-        } while (next_retrace == retrace);
-        retrace = next_retrace;
-    } while (*keybuf != 0x808);
-
-    retrace = VIGetRetraceCount();
-    do {
-        next_retrace = VIGetRetraceCount();
-    } while (next_retrace == retrace);
-    retrace = next_retrace;
-    do {
-        hsd_803975D4();
-        do {
-            next_retrace = VIGetRetraceCount();
-        } while (next_retrace == retrace);
-        retrace = next_retrace;
-    } while (*keybuf != 0);
-
-    retrace = VIGetRetraceCount();
-    do {
-        next_retrace = VIGetRetraceCount();
-    } while (next_retrace == retrace);
-    retrace = next_retrace;
-    do {
-        hsd_803975D4();
-        do {
-            next_retrace = VIGetRetraceCount();
-        } while (next_retrace == retrace);
-        retrace = next_retrace;
-    } while (*keybuf != 0x104);
-
-    retrace = VIGetRetraceCount();
-    do {
-        next_retrace = VIGetRetraceCount();
-    } while (next_retrace == retrace);
-    retrace = next_retrace;
-    do {
-        hsd_803975D4();
-        do {
-            next_retrace = VIGetRetraceCount();
-        } while (next_retrace == retrace);
-        retrace = next_retrace;
-    } while (*keybuf != 0);
-
-    retrace = VIGetRetraceCount();
-    do {
-        next_retrace = VIGetRetraceCount();
-    } while (next_retrace == retrace);
-    retrace = next_retrace;
-    do {
-        hsd_803975D4();
-        do {
-            next_retrace = VIGetRetraceCount();
-        } while (next_retrace == retrace);
-        retrace = next_retrace;
-    } while (*keybuf != 0x201);
-
-    retrace = VIGetRetraceCount();
-    do {
-        next_retrace = VIGetRetraceCount();
-    } while (next_retrace == retrace);
-    retrace = next_retrace;
-    do {
-        hsd_803975D4();
-        do {
-            next_retrace = VIGetRetraceCount();
-        } while (next_retrace == retrace);
-        retrace = next_retrace;
-    } while (*keybuf != 0);
-
-    retrace = VIGetRetraceCount();
-    do {
-        next_retrace = VIGetRetraceCount();
-    } while (next_retrace == retrace);
-    retrace = next_retrace;
-    do {
-        hsd_803975D4();
-        do {
-            next_retrace = VIGetRetraceCount();
-        } while (next_retrace == retrace);
-        retrace = next_retrace;
-    } while (*keybuf != 0x402);
-
-    /* Configure video */
-    sp->xD4 = ctx;
-    VIConfigure(sp->x30);
-    VISetBlack(0);
-
-    /* Clear display list */
-    {
-        ExcptNode** head = (ExcptNode**) &sp->xD0;
-        ExcptNode* next;
-        ExcptNode* cur;
-        keybuf = (u32*) head;
-        cur = *head;
-        while (cur != NULL) {
-            next = cur->next;
-            cur->next = NULL;
-            cur = next;
-        }
-        *head = NULL;
-    }
-
-    /* Link exception node */
-    if (base + 0xDC4 != NULL) {
-        ((ExcptNode*) (base + 0xDC4))->next = NULL;
-        *(void**) keybuf = base + 0xDC4;
-        if (((ExcptNode*) (base + 0xDC4))->callback != NULL) {
-            ((ExcptNode*) (base + 0xDC4))
-                ->callback((ExcptNode*) (base + 0xDC4));
-        }
-    }
-
-    /* Set initialized flag */
-    sp->x0_b5 = 1;
-    hsd_80394668();
-
-    /* Initial display setup */
-    {
-        s32* x3C_ptr;
-        s32* x20_ptr;
-        s32* x14_ptr;
-        s32* x18_ptr;
-        s32* c8_ptr;
-        s32* cc_ptr;
-        s32* fb_ptr;
-        s32* fb2_ptr;
-        u32 retrace2;
-        u32 next_retrace2;
-        s32* col_ptr;
-        s32* nrows_ptr;
-        s32* fb_array;
-        s32* size_ptr;
-        void* lbl_ptr;
-        s32 fb_idx;
-        PSNode* node;
-
-        fb_ptr = (s32*) &sp->x40;
-        fb2_ptr = (s32*) &sp->x44;
-        col_ptr = &sp->x34;
-        x18_ptr = &sp->x18;
-        x14_ptr = &sp->x14;
-        x20_ptr = &sp->x20;
-        nrows_ptr = &sp->x1C;
-        fb_array = &sp->x24;
-        x3C_ptr = &sp->x3C;
-
-        hsd_80394544(*x18_ptr, *x14_ptr, *x20_ptr, *nrows_ptr, 20,
-                     *fb_ptr - 40, fb_array[*col_ptr], *x3C_ptr, *fb_ptr,
-                     *fb2_ptr, (s32) lbl_804088B8, NULL);
-
-        sp->xC8 = 0;
-        c8_ptr = &sp->xC8;
-        cc_ptr = &sp->xCC;
-        *cc_ptr = *nrows_ptr - 1;
-
-        /* Process display node */
-        {
-            PSNode* tmp = *(PSNode**) keybuf;
-            node = tmp;
-        }
-        if (node != NULL) {
-            if (node->child != NULL) {
-                hsd_80397520(node->child);
-            }
-            if (node->callback != NULL) {
-                node->callback(node);
-            }
-        }
-
-        /* Flush and display first frame */
-        size_ptr = &sp->x48;
-        fb_idx = *col_ptr;
-        DCFlushRange((void*) fb_array[fb_idx], *size_ptr);
-        fb_idx = *col_ptr;
-        VISetNextFrameBuffer((void*) fb_array[fb_idx]);
-        VIFlush();
-
-        retrace2 = VIGetRetraceCount();
-        do {
-            next_retrace2 = VIGetRetraceCount();
-        } while (next_retrace2 == retrace2);
-        retrace2 = next_retrace2;
-        lbl_ptr = lbl_804088B8;
-
-        /* Main rendering loop */
-        while (*keybuf != 0) {
-            void* disp_node;
-            s32 result;
-
-            hsd_803975D4();
-        restart_walk:
-            sp->x0_b5 = 0;
-
-            /* Walk display callback list */
-            disp_node = *(void**) keybuf;
-            result = 0;
-            while (disp_node != NULL && !sp->x0_b5) {
-                if (*(void* (**) (void*) )((u8*) disp_node + 0xC) != NULL) {
-                    result = (s32) (*(void* (**) (void*) )((u8*) disp_node +
-                                                           0xC))(disp_node);
-                    switch (result) {
-                    case 0:
-                        break;
-                    default:
-                        goto walk_done;
-                    }
-                }
-                disp_node = *(void**) disp_node;
-            }
-
-            if (sp->x0_b5) {
-                goto restart_walk;
-            }
-            result = 0;
-        walk_done:
-
-            if (result != 0) {
-                /* Advance frame buffer index */
-                s32 idx = *col_ptr;
-                s32 nbufs = sp->x38;
-                idx = (idx + 1) % nbufs;
-                *col_ptr = idx;
-
-                hsd_80394668();
-
-                hsd_80394544(*x18_ptr, *x14_ptr, *x20_ptr, *nrows_ptr, 20,
-                             *fb_ptr - 40, (&sp->x24)[*col_ptr], *x3C_ptr,
-                             *fb_ptr, *fb2_ptr, (s32) lbl_ptr, NULL);
-
-                *c8_ptr = 0;
-                *cc_ptr = *nrows_ptr - 1;
-
-                /* Process display node */
-                node = *(PSNode**) keybuf;
-                if (node != NULL) {
-                    if (node->child != NULL) {
-                        hsd_80397520(ps_node_child(node));
-                    }
-                    if (node->callback != NULL) {
-                        node->callback(node);
-                    }
-                }
-
-                /* Flush and display */
-                fb_idx = *col_ptr;
-                DCFlushRange((void*) (&sp->x24)[fb_idx], *size_ptr);
-                fb_idx = *col_ptr;
-                VISetNextFrameBuffer((void*) (&sp->x24)[fb_idx]);
-                VIFlush();
-            }
-
-            do {
-                next_retrace2 = VIGetRetraceCount();
-            } while (next_retrace2 == retrace2);
-            retrace2 = next_retrace2;
-        }
-    }
-
-    OSPanic((char*) base + 0x13E0, 2228, (char*) base + 0x13F4);
-
-    return NULL;
-}
-
-u8 hsd_804CF8E8[0x1000];
-
-void hsd_80397DA4(OSContext* ctx)
-{
-    OSThread thread;
-    OSCreateThread(&thread, fn_80397814, ctx, hsd_804CF8E8 + 0xFFC,
-                   sizeof(hsd_804CF8E8), 0, 1);
-    OSResumeThread(&thread);
-}
-
-void Exception_StoreDebugLevel(int level)
-{
-    hsd_804D78C8 = level;
-}
-
-void hsd_80397DFC(u32 size)
-{
-    hsd_804D78CC = (size + 0xF) >> 4;
-}
-
-static s8 lbl_8040BF10[0x32] =
-    "unsupported no. of special purpose register (%d).";
-
-int baselib_mfspr(int spr)
-{
-    register int result;
-    switch (spr) {
-#ifdef MWERKS_GEKKO
-    case 0x1:
-        asm { mfxer result }
-        break;
-    case 0x8:
-        asm { mflr result }
-        break;
-    case 0x9:
-        asm { mfctr result }
-        break;
-    case 0x12:
-        asm { mfdsisr result }
-        break;
-    case 0x13:
-        asm { mfdar result }
-        break;
-    case 0x16:
-        asm { mfdec result }
-        break;
-    case 0x19:
-        asm { mfsdr1 result }
-        break;
-    case 0x1A:
-        asm { mfsrr0 result }
-        break;
-    case 0x1B:
-        asm { mfsrr1 result }
-        break;
-    case 0x110:
-        asm { mfsprg result, 0 }
-        break;
-    case 0x111:
-        asm { mfsprg result, 1 }
-        break;
-    case 0x112:
-        asm { mfsprg result, 2 }
-        break;
-    case 0x113:
-        asm { mfsprg result, 3 }
-        break;
-    case 0x118:
-        asm { mfspr result, 0x118 } ///< unknown spr?
-        break;
-    case 0x11A:
-        asm { mfear result }
-        break;
-    case 0x11F:
-        asm { mfspr result, pvr }
-        break;
-    case 0x210:
-        asm { mfibatu result, 0 }
-        break;
-    case 0x211:
-        asm { mfibatl result, 0 }
-        break;
-    case 0x212:
-        asm { mfibatu result, 1 }
-        break;
-    case 0x213:
-        asm { mfibatl result, 1 }
-        break;
-    case 0x214:
-        asm { mfibatu result, 2 }
-        break;
-    case 0x215:
-        asm { mfibatl result, 2 }
-        break;
-    case 0x216:
-        asm { mfibatu result, 3 }
-        break;
-    case 0x217:
-        asm { mfibatl result, 3 }
-        break;
-    case 0x218:
-        asm { mfdbatu result, 0 }
-        break;
-    case 0x219:
-        asm { mfdbatl result, 0 }
-        break;
-    case 0x21A:
-        asm { mfdbatu result, 1 }
-        break;
-    case 0x21B:
-        asm { mfdbatl result, 1 }
-        break;
-    case 0x21C:
-        asm { mfdbatu result, 2 }
-        break;
-    case 0x21D:
-        asm { mfdbatl result, 2 }
-        break;
-    case 0x21E:
-        asm { mfdbatu result, 3 }
-        break;
-    case 0x21F:
-        asm { mfdbatl result, 3 }
-        break;
-    case 0x390:
-        asm { mfspr result, gqr0 }
-        break;
-    case 0x391:
-        asm { mfspr result, gqr1 }
-        break;
-    case 0x392:
-        asm { mfspr result, gqr2 }
-        break;
-    case 0x393:
-        asm { mfspr result, gqr3 }
-        break;
-    case 0x394:
-        asm { mfspr result, gqr4 }
-        break;
-    case 0x395:
-        asm { mfspr result, gqr5 }
-        break;
-    case 0x396:
-        asm { mfspr result, gqr6 }
-        break;
-    case 0x397:
-        asm { mfspr result, gqr7 }
-        break;
-    case 0x398:
-        asm { mfspr result, SPR_HID2
-        } ///< hid2 gives incorrect codegen? compiler bug?
-        break;
-    case 0x399:
-        asm { mfspr result, wpar }
-        break;
-    case 0x39A:
-        asm { mfspr result, dma_u }
-        break;
-    case 0x39B:
-        asm { mfspr result, dma_l }
-        break;
-    case 0x3A8:
-        asm { mfspr result, ummcr0 }
-        break;
-    case 0x3A9:
-        asm { mfspr result, upmc1 }
-        break;
-    case 0x3AA:
-        asm { mfspr result, upmc2 }
-        break;
-    case 0x3AB:
-        asm { mfspr result, usia }
-        break;
-    case 0x3AC:
-        asm { mfspr result, ummcr1 }
-        break;
-    case 0x3AD:
-        asm { mfspr result, upmc3 }
-        break;
-    case 0x3AE:
-        asm { mfspr result, upmc4 }
-        break;
-    case 0x3AF:
-        asm { mfspr result, SPR_USDA } ///< usda not recognized by compiler?
-        break;
-    case 0x3B8:
-        asm { mfspr result, mmcr0 }
-        break;
-    case 0x3B9:
-        asm { mfspr result, pmc1 }
-        break;
-    case 0x3BA:
-        asm { mfspr result, pmc2 }
-        break;
-    case 0x3BB:
-        asm { mfspr result, sia }
-        break;
-    case 0x3BC:
-        asm { mfspr result, mmcr1 }
-        break;
-    case 0x3BD:
-        asm { mfspr result, pmc3 }
-        break;
-    case 0x3BE:
-        asm { mfspr result, pmc4 }
-        break;
-    case 0x3BF:
-        asm { mfspr result, sda }
-        break;
-    case 0x3F0:
-        asm { mfspr result, hid0 }
-        break;
-    case 0x3F1:
-        asm { mfspr result, hid1 }
-        break;
-    case 0x3F2:
-        asm { mfspr result, iabr }
-        break;
-    case 0x3F5:
-        asm { mfspr result, dabr }
-        break;
-    case 0x3F9:
-        asm { mfspr result, l2cr }
-        break;
-    case 0x3FB:
-        asm { mfspr result, ictc }
-        break;
-    case 0x3FC:
-        asm { mfspr result, thrm1 }
-        break;
-    case 0x3FD:
-        asm { mfspr result, thrm2 }
-        break;
-    case 0x3FE:
-        asm { mfspr result, thrm3 }
-        break;
-#endif
-    default:
-        OSReport("unsupported no. of special purpose register (%d).", spr);
-        return 0;
-    }
-    return result;
-}
-
-void fn_803982E4(HSD_GObj* gobj, int unused)
-{
-    HSD_CObjSetCurrent(gobj->hsd_obj);
-    fn_80392934();
-    hsd_8039254C();
-}
-
-static HSD_CObjDesc lbl_8040BF70 = { 0 };
-
-HSD_GObj* hsd_80398310(u16 class_id, u8 p_link, u8 obj_kind, u32 gx_link)
-{
-    HSD_GObj* gobj;
-    HSD_CObj* cobj;
-
-    gobj = GObj_Create(class_id, p_link, obj_kind);
-    if (gobj == NULL) {
-        return NULL;
-    }
-    cobj = HSD_CObjLoadDesc(&lbl_8040BF70);
-    HSD_GObjObject_80390A70(gobj, HSD_GObj_804D784B, cobj);
-    GObj_SetupGXLinkMax(gobj, fn_803982E4, gx_link);
-    hsd_80392528((Event) fn_80392A3C);
-    fn_80392A08(4, 1, 0);
-    return gobj;
-}
 
 void hsd_803983A4(HSD_Generator* gen)
 {
@@ -4629,8 +107,6 @@ void hsd_803983A4(HSD_Generator* gen)
         }
     }
 }
-
-HSD_JObj* hsd_804D08E8[8];
 
 // @TODO: Currently 96.40% match - lis hoisting and r29/r30 register swap
 void psInitDataBankLoad(int bank, int* cmdBank, int* texBank, u32* ref,
@@ -4897,17 +373,17 @@ void hsd_80398A08(u32 unused)
     hsd_804D0908[13] = NULL;
     hsd_804D0908[14] = NULL;
     hsd_804D0908[15] = NULL;
-    hsd_804D78E2 = i;
-    hsd_804D78DC = i;
-    for (; i < 0x41; i++) {
+    hsd_804D78E2[0] = 0;
+    hsd_804D78DC = 0;
+    for (i = 0; i < 0x41; i++) {
+        psCmdListArray[i] = NULL;
         psNumCmdList[i] = 0;
         psFormGroupArray[i] = NULL;
         ptclref_804D0E5C[i] = NULL;
         psTexGroupArray[i] = NULL;
-        psCmdListArray[i] = NULL;
-        hsd_804D0908[0x10 + i] = NULL;
+        hsd_804D0908[i] = NULL;
     }
-    hsd_804D78D4 = 0;
+    psCallback = NULL;
     hsd_804D08E8[0] = NULL;
     hsd_804D08E8[1] = NULL;
     hsd_804D08E8[2] = NULL;
@@ -4936,9 +412,9 @@ HSD_Particle* hsd_80398C04(HSD_Particle** head, int linkNo, int bank, u32 kind,
         return NULL;
     }
 
-    hsd_804D78E2 += 1;
-    if (hsd_804D78E2 > numPeakParticles) {
-        numPeakParticles = hsd_804D78E2;
+    hsd_804D78E2[0] += 1;
+    if (hsd_804D78E2[0] > numPeakParticles) {
+        numPeakParticles = hsd_804D78E2[0];
     }
 
     if (gp != NULL) {
@@ -7308,7 +2784,7 @@ do_life:
             /* Free particle */
             hsd_8039D048(pp);
             HSD_ObjFree(&hsd_804D0F60.alloc_data, pp);
-            numActiveParticles--;
+            hsd_804D78E2[0]--;
             return next_pp;
         }
     }
@@ -7548,1284 +3024,10 @@ void hsd_8039D0A0(HSD_Generator* gen)
             }
 
             HSD_ObjFree(&data->alloc_data, prt);
-            hsd_804D78E2--;
+            hsd_804D78E2[0]--;
         } else {
             prev = prt;
         }
         prt = next;
     }
-}
-
-static u16 lbl_804D6368 = 0x100;
-
-void hsd_8039D1E4(HSD_Generator* gen, void* userfunc)
-{
-    gen->userfunc = userfunc;
-}
-
-u16 hsd_8039D1EC(void)
-{
-    lbl_804D6368++;
-    if (lbl_804D6368 < 256) {
-        lbl_804D6368 = 256;
-    }
-    return lbl_804D6368;
-}
-
-void hsd_8039D214(HSD_Generator* gen)
-{
-    HSD_JObj* jobj;
-
-    if (!(gen->type & 0x100)) {
-        return;
-    }
-
-    if ((jobj = gen->jobj) == NULL) {
-        return;
-    }
-
-    HSD_JObjSetupMatrix(jobj);
-
-    if (gen->type & 0x200) {
-        gen->pos.x = gen->jobj->mtx[0][3];
-        gen->pos.y = gen->jobj->mtx[1][3];
-        gen->pos.z = gen->jobj->mtx[2][3];
-    }
-
-    if (gen->type & 0x800) {
-        if (gen->appsrt != NULL && gen->appsrt->gp == gen) {
-            gen->appsrt->translate.x = gen->jobj->mtx[0][3];
-            gen->appsrt->translate.y = gen->jobj->mtx[1][3];
-            gen->appsrt->translate.z = gen->jobj->mtx[2][3];
-        }
-    }
-
-    if (gen->type & 0x1000) {
-        if (gen->appsrt != NULL && gen->appsrt->gp == gen) {
-            HSD_MtxGetScale(gen->jobj->mtx, &gen->appsrt->scale);
-        }
-    }
-}
-
-void hsd_8039D354(u32 unused)
-{
-    HSD_ObjAllocInit(&hsd_804D0F90.alloc_data, 0x94, 4);
-    hsd_804D78FC = NULL;
-    hsd_804D78E0 = 0;
-    hsd_804D78DA = 0;
-    hsd_804D78F4 = 0;
-    hsd_804D78F0 = 0;
-    hsd_804D78E8 = 0;
-    hsd_804D78EC = 0;
-    hsd_804D78F8 = 0;
-    hsd_804D7900 = 0;
-}
-
-HSD_Generator* hsd_8039D3AC(HSD_Generator* gen, HSD_Generator* prev)
-{
-    if (gen->type & 0x80) {
-        hsd_8039D0A0(gen);
-    }
-    if (gen->numChild != 0) {
-        gen->random = 0.0F;
-        gen->genLife = 1;
-        return gen;
-    }
-    if (gen->type & 0x100) {
-        if (gen->jobj != NULL) {
-            if (gen->type & 0x1800) {
-                HSD_psAppSRT* srt = gen->appsrt;
-                if (srt != NULL) {
-                    if (srt->gp == gen) {
-                        if (srt->usedCount != 1) {
-                            gen->random = 0.0F;
-                            gen->genLife = 1;
-                            return gen;
-                        }
-                    }
-                }
-            }
-        }
-    }
-    if (prev == NULL) {
-        hsd_804D78FC = gen->next;
-    } else {
-        prev->next = gen->next;
-    }
-    if (gen->appsrt != NULL) {
-        psRemoveGeneratorSRT(gen);
-    }
-    if (gen->jobj != NULL) {
-        HSD_JObjUnref(gen->jobj);
-        gen->jobj = NULL;
-    }
-    HSD_ObjFree(&hsd_804D0F90.alloc_data, gen);
-    hsd_804D78E0--;
-    return prev;
-}
-
-void hsd_8039D4DC(HSD_Generator* gen)
-{
-    HSD_Generator* cur;
-
-    cur = hsd_804D78FC;
-    hsd_804D78F8 = 0;
-    while (cur != NULL) {
-        if (cur == gen) {
-            hsd_804D78F8 =
-                (u32) hsd_8039D3AC(gen, (HSD_Generator*) hsd_804D78F8);
-            if (hsd_804D78F8 != 0) {
-                while (((HSD_Generator*) hsd_804D78F8)->next != NULL) {
-                    hsd_804D78F8 = (u32) ((HSD_Generator*) hsd_804D78F8)->next;
-                }
-            } else if (hsd_804D78FC != NULL) {
-                hsd_804D78F8 = (u32) hsd_804D78FC;
-                while (((HSD_Generator*) hsd_804D78F8)->next != NULL) {
-                    hsd_804D78F8 = (u32) ((HSD_Generator*) hsd_804D78F8)->next;
-                }
-            }
-            return;
-        }
-        hsd_804D78F8 = (u32) cur;
-        cur = cur->next;
-    }
-}
-
-void hsd_8039D580(HSD_JObj* jobj)
-{
-    HSD_Generator* next;
-    HSD_Generator* cur;
-
-    if (jobj != NULL) {
-        cur = hsd_804D78FC;
-        while (cur != NULL) {
-            next = cur->next;
-            if (cur->jobj == jobj) {
-                hsd_8039D4DC(cur);
-            }
-            cur = next;
-        }
-    }
-}
-
-void hsd_8039D5DC(HSD_JObj* jobj)
-{
-    HSD_Generator* next;
-    HSD_Generator* gen;
-
-    if (jobj == NULL) {
-        return;
-    }
-    gen = hsd_804D78FC;
-    while (gen != NULL) {
-        next = gen->next;
-        if (gen->jobj == jobj) {
-            hsd_8039D4DC(gen);
-        }
-        gen = next;
-    }
-    if (jobj->flags & 0x1000) {
-        return;
-    }
-    jobj = jobj == NULL ? NULL : jobj->child;
-    while (jobj != NULL) {
-        hsd_8039D5DC(jobj);
-        jobj = jobj == NULL ? NULL : jobj->next;
-    }
-}
-
-void hsd_8039D688(HSD_JObj* jobj, f32** unused1, s32 unused2)
-{
-    HSD_Generator* next;
-    HSD_Generator* gen;
-
-    gen = hsd_804D78FC;
-    hsd_804D78F8 = 0;
-    while (gen != NULL) {
-        next = gen->next;
-        if (gen->jobj == jobj) {
-            if (gen->appsrt != NULL) {
-                u16 type = gen->type;
-                if (type & 0x100) {
-                    gen->type = type | 0x80;
-                }
-            }
-            hsd_804D78F8 =
-                (u32) hsd_8039D3AC(gen, (HSD_Generator*) hsd_804D78F8);
-        } else {
-            hsd_804D78F8 = (u32) gen;
-        }
-        gen = next;
-    }
-}
-
-void hsd_8039D71C(HSD_Generator* gen)
-{
-    HSD_JObj* jobj;
-    Mtx mtx;
-    Vec3 col;
-    f32 col0_x, col0_y, col0_z;
-    f32 col1_x, col1_y, col1_z;
-    PAD_STACK(12);
-
-    if (!(gen->type & 0x100)) {
-        return;
-    }
-
-    if ((jobj = gen->jobj) == NULL) {
-        return;
-    }
-
-    HSD_JObjSetupMatrix(jobj);
-
-    if (gen->type & 0x200) {
-        gen->pos.x = gen->jobj->mtx[0][3];
-        gen->pos.y = gen->jobj->mtx[1][3];
-        gen->pos.z = gen->jobj->mtx[2][3];
-    }
-
-    if (gen->type & 0x800) {
-        if (gen->appsrt != NULL && gen->appsrt->gp == gen) {
-            gen->appsrt->translate.x = gen->jobj->mtx[0][3];
-            gen->appsrt->translate.y = gen->jobj->mtx[1][3];
-            gen->appsrt->translate.z = gen->jobj->mtx[2][3];
-        }
-    }
-
-    if (gen->type & 0x1000) {
-        if (gen->appsrt != NULL && gen->appsrt->gp == gen) {
-            HSD_MtxGetScale(gen->jobj->mtx, &gen->appsrt->scale);
-        }
-    }
-
-    col.x = jobj->mtx[0][0];
-    col.y = jobj->mtx[1][0];
-    col.z = jobj->mtx[2][0];
-    PSVECNormalize(&col, &col);
-    col0_x = col.x;
-    col0_y = col.y;
-    col0_z = col.z;
-
-    col.x = jobj->mtx[0][1];
-    col.y = jobj->mtx[1][1];
-    col.z = jobj->mtx[2][1];
-    PSVECNormalize(&col, &col);
-    col1_x = col.x;
-    col1_y = col.y;
-    col1_z = col.z;
-
-    col.x = jobj->mtx[0][2];
-    col.y = jobj->mtx[1][2];
-    col.z = jobj->mtx[2][2];
-    PSVECNormalize(&col, &col);
-
-    {
-        f32 cx = col.x;
-        f32 cy = col.y;
-        f32 cz = col.z;
-        f32 vx = gen->vel.x;
-        f32 vy = gen->vel.y;
-        f32 vz = gen->vel.z;
-        gen->vel.x = col0_x * vx + col1_x * vy + cx * vz;
-        gen->vel.y = col0_y * vx + col1_y * vy + cy * vz;
-        gen->vel.z = col0_z * vx + col1_z * vy + cz * vz;
-    }
-
-    if ((gen->type & 0xF) == 1) {
-        f32 ax = gen->aux.line.x2;
-        f32 ay = gen->aux.line.y2;
-        f32 az = gen->aux.line.z2;
-        gen->aux.line.x2 =
-            jobj->mtx[0][0] * ax + jobj->mtx[0][1] * ay + jobj->mtx[0][2] * az;
-        gen->aux.line.y2 =
-            jobj->mtx[1][0] * ax + jobj->mtx[1][1] * ay + jobj->mtx[1][2] * az;
-        gen->aux.line.z2 =
-            jobj->mtx[2][0] * ax + jobj->mtx[2][1] * ay + jobj->mtx[2][2] * az;
-    }
-}
-
-HSD_Generator* hsd_8039D9C8(void)
-{
-    HSD_Generator* gen;
-    PAD_STACK(8);
-
-    gen = HSD_ObjAlloc(&hsd_804D0F90.alloc_data);
-    if (gen != NULL) {
-        memset(gen, 0, 0x94);
-    }
-    if (gen == NULL) {
-        return NULL;
-    }
-
-    hsd_804D78E0++;
-    if (hsd_804D78E0 > hsd_804D78DA) {
-        hsd_804D78DA = hsd_804D78E0;
-    }
-
-    if ((HSD_Generator*) hsd_804D78F8 == NULL ||
-        ((HSD_Generator*) hsd_804D78F8)->next == NULL)
-    {
-        if (hsd_804D78FC == NULL) {
-            gen->next = NULL;
-            hsd_804D78FC = gen;
-        } else {
-            gen->next = hsd_804D78FC->next;
-            hsd_804D78FC->next = gen;
-        }
-    } else {
-        gen->next = ((HSD_Generator*) hsd_804D78F8)->next->next;
-        ((HSD_Generator*) hsd_804D78F8)->next->next = gen;
-    }
-
-    lbl_804D6368++;
-    if (lbl_804D6368 < 0x100) {
-        lbl_804D6368 = 0x100;
-    }
-    gen->idnum = lbl_804D6368;
-    gen->appsrt = NULL;
-    gen->userfunc = NULL;
-    return gen;
-}
-
-// @TODO: Currently 97.23% match - register allocation, shape/velocity
-// switch case logic, Newton-Raphson sqrt inlining, trig matrix layout
-f32 hsd_8039DAD4(HSD_Generator* gen)
-{
-    Mtx rot_mtx;
-    Vec3 vel_out;
-    Vec3 tmpvec;
-    Vec3 emit_pos;
-    Vec3 vel_copy;
-    Mtx jobj_mtx;
-    Vec3 look_dir;
-    Vec3 cam_up;
-    Vec3 cross1;
-    Vec3 vel_norm;
-    Mtx trig_mtx;
-    f64 eps;
-    f32 vel_mag_sq;
-    f32 angle1;
-    f32 sin_az;
-    f32 cos_az;
-    f32 sin_el;
-    f32 angle_step;
-    f32 cur_angle;
-    f32 radius;
-    f32 cone_angle;
-    f32 elevation;
-    f32 tmp;
-    f32 angle3;
-    PAD_STACK(8);
-
-    angle1 = angle3 = 0.0F;
-    if (gen->count < 1.0F) {
-        return gen->count;
-    }
-
-    vel_mag_sq = angle3;
-
-    /* Copy velocity */
-    vel_copy.x = gen->vel.x;
-    vel_copy.y = gen->vel.y;
-    vel_copy.z = gen->vel.z;
-
-    /* Compute velocity magnitude */
-    vel_mag_sq = vel_copy.z * vel_copy.z;
-    vel_mag_sq += vel_copy.x * vel_copy.x + vel_copy.y * vel_copy.y;
-    if (vel_mag_sq > 0.0F) {
-        f64 e = __frsqrte(vel_mag_sq);
-        e = 0.5 * e * -(vel_mag_sq * (e * e) - 3.0);
-        e = 0.5 * e * -(vel_mag_sq * (e * e) - 3.0);
-        e = 0.5 * e * -(vel_mag_sq * (e * e) - 3.0);
-        tmp = (f32) (vel_mag_sq * e);
-        vel_mag_sq = tmp;
-    }
-
-    /* Initialize rotation matrix */
-    PSMTXIdentity(rot_mtx);
-
-    /* JObj-based rotation: type & 0x100 and jobj!=NULL and type & 0x400
-       and !(kind & 0x3C000) */
-    if ((gen->type & 0x100) && gen->jobj != NULL && (gen->type & 0x400) &&
-        !(gen->kind & 0x30000))
-    {
-        PSMTXCopy((void*) ((u8*) gen->jobj + 0x44), jobj_mtx);
-
-        /* Extract and normalize each column of the 3x3 rotation */
-        tmpvec.x = jobj_mtx[0][0];
-        tmpvec.y = jobj_mtx[1][0];
-        tmpvec.z = jobj_mtx[2][0];
-        PSVECNormalize(&tmpvec, &tmpvec);
-        rot_mtx[0][0] = tmpvec.x;
-        rot_mtx[1][0] = tmpvec.y;
-        rot_mtx[2][0] = tmpvec.z;
-
-        tmpvec.x = jobj_mtx[0][1];
-        tmpvec.y = jobj_mtx[1][1];
-        tmpvec.z = jobj_mtx[2][1];
-        PSVECNormalize(&tmpvec, &tmpvec);
-        rot_mtx[0][1] = tmpvec.x;
-        rot_mtx[1][1] = tmpvec.y;
-        rot_mtx[2][1] = tmpvec.z;
-
-        tmpvec.x = jobj_mtx[0][2];
-        tmpvec.y = jobj_mtx[1][2];
-        tmpvec.z = jobj_mtx[2][2];
-        PSVECNormalize(&tmpvec, &tmpvec);
-        rot_mtx[0][2] = tmpvec.x;
-        rot_mtx[1][2] = tmpvec.y;
-        rot_mtx[2][2] = tmpvec.z;
-
-        rot_mtx[0][3] = 0.0F;
-        rot_mtx[1][3] = 0.0F;
-        rot_mtx[2][3] = 0.0F;
-    }
-
-    /* Billboard orientation: kind & 0x10000 */
-    if (gen->kind & 0x10000) {
-        HSD_ASSERT(677, hsd_804D78F0);
-        {
-            HSD_CObj* cobj = (HSD_CObj*) hsd_804D78F0;
-            void* view = *(void**) ((u8*) cobj + 0x24);
-            look_dir.x = *(f32*) ((u8*) view + 0x0C) - gen->pos.x;
-            view = *(void**) ((u8*) cobj + 0x24);
-            look_dir.y = *(f32*) ((u8*) view + 0x10) - gen->pos.y;
-            view = *(void**) ((u8*) cobj + 0x24);
-            look_dir.z = *(f32*) ((u8*) view + 0x14) - gen->pos.z;
-        }
-        PSVECNormalize(&look_dir, &look_dir);
-        HSD_CObjGetUpVector((HSD_CObj*) hsd_804D78F0, &cam_up);
-        PSVECNormalize(&cam_up, &cam_up);
-        PSVECCrossProduct(&cam_up, &look_dir, &cross1);
-        PSVECCrossProduct(&look_dir, &cross1, &cam_up);
-
-        rot_mtx[0][0] = cross1.x;
-        rot_mtx[1][0] = cross1.y;
-        rot_mtx[2][0] = cross1.z;
-        rot_mtx[0][1] = cam_up.x;
-        rot_mtx[1][1] = cam_up.y;
-        rot_mtx[2][1] = cam_up.z;
-        rot_mtx[0][2] = look_dir.x;
-        rot_mtx[1][2] = look_dir.y;
-        rot_mtx[2][2] = look_dir.z;
-    }
-
-    /* Velocity-based rotation */
-    if ((gen->type & 0xF) != 1 && vel_mag_sq > 0.0F) {
-        vel_norm.x = gen->vel.x;
-        vel_norm.y = gen->vel.y;
-        vel_norm.z = gen->vel.z;
-        PSVECNormalize(&vel_norm, &vel_norm);
-
-        tmp = vel_norm.z;
-        *(s32*) &tmp &= 0x7FFFFFFF;
-        if (tmp < 1.1754944e-38F) {
-            if (vel_norm.y >= 0.0F) {
-                sin_az = 1.5707964F;
-            } else {
-                sin_az = -1.5707964F;
-            }
-        } else {
-            sin_az = atan2f(vel_norm.y, vel_norm.z);
-        }
-        cos_az = sinf(sin_az);
-        sin_az = cosf(sin_az);
-
-        {
-            f32 projected_z = vel_norm.z * sin_az;
-            projected_z += vel_norm.y * cos_az;
-            tmp = projected_z;
-            *(s32*) &tmp &= 0x7FFFFFFF;
-            if (tmp < 1.1754944e-38F) {
-                if (vel_norm.x >= 0.0F) {
-                    elevation = 1.5707964F;
-                } else {
-                    elevation = -1.5707964F;
-                }
-            } else {
-                elevation = atan2f(vel_norm.x, projected_z);
-            }
-        }
-        sin_el = sinf(elevation);
-        {
-            f32 cos_el = cosf(elevation);
-            f32 neg_cos_az = -cos_az;
-            f32 neg_sin_az_s = -sin_az;
-
-            trig_mtx[0][0] = cos_el;
-            trig_mtx[0][1] = 0.0F;
-            trig_mtx[0][2] = sin_el;
-            trig_mtx[0][3] = 0.0F;
-            trig_mtx[1][0] = neg_cos_az * sin_el;
-            trig_mtx[1][1] = sin_az;
-            trig_mtx[1][2] = cos_az * cos_el;
-            trig_mtx[1][3] = 0.0F;
-            trig_mtx[2][0] = neg_sin_az_s * sin_el;
-            trig_mtx[2][1] = neg_cos_az;
-            trig_mtx[2][2] = sin_az * cos_el;
-            trig_mtx[2][3] = 0.0F;
-        }
-        PSMTXConcat(rot_mtx, trig_mtx, rot_mtx);
-    }
-
-    /* Type 2: extract Euler angles from rotation matrix */
-    if ((gen->type & 0xF) == 2) {
-        tmp = rot_mtx[2][2];
-        *(s32*) &tmp &= 0x7FFFFFFF;
-        if (tmp < 1.1754944e-38F) {
-            if (rot_mtx[1][2] >= 0.0F) {
-                angle1 = 1.5707964F;
-            } else {
-                angle1 = -1.5707964F;
-            }
-        } else {
-            angle1 = atan2f(rot_mtx[1][2], rot_mtx[2][2]);
-        }
-        {
-            f32 comb = rot_mtx[2][2] * cosf(angle1);
-            comb += rot_mtx[1][2] * sinf(angle1);
-            tmp = comb;
-            *(s32*) &tmp &= 0x7FFFFFFF;
-            if (tmp < 1.1754944e-38F) {
-                if (rot_mtx[0][2] >= 0.0F) {
-                    angle3 = 1.5707964F;
-                } else {
-                    angle3 = -1.5707964F;
-                }
-            } else {
-                angle3 = atan2f(rot_mtx[0][2], comb);
-            }
-        }
-    }
-
-    /* Angle step computation (pre-loop) */
-    if (gen->angle < 0.0F) {
-        switch (gen->type & 0xF) {
-        case 0:
-        case 3:
-        case 4: {
-            f32 min_a = gen->aux.disc.minAngle;
-            f32 rnd = HSD_Randf();
-            f32 range = gen->aux.disc.maxAngle - min_a;
-            angle_step = range / (f32) (s32) gen->count;
-            cur_angle = angle_step * rnd + min_a;
-            break;
-        }
-        case 6:
-        case 7: {
-            f32 min_a = gen->aux.cone.minAngle;
-            f32 rnd = HSD_Randf();
-            f32 range = gen->aux.cone.maxAngle - min_a;
-            angle_step = range / (f32) (s32) gen->count;
-            cur_angle = angle_step * rnd + min_a;
-            break;
-        }
-        default: {
-            f32 rnd = HSD_Randf();
-            cur_angle = (f32) (2.0 * (M_PI * (f64) rnd));
-            angle_step = (f32) ((2.0 * M_PI) / (f64) (s32) gen->count);
-            break;
-        }
-        }
-    }
-
-    /* Main particle emission loop */
-    eps = 0.001F;
-    while (gen->count >= 1.0F) {
-        switch (gen->type & 0xF) {
-        case 0: /* point, disc, cone, sphere, etc. */
-        case 3:
-        case 4:
-        case 6:
-        case 7: {
-            /* Compute radius */
-            if (gen->radius < 0.0F) {
-                sin_az = -gen->radius;
-                radius = 0.0F;
-            } else {
-                radius = HSD_Randf();
-                if ((gen->type & 0xF) == 3 || (gen->type & 0xF) == 4) {
-                    radius = sqrtf(radius);
-                }
-                sin_az = radius * gen->radius;
-            }
-
-            /* Compute emission angle per shape */
-            switch (gen->type & 0xF) {
-            case 6:
-                if (gen->angle < 0.0F) {
-                    emit_pos.z = sin_az;
-                    cur_angle += angle_step;
-                    tmp = sin_az;
-                    *(s32*) &tmp &= 0x7FFFFFFF;
-                    if (tmp < 1.1754944e-38F) {
-                        if (gen->aux.cone.height >= 0.0F) {
-                            cone_angle = -gen->angle;
-                        } else {
-                            cone_angle = (f32) (M_PI - gen->angle);
-                        }
-                    } else {
-                        cone_angle = (f32) (M_PI - atan2f(gen->aux.cone.height,
-                                                          sin_az)) -
-                                     gen->angle;
-                    }
-                } else {
-                    cur_angle = gen->aux.cone.minAngle;
-                    {
-                        f32 rnd = HSD_Randf();
-                        f32 range = gen->aux.cone.maxAngle - cur_angle;
-                        cone_angle = radius * gen->angle;
-                        cur_angle = range * rnd + cur_angle;
-                    }
-                    emit_pos.z = sin_az;
-                    tmp = sin_az;
-                    *(s32*) &tmp &= 0x7FFFFFFF;
-                    if (tmp < 1.1754944e-38F) {
-                        if (gen->aux.cone.height >= 0.0F) {
-                            cone_angle = gen->angle;
-                        } else {
-                            cone_angle = (f32) (M_PI + gen->angle);
-                        }
-                    } else {
-                        cone_angle = gen->angle +
-                                     (f32) (M_PI - atan2f(gen->aux.cone.height,
-                                                          sin_az));
-                    }
-                }
-                break;
-            case 7:
-                if (gen->angle < 0.0F) {
-                    cone_angle = (f32) (M_PI - gen->angle);
-                    cur_angle += angle_step;
-                } else {
-                    cur_angle = gen->aux.disc.minAngle;
-                    {
-                        f32 rnd = HSD_Randf();
-                        f32 range = gen->aux.disc.maxAngle - cur_angle;
-                        cone_angle = (f32) (M_PI + gen->angle);
-                        cur_angle = range * rnd + cur_angle;
-                    }
-                }
-                break;
-            default:
-                if (gen->angle < 0.0F) {
-                    cone_angle = -gen->angle;
-                    cur_angle += angle_step;
-                    cone_angle = radius * cone_angle;
-                } else {
-                    cur_angle = gen->aux.disc.minAngle;
-                    {
-                        f32 rnd = HSD_Randf();
-                        f32 range = gen->aux.disc.maxAngle - cur_angle;
-                        cone_angle = radius * gen->angle;
-                        cur_angle = range * rnd + cur_angle;
-                    }
-                }
-                break;
-            }
-
-            /* Compute emission position and velocity from angle */
-            emit_pos.x = sin_az * cosf(cur_angle);
-            emit_pos.y = sin_az * sinf(cur_angle);
-
-            /* Shape 6,7: z from random, scale by height */
-            if ((gen->type & 0xF) == 6 || (gen->type & 0xF) == 7) {
-                emit_pos.z = HSD_Randf();
-                if ((gen->type & 0xF) == 6) {
-                    emit_pos.x *= (1.0F - emit_pos.z);
-                    emit_pos.y *= (1.0F - emit_pos.z);
-                }
-                emit_pos.z *= gen->aux.cone.height;
-            } else {
-                emit_pos.z = 0.0F;
-            }
-
-            /* Compute velocity direction from cone angle */
-            {
-                f32 sin_ca = vel_mag_sq * sinf(cone_angle);
-                vel_out.x = sin_ca * cosf(cur_angle);
-                vel_out.y = sin_ca * sinf(cur_angle);
-                vel_out.z = vel_mag_sq * cosf(cone_angle);
-            }
-
-            /* Shape 3: scale velocity by radius factor */
-            if ((gen->type & 0xF) == 3) {
-                vel_out.x *= radius;
-                vel_out.y *= radius;
-                vel_out.z *= radius;
-            }
-
-            /* Apply rotation matrix to emission position */
-            PSMTXMultVec(rot_mtx, &emit_pos, &emit_pos);
-            emit_pos.x += gen->pos.x;
-            emit_pos.y += gen->pos.y;
-            emit_pos.z += gen->pos.z;
-
-            /* Apply rotation matrix to velocity */
-            PSMTXMultVec(rot_mtx, &vel_out, &vel_out);
-
-            /* Spawn particle */
-            hsd_80398F0C(gen->linkNo, gen->bank, gen->kind, gen->texGroup,
-                         (s32) gen->cmdList, gen->life, 0, (s32) gen,
-                         emit_pos.x, emit_pos.y, emit_pos.z, vel_out.x,
-                         vel_out.y, vel_out.z, gen->size, gen->grav,
-                         gen->fric);
-            break;
-        }
-
-        case 1: /* line */
-        {
-            f32 rnd = HSD_Randf();
-            emit_pos.x = rnd * gen->aux.line.x2;
-            emit_pos.y = rnd * gen->aux.line.y2;
-            emit_pos.z = rnd * gen->aux.line.z2;
-
-            PSMTXMultVec(rot_mtx, &emit_pos, &emit_pos);
-            emit_pos.x += gen->pos.x;
-            emit_pos.y += gen->pos.y;
-            emit_pos.z += gen->pos.z;
-
-            PSMTXMultVec(rot_mtx, &vel_copy, &vel_out);
-
-            hsd_80398F0C(gen->linkNo, gen->bank, gen->kind, gen->texGroup,
-                         (s32) gen->cmdList, gen->life, 0, (s32) gen,
-                         emit_pos.x, emit_pos.y, emit_pos.z, vel_out.x,
-                         vel_out.y, vel_out.z, gen->size, gen->grav,
-                         gen->fric);
-            break;
-        }
-
-        case 2: /* tornado */
-        {
-            if (gen->radius < 0.0F) {
-                sin_az = 0.0F;
-            } else {
-                sin_az = HSD_Randf();
-            }
-            if (gen->angle < 0.0F) {
-                cur_angle += angle_step;
-            } else {
-                f32 rnd = HSD_Randf();
-                cur_angle = (f32) (2.0 * (M_PI * (f64) rnd));
-            }
-            gen->aux.line.x2 = vel_mag_sq;
-            hsd_80398F0C(gen->linkNo, gen->bank, gen->kind | 4, gen->texGroup,
-                         (s32) gen->cmdList, gen->life, 0, (s32) gen, 0.0F,
-                         0.0F, 0.0F, cur_angle, sin_az, 0.0F, gen->size,
-                         angle1, angle3);
-            break;
-        }
-
-        case 5: /* rect */
-        {
-            Vec3 vel_temp;
-            emit_pos.x = HSD_Randf();
-            emit_pos.y = HSD_Randf();
-            emit_pos.z = HSD_Randf();
-
-            switch (gen->aux.rect.flag) {
-            case 0: /* default: no modification */
-                break;
-            case 1: /* +x */
-                emit_pos.x = emit_pos.x > 0.5F ? 1.0F : 0.0F;
-                break;
-            case 2: /* +y */
-                emit_pos.y = emit_pos.y > 0.5F ? 1.0F : 0.0F;
-                break;
-            case 3: /* xy split */
-            {
-                f32 rnd = HSD_Randf();
-                f32 a2 = gen->aux.rect.x;
-                f32 b2 = gen->aux.rect.y;
-                if (rnd > a2 / (a2 + b2)) {
-                    emit_pos.y = emit_pos.y > 0.5F ? 1.0F : 0.0F;
-                } else {
-                    emit_pos.x = emit_pos.x > 0.5F ? 1.0F : 0.0F;
-                }
-                break;
-            }
-            case 4: /* +z */
-                emit_pos.z = emit_pos.z > 0.5F ? 1.0F : 0.0F;
-                break;
-            case 5: /* xz split */
-            {
-                f32 rnd = HSD_Randf();
-                f32 a2 = gen->aux.rect.x;
-                f32 b2 = gen->aux.rect.z;
-                if (rnd > a2 / (a2 + b2)) {
-                    emit_pos.z = emit_pos.z > 0.5F ? 1.0F : 0.0F;
-                } else {
-                    emit_pos.x = emit_pos.x > 0.5F ? 1.0F : 0.0F;
-                }
-                break;
-            }
-            case 6: /* yz split */
-            {
-                f32 rnd = HSD_Randf();
-                f32 a2 = gen->aux.rect.y;
-                f32 b2 = gen->aux.rect.z;
-                if (rnd > a2 / (a2 + b2)) {
-                    emit_pos.z = emit_pos.z > 0.5F ? 1.0F : 0.0F;
-                } else {
-                    emit_pos.y = emit_pos.y > 0.5F ? 1.0F : 0.0F;
-                }
-                break;
-            }
-            case 7: /* 3-way split */
-            {
-                f32 rnd = HSD_Randf();
-                f32 a2 = gen->aux.rect.z;
-                f32 b2 = gen->aux.rect.y;
-                f32 c2 = gen->aux.rect.x;
-                f32 ab = b2 * a2;
-                f32 ca = c2 * b2;
-                f32 total = c2 * (b2 + a2) + ab;
-                f32 r0 = 1.0F / total;
-                f32 t1 = r0 * ca;
-                if (rnd < t1) {
-                    emit_pos.z = emit_pos.z > 0.5F ? 1.0F : 0.0F;
-                } else {
-                    f32 t2 = 1.0F - r0 * c2 * a2;
-                    if (rnd > t2) {
-                        emit_pos.y = emit_pos.y > 0.5F ? 1.0F : 0.0F;
-                    } else {
-                        emit_pos.x = emit_pos.x > 0.5F ? 1.0F : 0.0F;
-                    }
-                }
-                break;
-            }
-            default:
-                break;
-            }
-
-            /* Common rect post-processing */
-            emit_pos.x -= 0.5F;
-            emit_pos.y -= 0.5F;
-            emit_pos.z -= 0.5F;
-
-            /* Multiply by rect's 3x3 matrix */
-            vel_temp.x = gen->aux.rect.yx * emit_pos.y +
-                         gen->aux.rect.xx * emit_pos.x +
-                         gen->aux.rect.zx * emit_pos.z;
-            vel_temp.y = gen->aux.rect.yy * emit_pos.y +
-                         gen->aux.rect.xy * emit_pos.x +
-                         gen->aux.rect.zy * emit_pos.z;
-            vel_temp.z = gen->aux.rect.yz * emit_pos.y +
-                         gen->aux.rect.xz * emit_pos.x +
-                         gen->aux.rect.zz * emit_pos.z;
-
-            PSMTXMultVec(rot_mtx, &vel_temp, &emit_pos);
-            emit_pos.x += gen->pos.x;
-            emit_pos.y += gen->pos.y;
-            emit_pos.z += gen->pos.z;
-
-            /* Compute velocity from rect axes */
-            {
-                f32 mag = gen->aux.rect.zx * gen->aux.rect.zx +
-                          gen->aux.rect.zy * gen->aux.rect.zy +
-                          gen->aux.rect.zz * gen->aux.rect.zz;
-                mag = sqrtf(mag);
-                {
-                    f32 scale = vel_mag_sq / mag;
-                    vel_out.x = gen->aux.rect.zx * scale;
-                    vel_out.y = gen->aux.rect.zy * scale;
-                    vel_out.z = gen->aux.rect.zz * scale;
-                }
-            }
-
-            PSMTXMultVec(rot_mtx, &vel_out, &vel_out);
-
-            hsd_80398F0C(gen->linkNo, gen->bank, gen->kind, gen->texGroup,
-                         (s32) gen->cmdList, gen->life, 0, (s32) gen,
-                         emit_pos.x, emit_pos.y, emit_pos.z, vel_out.x,
-                         vel_out.y, vel_out.z, gen->size, gen->grav,
-                         gen->fric);
-            break;
-        }
-
-        case 8: /* sphere emission */
-        {
-            f32 r0 = gen->aux.cone.height;
-            if (r0 == 0.0F || __fabs(r0 - M_PI) < eps) {
-                f32 rnd = sqrtf(HSD_Randf());
-                radius = (f32) (M_PI * rnd);
-                {
-                    f32 rnd2 = HSD_Randf();
-                    if (rnd2 < 0.5F) {
-                        radius = (f32) (M_PI - radius);
-                    }
-                }
-            } else {
-                f32 rnd = sqrtf(HSD_Randf());
-                radius = gen->aux.cone.height * rnd;
-            }
-            {
-                f32 rnd = HSD_Randf();
-                f32 az_angle;
-                f32 r2 = gen->radius;
-                az_angle = (f32) (2.0 * (M_PI * (f64) rnd));
-                cone_angle = az_angle;
-
-                if (r2 < 0.0F) {
-                    sin_az = -r2;
-                } else {
-                    rnd = sqrtf(HSD_Randf());
-                    sin_az = r2 * rnd;
-                }
-            }
-            {
-                f32 cos_r;
-                f32 sin_r;
-                cos_r = cosf(cone_angle);
-                cos_az = sinf(radius);
-                vel_out.x = cos_az * cos_r;
-                sin_el = sinf(cone_angle);
-                sin_r = sinf(radius);
-                vel_out.y = sin_el * sin_r;
-                vel_out.z = cosf(radius);
-            }
-
-            PSMTXMultVec(rot_mtx, &vel_out, &emit_pos);
-
-            /* Scale position by aux.rect.x */
-            vel_out.x = emit_pos.x * gen->aux.rect.x;
-            vel_out.y = emit_pos.y * gen->aux.rect.x;
-            vel_out.z = emit_pos.z * gen->aux.rect.x;
-
-            /* Scale velocity if needed */
-            if (gen->radius >= 0.0F && gen->aux.rect.x < 0.0F) {
-                f32 scale = sin_az / gen->radius;
-                vel_out.x *= scale;
-                vel_out.y *= scale;
-                vel_out.z *= scale;
-            }
-
-            emit_pos.x = sin_az * emit_pos.x + gen->pos.x;
-            emit_pos.y = sin_az * emit_pos.y + gen->pos.y;
-            emit_pos.z = sin_az * emit_pos.z + gen->pos.z;
-
-            hsd_80398F0C(gen->linkNo, gen->bank, gen->kind, gen->texGroup,
-                         (s32) gen->cmdList, gen->life, 0, (s32) gen,
-                         emit_pos.x, emit_pos.y, emit_pos.z, vel_out.x,
-                         vel_out.y, vel_out.z, gen->size, gen->grav,
-                         gen->fric);
-            break;
-        }
-
-        default: /* shape > 8 */
-            if ((void (*)(HSD_Generator*, Mtx)) hsd_804D78E8 != NULL) {
-                ((void (*)(HSD_Generator*, Mtx)) hsd_804D78E8)(gen, rot_mtx);
-            }
-            break;
-        }
-
-        gen->count -= 1.0F;
-    }
-
-    return gen->count;
-}
-
-void hsd_8039EE24(u32 mask)
-{
-    HSD_Generator* gp;
-    HSD_Generator* gen;
-
-    while (hsd_804D78F4 != 0) {
-        gp = (HSD_Generator*) ((HSD_SList*) hsd_804D78F4)->data;
-        hsd_8039D71C(gp);
-        hsd_804D78F4 = (u32) HSD_SListRemove((HSD_SList*) hsd_804D78F4);
-        if (gp->jobj != NULL) {
-            HSD_JObjUnref(gp->jobj);
-            gp->jobj = NULL;
-        }
-    }
-
-    gen = hsd_804D78FC;
-    hsd_804D78F8 = 0;
-
-    while (gen != NULL) {
-        if (mask & (1 << (gen->linkNo + 16))) {
-            hsd_804D78F8 = (u32) gen;
-            gen = gen->next;
-            continue;
-        }
-        if (gen->kind & 0x800) {
-            hsd_804D78F8 = (u32) gen;
-            gen = gen->next;
-            continue;
-        }
-        hsd_8039D214(gen);
-        if (gen->random < 0.0F) {
-            gen->count -= gen->random;
-        } else {
-            gen->count += gen->random * HSD_Randf();
-        }
-        if (gen->count >= 1.0F) {
-            gen->count = hsd_8039DAD4(gen);
-        }
-        if (gen->genLife != 0) {
-            u16 life = gen->genLife - 1;
-            gen->genLife = life;
-            if (life == 0) {
-                hsd_804D78F8 =
-                    (u32) hsd_8039D3AC(gen, (HSD_Generator*) hsd_804D78F8);
-                if (hsd_804D78F8 != 0) {
-                    gen = ((HSD_Generator*) hsd_804D78F8)->next;
-                } else {
-                    gen = hsd_804D78FC;
-                }
-                continue;
-            }
-        }
-        hsd_804D78F8 = (u32) gen;
-        gen = gen->next;
-    }
-}
-
-HSD_Generator* hsd_8039EFAC(s32 linkNo, s32 bank, s32 gfx_id, HSD_JObj* jobj)
-{
-    HSD_Generator* gen;
-
-    gen = hsd_8039F05C(linkNo, bank, gfx_id);
-    if (gen == NULL) {
-        return NULL;
-    }
-    if (gen != NULL) {
-        gen->jobj = jobj;
-        if (jobj != NULL) {
-            ref_INC(jobj);
-        }
-    }
-    gen->type |= (gen->kind & 0x20000) ? 0x500 : 0x700;
-    return gen;
-}
-
-// @TODO: Currently 96.66% match - command-array register swap and stack
-// frame differences remain
-HSD_Generator* hsd_8039F05C(s32 linkNo, s32 bank, s32 idx)
-{
-    HSD_PSCmdList*** cmdListArr;
-    HSD_PSTexGroup* tg;
-    HSD_Generator* gen;
-    s32 ofs;
-    f32 mag;
-    f32 f0, f1, f3;
-    u32 shape;
-
-    if (bank >= 0x41) {
-        return NULL;
-    }
-    if (linkNo >= 8) {
-        return NULL;
-    }
-    if (idx >= psNumCmdList[bank]) {
-        return NULL;
-    }
-
-    cmdListArr = (HSD_PSCmdList***) &ptclref_804D0E5C[bank];
-    ofs = idx * 4;
-    if ((*cmdListArr)[idx] == NULL) {
-        return NULL;
-    }
-
-    gen = hsd_8039D9C8();
-    if (gen != NULL) {
-        gen->type = (*cmdListArr)[idx]->type;
-        gen->bank = bank;
-        gen->linkNo = linkNo;
-        gen->kind = (*cmdListArr)[idx]->kind;
-        gen->texGroup = (*cmdListArr)[idx]->texGroup;
-        gen->life = (*cmdListArr)[idx]->life;
-        gen->genLife = (*cmdListArr)[idx]->genLife;
-        gen->pos.z = 0.0F;
-        gen->pos.y = 0.0F;
-        gen->pos.x = 0.0F;
-        gen->vel.x = (*cmdListArr)[idx]->vx;
-        gen->vel.y = (*cmdListArr)[idx]->vy;
-        gen->vel.z = (*cmdListArr)[idx]->vz;
-        gen->grav = (*cmdListArr)[idx]->grav;
-        gen->fric = (*cmdListArr)[idx]->fric;
-        gen->size = (*cmdListArr)[idx]->size;
-        gen->cmdList = (*cmdListArr)[idx]->cmdList;
-        gen->radius = (*cmdListArr)[idx]->radius;
-        gen->angle = (*cmdListArr)[idx]->angle;
-        gen->random = (*cmdListArr)[idx]->random;
-
-        if (gen->kind & 0x100) {
-            f1 = gen->random;
-            if (f1 < 0.0F) {
-                f3 = 1.0F;
-                if ((1.0F + f1) > 1.1920929e-7F) {
-                } else {
-                    f3 = 0.0F;
-                }
-                gen->count = f3;
-            } else {
-                gen->count = 0.9999999F;
-            }
-        } else if (gen->random < 0.0F) {
-            gen->count = 0.0F;
-        } else {
-            gen->count = HSD_Randf();
-        }
-
-        tg = ((HSD_PSTexGroup***) ((uintptr_t) cmdListArr - bank * 4 -
-                                   0x30C))[bank][gen->texGroup];
-        if (tg != NULL && tg->palflag != 0) {
-            gen->kind |= 0x10;
-        }
-
-        gen->jobj = NULL;
-        gen->numChild = 0;
-
-        shape = gen->type & 0xF;
-        switch (shape) {
-        case 0:
-        case 3:
-        case 4: {
-            HSD_PSCmdList* c = (*cmdListArr)[idx];
-            f32 p1 = c->param1;
-            if (p1 == 0.0F && c->param2 == 0.0F) {
-                gen->aux.disc.minAngle = 0.0F;
-                gen->aux.disc.maxAngle = 6.2831855F;
-            } else {
-                gen->aux.disc.minAngle = p1;
-                gen->aux.disc.maxAngle = (*cmdListArr)[idx]->param2;
-            }
-            break;
-        }
-        case 1:
-            gen->aux.line.x2 = (*cmdListArr)[idx]->param1;
-            gen->aux.line.y2 = (*cmdListArr)[idx]->param2;
-            gen->aux.line.z2 = (*cmdListArr)[idx]->param3;
-            break;
-        case 6:
-        case 7: {
-            HSD_PSCmdList* c = (*cmdListArr)[idx];
-            f32 p1 = c->param1;
-            if (p1 == 0.0F && c->param2 == 0.0F) {
-                gen->aux.cone.minAngle = 0.0F;
-                gen->aux.cone.maxAngle = 6.2831855F;
-            } else {
-                gen->aux.cone.minAngle = p1;
-                gen->aux.cone.maxAngle = (*cmdListArr)[idx]->param2;
-            }
-            gen->aux.cone.height = (*cmdListArr)[idx]->param3;
-            break;
-        }
-        case 5: {
-            f0 = (*cmdListArr)[idx]->param1;
-            gen->aux.rect.x = f0;
-            gen->aux.rect.xx = f0;
-            f0 = (*cmdListArr)[idx]->param2;
-            gen->aux.rect.y = f0;
-            gen->aux.rect.zx = f0;
-            f0 = (*cmdListArr)[idx]->param3;
-            gen->aux.rect.z = f0;
-            gen->aux.rect.zy = f0;
-            gen->aux.rect.zz = 0.0F;
-            gen->aux.rect.yz = 0.0F;
-            gen->aux.rect.yy = 0.0F;
-            gen->aux.rect.xy = 0.0F;
-            gen->aux.rect.xz = 0.0F;
-            gen->aux.rect.yx = 0.0F;
-            gen->aux.rect.flag = 0;
-            if ((*cmdListArr)[idx]->param1 < 0.0F) {
-                gen->aux.rect.flag |= 1;
-            }
-            if ((*cmdListArr)[idx]->param2 < 0.0F) {
-                gen->aux.rect.flag |= 2;
-            }
-            if ((*cmdListArr)[idx]->param3 < 0.0F) {
-                gen->aux.rect.flag |= 4;
-            }
-            break;
-        }
-        case 8: {
-            f32 vx = gen->vel.x;
-            f32 vy = gen->vel.y;
-            f32 vz = gen->vel.z;
-            f32 vx_sq = vx * vx;
-            f32 vy_sq = vy * vy;
-            f32 vz_sq = vz * vz;
-            mag = vx_sq + vy_sq;
-            mag += vz_sq;
-            gen->aux.sphere.speed = sqrtf(mag);
-
-            f0 = gen->vel.x;
-            f1 = gen->vel.z;
-            vx_sq = f0 * f0;
-            vz_sq = f1 * f1;
-            mag = vx_sq + vz_sq;
-            mag = sqrtf(mag);
-
-            if (mag < 1.1754944e-38F) {
-                if (gen->vel.y >= 0.0F) {
-                    gen->aux.sphere.latMid = 1.5707964F;
-                } else {
-                    gen->aux.sphere.latMid = -1.5707964F;
-                }
-            } else {
-                gen->aux.sphere.latMid = atan2f(gen->vel.y, mag);
-            }
-
-            if (fabsf_bitwise(gen->vel.x) < 1.1754944e-38F) {
-                if (gen->vel.z >= 0.0F) {
-                    gen->aux.sphere.lonMid = 1.5707964F;
-                } else {
-                    gen->aux.sphere.lonMid = -1.5707964F;
-                }
-            } else {
-                gen->aux.sphere.lonMid = atan2f(gen->vel.z, gen->vel.x);
-            }
-            gen->aux.sphere.latRange = (*cmdListArr)[idx]->param1;
-            f1 = gen->aux.sphere.latRange;
-            if (f1 < 0.0F) {
-                gen->aux.sphere.latRange = -f1;
-                gen->aux.sphere.speed = -gen->aux.sphere.speed;
-            }
-            gen->aux.sphere.lonRange = (*cmdListArr)[idx]->param2;
-            break;
-        }
-        default:
-            if ((void (*)(HSD_Generator*)) hsd_804D78EC != NULL) {
-                ((void (*)(HSD_Generator*)) hsd_804D78EC)(gen);
-            }
-            break;
-        }
-
-        if (gen->kind & 0x20000) {
-            gen->type |= 0x800;
-            psAddGeneratorAppSRT_begin(gen, 0);
-            if (gen->appsrt != NULL) {
-                gen->appsrt->xA2 = 1;
-                gen->appsrt->gp = gen;
-            }
-        }
-
-        if ((void (*)(HSD_Generator*)) hsd_804D7900 != NULL) {
-            ((void (*)(HSD_Generator*)) hsd_804D7900)(gen);
-        }
-    }
-    return gen;
-}
-
-HSD_Generator* hsd_8039F6CC(s32 linkNo, s32 bank, s32 gfx_id, HSD_JObj* jobj)
-{
-    HSD_Generator* gen;
-
-    gen = hsd_8039F05C(linkNo, bank, gfx_id);
-    if (gen == NULL) {
-        return NULL;
-    }
-    if (gen != NULL) {
-        gen->jobj = jobj;
-        if (jobj != NULL) {
-            ref_INC(jobj);
-        }
-    }
-    gen->type |= (gen->kind & 0x20000) ? 0x500 : 0x700;
-    hsd_804D78F4 =
-        (u32) HSD_SListAllocAndAppend((HSD_SList*) hsd_804D78F4, gen);
-    return gen;
 }
